% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aras.R
\name{aras}
\alias{aras}
\title{Additive Ratio Assessment (ARAS) method for multiple‑criteria decision making}
\description{
Another method from the \link{wsm} family. In its requirements it is perhaps slightly closer to \link{saw}, as the sum of provided weights must equal 1.

As all methods in this family, it requires a performance matrix to be normalized.

ARAS differs from other methods in the optimality estimation step,
where it computes the ratio between an alternative’s utility and that of the ideal variant.

The computation process:

step 1) Formulate DMM (Decision Making Matrix)

Package usually uses a different name for it, PM (Performance Matrix). The idea
 is the same; we work with m alternatives in n criteria describing  performance
 of these alternatives.

Performance of the i‑th alternative in criterion j is denoted as \mjseqn{x_{ij}}
 and \mjseqn{x_{0j}} represents the optimal value of the j‑th criterion.

Generally speaking if optimal value of j-th criterion is not known then for
 benefit criteria we can compute it as:

\mjsdeqn{x_{0j} = max_i x_{ij}}

and for cost criteria:

\mjsdeqn{x_{0j} = min_i x_{ij}}

Actual implementation in this package is slightly different as it transforms
 scales of cost criteria to benefit scale, simplifying the computation
 process.

step 2) Normalize DMM

for benefit criteria

\mjsdeqn{x_{ij}' = \frac{x_{ij}}{\sum_{i=1}^m x_{ij}}}

step 3) normalized weight matrix

For weights following constraint must hold:

\mjsdeqn{\sum_{j=1}^n w_j = 1}

then we can compute weighted performance as

\mjsdeqn{\hat{x}_{ij} = x_{ij}' w_j}

step 4) optimality estimation

\mjsdeqn{S_i = \sum_{j=1}^n \hat{x}_{ij}}

Alternatives with larger \mjseqn{S_i} are considered better than those with
 a lower value of it.

Up to this point, the procedure is basically same as for \link{wsm} or
 \link{saw}. ARAS adds comparison of the ratio between value of \mjseqn{S_i}
 and ideal variant \mjseqn{S_0} as

\mjsdeqn{K_i = \frac{S_i}{S_0}}

It can be argued that the ranking of alternatives based on sum \mjseqn{S_i}
 and value of ratio \mjseqn{K_i} will be same in both cases. So the
 computation of \mjseqn{K_i} allows us to see the optimality through optics
 of distance to ideal variant.
}
\examples{

## ------------------------------------------------
## Method `aras$new`
## ------------------------------------------------

alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
t <- aras$new(M, w)
}
\references{
Zavadskas, E. K., Turskis, Z.: A New Additive Ratio Assessment (ARAS) method
 in multicriteria decision-making. Technological and Economic Development
 of Economy, 16(2), pp. 159-172, DOI:
 \url{https://dx.doi.org/10.3846/tede.2010.10}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ARAS}
\keyword{SAW}
\keyword{WSM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not normalized) performance matrix}

\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights (there are no other limitations on weights)}

\item{\code{minmax}}{vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{result_table}}{has weighted performance matrix with added columns
to summarize performance and express this sum as a percentage of the best
alternative}

\item{\code{weighted_sum_prc}}{vector specifying how close the alternatives are
to the best alternative (expressed as the percentage of best) sorted
descending (from best to worst)}

\item{\code{scoreM}}{results of the WSM presented as stacked bar chart. The
graph clearly shows contribution of the criteria to overall performance
of the alternative.}

\item{\code{assessment_ratio}}{is ratio between overall performance of the
alternative to the performance of best alternative}

\item{\code{result}}{dataframe with results (just assessment ratio and rank)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aras-new}{\code{aras$new()}}
\item \href{#method-aras-compute}{\code{aras$compute()}}
\item \href{#method-aras-summary}{\code{aras$summary()}}
\item \href{#method-aras-clone}{\code{aras$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aras-new"></a>}}
\if{latex}{\out{\hypertarget{method-aras-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aras$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
t <- aras$new(M, w)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aras-compute"></a>}}
\if{latex}{\out{\hypertarget{method-aras-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of ARAS model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aras$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aras-summary"></a>}}
\if{latex}{\out{\hypertarget{method-aras-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the SAW method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aras$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aras-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aras-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aras$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
