% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_spotis.R
\name{balanced_spotis}
\alias{balanced_spotis}
\title{Balanced Stable Preference Ordering Super Intelligence}
\description{
introduces Expected solution point (ESP) and alpha confidence coefficient
 into the SPOTIS method.

Uses distance d(Ai, S*) as in normal SPOTIS and d(Ai, S+), where S+ is ESP
 measured as best provided performance in the criterion.

Alpha then denotes confidence of decision maker to provided values. The
 coefficient is always in interval <0; 1>, balancing d(Ai, S*) and d(Ai, S+)
 components:

\mjsdeqn{P_i = \alpha \cdot d(A_i, S^+) + (1 - \alpha) \cdot d(A_i, S^*)}
}
\examples{

## ------------------------------------------------
## Method `balanced_spotis$new`
## ------------------------------------------------

# Example from the book (see references)
PM <- rbind(
   c(10.5, -3.1, 1.7),
   c(-4.7, 0, 3.4),
   c(8.1, 0.3, 1.3),
   c(3.2, 7.3, -5.3)
)
rownames(PM) <- c('A1', 'A2', 'A3', 'A4')
colnames(PM) <- c('C1', 'C2', 'C3')
minmax <- c('max', 'min', 'max')
w <- c(0.2, 0.3, 0.5)
bounds <- rbind(
   c(-5, 12),
   c(-6, 10),
   c(-8, 5)
)
rownames(bounds) <- colnames(PM)
colnames(bounds) <- c('min', 'max')
result <- balanced_spotis$new(PM, w, minmax, bounds, alpha = 0.4)
}
\references{
Shekhovtsov, A., Dezert, J. Sałabun, W. Enhancing Personalized Decision-Making
 with the Balanced SPOTIS Algorithm. In Proceedings of the 17th International
 Conference on Agents and Artificial Intelligence (ICAART 2025) - Vol. 3,
 pp. 264-271, DOI: 10.5220/0013119800003890, ISBN: 978-989-758-737-5;
 ISSN: 2184-433X
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{B-SPOTIS}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{bounds}}{dataframe of bounds of criteria (criteria in rows with
min and max columns)}

\item{\code{alpha}}{confidence value for pm (in interval <0; 1>)}

\item{\code{esp}}{Expected solution point as vector of best values of the
criteria}

\item{\code{score_raw}}{vector of average distance to ideal solution}

\item{\code{score}}{vector of average distance to ideal solution ordered}

\item{\code{result}}{dataframe with score and ranks of the alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-balanced_spotis-new}{\code{balanced_spotis$new()}}
\item \href{#method-balanced_spotis-compute}{\code{balanced_spotis$compute()}}
\item \href{#method-balanced_spotis-summary}{\code{balanced_spotis$summary()}}
\item \href{#method-balanced_spotis-clone}{\code{balanced_spotis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-balanced_spotis-new"></a>}}
\if{latex}{\out{\hypertarget{method-balanced_spotis-new}{}}}
\subsection{Method \code{new()}}{
public constructor allowing the user to construct SPOTIS decision
 analysis problem and compute it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{balanced_spotis$new(pm, w, minmax = "max", bounds, alpha = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{bounds}}{dataframe of bounds of criteria (criteria in rows with
min and max columns)}

\item{\code{alpha}}{confidence value for pm (in interval <0; 1>)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialized R6 class with computed results for SPOTIS
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example from the book (see references)
PM <- rbind(
   c(10.5, -3.1, 1.7),
   c(-4.7, 0, 3.4),
   c(8.1, 0.3, 1.3),
   c(3.2, 7.3, -5.3)
)
rownames(PM) <- c('A1', 'A2', 'A3', 'A4')
colnames(PM) <- c('C1', 'C2', 'C3')
minmax <- c('max', 'min', 'max')
w <- c(0.2, 0.3, 0.5)
bounds <- rbind(
   c(-5, 12),
   c(-6, 10),
   c(-8, 5)
)
rownames(bounds) <- colnames(PM)
colnames(bounds) <- c('min', 'max')
result <- balanced_spotis$new(PM, w, minmax, bounds, alpha = 0.4)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-balanced_spotis-compute"></a>}}
\if{latex}{\out{\hypertarget{method-balanced_spotis-compute}{}}}
\subsection{Method \code{compute()}}{
Computes Balanced SPOTIS model based on parameters specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation is required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{balanced_spotis$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-balanced_spotis-summary"></a>}}
\if{latex}{\out{\hypertarget{method-balanced_spotis-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the balanced SPOTIS method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{balanced_spotis$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-balanced_spotis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-balanced_spotis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{balanced_spotis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
