% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copeland.R
\name{copeland}
\alias{copeland}
\title{Copeland's Method AKA Llull method}
\description{
ranked-choice voting system based on counting each candidate's pairwise wins
 and losses. System can be applied to general decision situations. First we
 rank the alternatives in criteria based on their performance.

Then we compute result matrix r.

For each comparison we compare respective ranks with rij = 1 if alternative
 i is ranked better than j, rij = 0 if alternative j is ranked better than i
 and finally rij = 0.5 if alternatives i and j have same rank. rii = 0
 (although technically any value will not have any impact on ranking).

We sum r per j (criteria) to compute Copeland score. Which is directly
 translatable into ranking (higher values are better).
}
\examples{

## ------------------------------------------------
## Method `copeland$new`
## ------------------------------------------------

cri <- c("C1", "C2", "C3")
alt <- c("A1", "A2", "A3", "A4", "A5")
pm <- cbind(
  c(5, 10, 4, 8, 1),
  c(4, 8, 2, 16, 4),
  c(7, 7, 6, 8, 5)
)
rownames(pm) <- alt
colnames(pm) <- cri
minmax <- c("max", "max", "min")
result <- copeland$new(pm, minmax)
}
\references{
Copeland's method. Wikipedia. available from \url{https://en.wikipedia.org/wiki/Copeland's_method}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{Copeland}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{CopelandScore}}{summed results per criteria}

\item{\code{result}}{matrix with points based on pairwise comparison and 
Copeland Score with its accompanying ranking}

\item{\code{rank}}{vector of asigned ranks based on Copeland's score}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-copeland-new}{\code{copeland$new()}}
\item \href{#method-copeland-compute}{\code{copeland$compute()}}
\item \href{#method-copeland-summary}{\code{copeland$summary()}}
\item \href{#method-copeland-clone}{\code{copeland$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copeland-new"></a>}}
\if{latex}{\out{\hypertarget{method-copeland-new}{}}}
\subsection{Method \code{new()}}{
public constructor allowing the user to construct Copeland decision
 analysis problem and compute it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copeland$new(pm, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion. only
numeric values expected. Rows and columns are expected to be named.}

\item{\code{minmax}}{value or vector of values 'min' or 'max' specifying
optimization direction for the criterium}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialized R6 class with computed results for Copeland
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cri <- c("C1", "C2", "C3")
alt <- c("A1", "A2", "A3", "A4", "A5")
pm <- cbind(
  c(5, 10, 4, 8, 1),
  c(4, 8, 2, 16, 4),
  c(7, 7, 6, 8, 5)
)
rownames(pm) <- alt
colnames(pm) <- cri
minmax <- c("max", "max", "min")
result <- copeland$new(pm, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copeland-compute"></a>}}
\if{latex}{\out{\hypertarget{method-copeland-compute}{}}}
\subsection{Method \code{compute()}}{
Computes Copeland model based on parameters specified in constructor.
 Normally this method does not need to be run manually as constructor
 calls it automatically.

Manual re-computation is required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copeland$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copeland-summary"></a>}}
\if{latex}{\out{\hypertarget{method-copeland-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the Copeland method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copeland$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copeland-clone"></a>}}
\if{latex}{\out{\hypertarget{method-copeland-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copeland$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
