% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwm.R
\name{cwm}
\alias{cwm}
\title{Combination Weight Method}
\usage{
cwm(w1, w2, alpha = 0.5)
}
\description{
Approach derives weights from two sets of weights. Usually one set is derived
 by objective weighting method such as EWM, COPRAS (and many other)
 and the other based on weights derived from preferences of decision maker
 such as AHP, BWM and many other methods.

It is claimed that the derived weights are more stable, robust, and reliable
 evaluation by reducing randomness and bias, but there is no "hard" scientific
 evidence of the method being able to do that.

Mathematically:

\mjsdeqn{w = \alpha w_1 + (1 - \alpha) w_2}

Both weights do not need to be normalized - the function is normalizing provided
 weights.
}
\examples{
w1 <- c(1, 3, 4)
w2 <- c(1, 4, 3)
w <- cwm(w1, w2, alpha = 0.5)
}
