% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electre2.R
\name{electre2}
\alias{electre2}
\title{ELECTRE II method used to solve multiple criteria decision making}
\description{
The acronym ELECTRE stands for: ELimination Et Choix Traduisant la REalite
 (ELimination and Choice Expressing REality). ELECTRE II method is then
 designed for ranking purposes as oposed to ELECTRE I method which is
 useful for identification of kernel - as alternatives, which cannot be
 eliminated, are not obviously dominated by other alternatives.

ELECTRE II uses concordance and discordance indexes to construct two partial
 pre-orders by trying do formally describe so called strong and weak
 preferences between the alternatives. Strenght is this regard is taken as
 strength of belief, that alternative a is better then alternative b.

Two partial preorders and then used in aggregation procedure to construct
 final patrial preorder.

The code is partially inspired by code authored by M. Balamurali in
 pyDecisions (in Python) and reimplements it in R, thou some portions of the
 function are solved differently in here. For example whole graph
 simplification process in here relies on functionality of
 iGraph package. This method also differently implements the final
 recommendation (total partial order).

Considering that the ELECTRE II method is based on ELECTRE I method with
 additions to produce ranking, the computation starts with computing
 concordance \code{\link{ELECTRE_ConcordanceMatrix}} and discordance
 \code{\link{ELECTRE_DiscordanceMatrix}} matrixes. See
 \code{\link{electre1}} documentation for mathematical aparatus.

Based on concordance and discordance matrixes and comparing its values with
 the thresholds, we can establish strong \mjseqn{dom_s} or weak
 \mjseqn{dom_w} dominance.

\mjsdeqn{dom_s(a,b) = \left\lbrace\begin{array}{ll} 1 & if \; (C(a,b) \ge c^+ \; and \; d(a,b) \le d^+) \; or \; (C(a,b) \ge c^0 \; and \; d(a,b) \le d^-) \cr 0 & otherwise\end{array}\right.}

\mjsdeqn{dom_w(a,b) = \left\lbrace\begin{array}{ll} 1 & if \; (C(a,b) \ge c^0 \; and \; d(a,b) \le d^+) \; or \; (C(a,b) \ge c^- \; and \; d(a,b) \le d^-) \cr 0 & otherwise\end{array}\right.}

Next preference ratio S(a,b) needs to be computed for \mjseqn{dom_s} and
 \mjseqn{dom_w} matrixes. We establish sum of weights of criteria
 \mjseqn{w^+}, where a outperforms b and sum of criteria \mjseqn{w^-},
 where the opposite applies.

\mjsdeqn{w^+ = \sum_{\forall k: PM_k(a)>PM_k(b)} w_k}

and

\mjsdeqn{w^- = \sum_{\forall k: PM_k(a)<PM_k(b)} w_k}

\mjsdeqn{S(a,b) = \left\lbrace\begin{array}{ll} 1 & if \; w^- = 0 \; or \; \frac{w^+}{w^-} \ge 1 \cr 0 & otherwise\end{array}\right.}

Technically the computation needs to be performed only on pairs of (a,b)
 where either \mjseqn{dom_s} and or \mjseqn{dom_w} showed possibility of a
 outranking b, otherwise we can directly establish, that S(a,b) = 0.

Strong and weak dominances are used to then to form first and second total
 preorder. Both matrixes need to be simplified to remove loops from it first.

These preorders aim to measure how much alternatives over-rank each
 other. The procedure is same for both orders. First we consolidate
 information on dominances by

\mjsdeqn{dom = 2 \cdot dom_s + dom_w}

Then we create from all options not strongly outranked by any other option
 the set D. Those options within D that are linked to each other by a weak
 outranking relationship constitute the set U. The set B consists of all
 options from U not weakly outranked by any other from within U. The single
 option or group of equally ranked options is defined by the union of the
 sets (D - U) and B. This provides enough information to form full ranking.

Second preorder uses same approach described in previous paragraph, but in
 reversed order. We achieve this by providing ranking algorithm transposed
 dominance matrixes and reversing resulting order.

Since we got two separate preorders, agregation procedure process these into
 single ranking order. To do that some simple rules are used:

\itemize{
  \item if a is preferred to b in both pre-orders, then this will also be
 the case in final order
  \item if a has an equivalent ranking to b in the one of the complete
 pre-orders, but is preferred in the other, then a precedes b in the final
 order
  \item if a is preferred to b in the first complete pre-order, but b is
 preferred to a in the second, then the two alternatives are incomparable
 in the final order.
}
}
\examples{

## ------------------------------------------------
## Method `electre2$new`
## ------------------------------------------------

PM <- cbind(
 c(103000,101300,156400,267400,49900,103600,103000,170100,279700,405000),
 c(171.3,205.3,221.7,230.7,122.6,205.1,178.0,226.0,233.8,265.0),
 c(7.65,7.90,7.90,10.50,8.30,8.20,7.20,9.10,10.90,10.30),
 c(352,203,391,419,120,265,419,419,359,265),
 c(11.6,8.4,8.4,8.6,23.7,8.1,11.4,8.1,7.8,6.0),
 c(88.0,78.3,81.5,64.7,74.1,81.7,77.6,74.7,75.5,74.7),
 c(69.7,73.4,69.0,65.6,76.4,73.6,66.2,71.7,70.9,72.0))
rownames(PM) <- c("CBX16","P205G","P405M","P605S",
 "R4GTL","RCLIO","R21TS","R21TU","R25BA","ALPIN")
colnames(PM) <- c("Prix","Vmax","C120","Coff","Acce","Frei","Brui")
minmaxcriteria <-c("min","max","min","max","min","min","min")
w <- c(0.3,0.1,0.3,0.2,0.1,0.2,0.1)
t <- electre2$new(PM, w, minmaxcriteria)
}
\references{
Balamurali, M.: pyDecisions - A Python Library of management decision
 making techniques. Available on-line from
 \url{https://github.com/Valdecy/pyDecisions}

Rogers, Martin and Myastre, Lucien-Yves. ELECTRE and Decision Support:
 Methods and Applications in Engineering and Infrastructure investment.
 Springer 2000, 208 p., ISBN 978-1-4757-5057-7

Prombo, M. Package OutrankingTools, CRAN: 2015, available from:
 \url{https://cran.r-project.org/web/packages/OutrankingTools/}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{I}
\keyword{II}
\keyword{concordance}
\keyword{discordance}
\keyword{matrix}
\keyword{outranking}
\keyword{strong}
\keyword{weak}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not normalized) performance matrix}

\item{\code{pm}}{normalized performance matrix (direction of all criteria is
maximalize)}

\item{\code{w}}{weght vector}

\item{\code{minmaxcriteria}}{vector of criteria direction to be used during
performance matrix normalization. Can be replaced with single max or
min if all criteria are to be maximized or minimized. Implicitly set
to max.}

\item{\code{c_minus}}{first of three concordance threshold parameters.
ELECTRE II method implements the concotdance threshold as fuzzy
triangle. Value is alway between 0-1 and \mjsdeqn{0 \le c^- \le c^0 \le c^+ \le 1}
default 0.65.}

\item{\code{c_zero}}{second of three concordance threshold parameters.
ELECTRE II method implements the concotdance threshold as fuzzy
triangle. Default value 0.75.}

\item{\code{c_plus}}{third of three concordance threshold parameters.
ELECTRE II method implements the concotdance threshold as fuzzy
triangle. Default calue 0.85.}

\item{\code{d_minus}}{first of two parameters defining discordance threshold.
 Threshold is defined as range, where \mjsdeqn{0 \le d^- \le d^+ \le 1}
Default value 0.25.}

\item{\code{d_plus}}{first of two parameters defining discordance threshold.
Default value 0.5.}

\item{\code{ConcordanceMatrix}}{concordance matrix for decision problem}

\item{\code{DiscordanceMatrix}}{discordance matrix for decision problem}

\item{\code{StrongOutranking}}{adjacancy matrix for strong outranking}

\item{\code{WeakOutranking}}{adjacancy matrix for weak outranking}

\item{\code{firstTotalPreorder}}{first total preorder}

\item{\code{secondTotalPreorder}}{second total preorder}

\item{\code{finalPreorderMatrix}}{final preorder matrix}

\item{\code{incomparableAlternatives}}{matrix of incomparable alternatives}

\item{\code{graphResult}}{visualization of final preorder matrix}

\item{\code{finalPreorder}}{final preorder vector}

\item{\code{finalPreorderSorted}}{sorted final preoder vector of alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-electre2-new}{\code{electre2$new()}}
\item \href{#method-electre2-compute}{\code{electre2$compute()}}
\item \href{#method-electre2-summary}{\code{electre2$summary()}}
\item \href{#method-electre2-sensitivity}{\code{electre2$sensitivity()}}
\item \href{#method-electre2-clone}{\code{electre2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre2-new"></a>}}
\if{latex}{\out{\hypertarget{method-electre2-new}{}}}
\subsection{Method \code{new()}}{
public constructor, creates electre2 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre2$new(
  pm,
  w,
  minmaxcriteria = "max",
  c_minus = 0.65,
  c_zero = 0.75,
  c_plus = 0.85,
  d_minus = 0.25,
  d_plus = 0.5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{weght vector}

\item{\code{minmaxcriteria}}{vector of criteria direction to be used during
performance matrix normalization. Can be replaced with single max or
min if all criteria are to be maximized or minimized. Implicitly set
to max.}

\item{\code{c_minus}}{first of three concordance threshold parameters.
ELECTRE II method implements the concotdance threshold as fuzzy
triangle. Value is alway between 0-1 and \mjsdeqn{0 \le c^- \le c^0 \le c^+ \le 1}
default 0.65.}

\item{\code{c_zero}}{second of three concordance threshold parameters.
ELECTRE II method implements the concotdance threshold as fuzzy
triangle. Default value 0.75.}

\item{\code{c_plus}}{third of three concordance threshold parameters.
ELECTRE II method implements the concotdance threshold as fuzzy
triangle. Default calue 0.85.}

\item{\code{d_minus}}{first of two parameters defining discordance threshold.
 Threshold is defined as range, where \mjsdeqn{0 \le d^- \le d^+ \le 1}
Default value 0.25.}

\item{\code{d_plus}}{first of two parameters defining discordance threshold.
Default value 0.5.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{PM <- cbind(
 c(103000,101300,156400,267400,49900,103600,103000,170100,279700,405000),
 c(171.3,205.3,221.7,230.7,122.6,205.1,178.0,226.0,233.8,265.0),
 c(7.65,7.90,7.90,10.50,8.30,8.20,7.20,9.10,10.90,10.30),
 c(352,203,391,419,120,265,419,419,359,265),
 c(11.6,8.4,8.4,8.6,23.7,8.1,11.4,8.1,7.8,6.0),
 c(88.0,78.3,81.5,64.7,74.1,81.7,77.6,74.7,75.5,74.7),
 c(69.7,73.4,69.0,65.6,76.4,73.6,66.2,71.7,70.9,72.0))
rownames(PM) <- c("CBX16","P205G","P405M","P605S",
 "R4GTL","RCLIO","R21TS","R21TU","R25BA","ALPIN")
colnames(PM) <- c("Prix","Vmax","C120","Coff","Acce","Frei","Brui")
minmaxcriteria <-c("min","max","min","max","min","min","min")
w <- c(0.3,0.1,0.3,0.2,0.1,0.2,0.1)
t <- electre2$new(PM, w, minmaxcriteria)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre2-compute"></a>}}
\if{latex}{\out{\hypertarget{method-electre2-compute}{}}}
\subsection{Method \code{compute()}}{
computes the ELECTRE II problem. Usually doesn't need to be run manually
 as it is called automatically by class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre2$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre2-summary"></a>}}
\if{latex}{\out{\hypertarget{method-electre2-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the ELECTRE II method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre2$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre2-sensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-electre2-sensitivity}{}}}
\subsection{Method \code{sensitivity()}}{
sensitivity analysis for ELECTRE II is performed on values of
 concordance and discordance thresholds. Since these thresholds are
 defined as fuzzy numbers, the analysis is performed on five parameters
 c-, c0 and c+ for concordance thresholds, d- and d+ for discordance
 threshold.

Analysis always explores changes of a single parametr, while all other
 are held static (non changing).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre2$sensitivity(step = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{iteration step the function uses to generate the testing
interval}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns dataframe specifying upper and lower limits of the solution's
 sensitivity the thresholds. Limits values can be replaced by
 insens. if no limit has been identified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-electre2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
