% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electre3.R
\name{electre3}
\alias{electre3}
\title{electre3 : ELECTRE III method for ranking alternatives}
\description{
ELECTRE III  method aims to answer the following question: considering a
 finite set of actions, A, evaluated on a coherent family of
 pseudo-criteria, F, how to make a partition of A in classes of  quivalence
 and provide a necessarily complete pre-order expressing the relative
 position of these classes? In the first phase, ELECTRE III method involves
 the construction of a fuzzy outranking relation.

In the second phase, an algorithm is used for making a ranking in a final
 partial pre-order, that combines two complete pre-orders.

This implementation presents simplified version of the method using single
 value alpha, beta parameters. If more complex approach to ELECTRE III
 computation is required use OutrankingTools package for R, which provides
 variant allowing such computations.

The procedure itself is as follows. First we compute Concordance matrix.
Concordance matrix (index) measures strength of the statement that
 alternative a outranks alternative b.

To compute concordance matrix, partial concordance matrix \mjseqn{c_j} of
 the criteria needs to be computed first. The partial concordance matrix
 considers concordance by comparing performances of the alternatives in
 criterium \mjseqn{j}.

\mjsdeqn{c_j(a,b) = \left\lbrace\begin{array}{ll} 1 & \;if\; PM_j(a) + Q_j(a) \ge PM_j(b) \cr 0 & \;if\; PM_j(a) + P_j(a) < PM_j(b)) \cr \frac{PM_j(a) - PM_j(b) + P_j(a)}{P_j(a) - Q_j(a)} & else\end{array}\right.}

Based on partial concordance matrixes we compute concordence matrix
 \mjseqn{C} as:

\mjsdeqn{C(a,b) = \frac{\sum (w_j \cdot c_j(a,b))}{\sum w_j}}

Then we compute discordance matrixes \mjseqn{d_j} separately for each
 criterion using:

\mjsdeqn{d_j(a,b) = \left\lbrace\begin{array}{ll} 1 & \;if\; PM_j(b) > PM_j(a) + V_j(a) \cr 0 & \;if\; PMj(b) \le PM_j(a) + P_j(a) \cr \frac{PM_j(b) - PM_j(a) - P_j(a)}{V_j(a) - P_j(a)} & else\end{array}\right.}

where

PM ... performance of alternative in criterion, Q ... indifference
 threshold, P ... preference threshold, V ... veto threshold, w ...
 weights.

Based on discordance matrixes \mjseqn{d_j} we compute credibility matrix
 \mjseqn{S}

\mjsdeqn{S(a,b) = \left\lbrace\begin{array}{ll} C(a,b) & \;if\; d_j(a,b) \le C(a,b) \cr C(a,b) \cdot \prod_{\forall j \in J} \frac{1 - d_j(a,b)}{1 - C(a,b)} & \;otherwise\; \end{array}\right.}

where J is set of criteria \mjseqn{j} for which \mjseqn{d_j(a,b) > C(a,b)}

Finally we may use the information to derive ranking of the alternatives.To
 do that we will perform descending rank distillation
 \code{\link{Electre_desc_dist}}, followed by ascending rank distillation
 \code{\link{Electre_asc_dist}}, then we agregate both into final rank
 \code{\link{finalRanking}}.

Preorder for descending rank distillation is achieved by distilling
 alternatives from credibility matrix using progressively lower cutoff
 thresholds \mjseqn{\lambda}. first value of the \mjseqn{\lambda} is
 computed:

\mjsdeqn{\lambda_1 = max_{a,b \in A}S(a,b)}

where A ... is set of alternatives.

Ascending distillation works similarly, but going up from lowest values in
 credibility matrix using progressively higher cutoff thresholds
 \mjseqn{\lambda}.

Since both preorders can be different gathered information needs to be
 further processed by creation of pre-order matrix. This matrix provides
 comparison between all alternatives by comparing their mutual rankings in
 descending and ascending preorders.

Suppose we compare alternatives a and b. We already have available its ranks
 in descending (\mjseqn{D_a, D_b}) and ascending (\mjseqn{A_a, A_b})
 preorders. We will use this information to derive relation between the two
 alternatives.

If (\mjseqn{D_a < D_b} and at same time \mjseqn{A_a < A_b}) or
 (\mjseqn{D_a == D_b} and at same time \mjseqn{A_a < A_b})  or
  (\mjseqn{D_a < B_b} and at same time \mjseqn{A_a == A_b}) we can say that
 that a is preferred to b (aPb). In other words aPb if it is either preferred
 in both preorders or when it is better ranked in one while being evaluated
 as equal in other.

Since we work with matrix we use [a, b] = P+ and [b, a] = P- notation to
 express it.

Same rules apply for bPa, leading to [a, b] = P- and [b, a] = P+.

We can also say that if the ranks of the alternatives in both preorders are
 same, then we are indifferent to them preferantion wise: aIb, leading to
 [a, b] = [b, a] = I.

Otherwise the alternatives are incomparable: aRb, leading to
 [a, b] = [b, a] = R. That can hapen if in one preorder aPb, while in second
 one bPa.

This information is directly usable for final ranking derivation.

Please note, that ELECTRE IV (see \code{\link{electre4}}) uses same
 procedure for rank derivation, but does not use weights.
}
\examples{

## ------------------------------------------------
## Method `electre3$new`
## ------------------------------------------------

PM <- cbind(
  c(-14,129,-10,44,-14),
  c(90,100,50,90,100),
  c(40,0,10,5,20),
  c(40,0,10,5,20),
  c(100,0,100,20,40)
)
rownames(PM) <- c("Project1","Project2","Project3","Project4","Project5")
colnames(PM) <- c( "CR1","CR2","CR3","CR4","CR5")
minmaxcriteria <- 'max'
Q <- c(25,16,0,12,10) #Indifference thresholds
P <- c(50,24,1,24,20) #Preference thresholds
V <- c(100,60,2,48,90) #Veto thresholds
w <- c(1,1,1,1,1) #weights
t <- electre3$new(PM, w, P, Q, V, minmaxcriteria)
}
\references{
Balamurali, M.: pyDecisions - A Python Library of management decision making
 techniques. Available on-line from
 \url{https://github.com/Valdecy/pyDecisions}

Rogers, Martin and Myastre, Lucien-Yves. ELECTRE and Decision Support:
 Methods and Applications in Engineering and Infrastructure investment.
 Springer 2000, 208 p., ISBN 978-1-4757-5057-7

Roy B. : "The outranking approach and the foundations of ELECTRE methods",
 Theory and Decision 31, 1991, 49-73.

Vallée, D.; Zielniewicz, P. 1994. ELECTRE III-IV, version 3.x, Aspects
 Méthodologiques (tome 1), Guide d’utilisation (tome 2). Document du LAMSADE
 85 et 85bis, Université Paris Dauphine

Prombo, M. Package OutrankingTools, CRAN: 2015, available from:
 \url{https://cran.r-project.org/web/packages/OutrankingTools/}

Meyer, P. at al. MCDA package. GitHub: 2021, available from:
 \url{https://github.com/paterijk/MCDA/blob/master/}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{Discrimination}
\keyword{ELECTRE}
\keyword{III}
\keyword{Outranking}
\keyword{analysis}
\keyword{approaches}
\keyword{concordance}
\keyword{credibility}
\keyword{global}
\keyword{matrix}
\keyword{methods}
\keyword{modelling}
\keyword{multicriteria}
\keyword{preference}
\keyword{pseudo-criterion}
\keyword{thresholds}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not normalized) performance matrix}

\item{\code{pm}}{normalized performance matrix (direction of all criteria is
maximalize)}

\item{\code{w}}{weght vector}

\item{\code{p}}{Preference threshold - Vector containing the preference
thresholds constraints defined for each criterion.}

\item{\code{q}}{Indifference threshold - Vector containing the indifference
thresholds constraints defined for each criterion.}

\item{\code{v}}{Veto threshold - Vector containing the veto thresholds
constraints defined for each criterion.}

\item{\code{minmaxcriteria}}{vector of criteria direction to be used during
performance matrix normalization. Can be replaced with single max or
min if all criteria are to be maximized or minimized. Implicitly set
to max.}

\item{\code{alpha}}{alpha and beta coefficients are used in downward and upward
distillation procedure as well as final ranking procedure to construct
orders. Both coef. are used iteratively change thresholds limiting
evaluation of the outranking relations between the alternatives.}

\item{\code{beta}}{see alpha. Preset values of alpha = 0.3 and beta = 0.15 are
set as per Vallee and Zielniewicz (1994). This version of method allows
only to set single value for the purpose.}

\item{\code{rank_d}}{Descending distillation ranking - final partial preorder
orders the alternatives from the best to the worst.}

\item{\code{rank_a}}{Ascending distillation ranking - final partial preorder
orders the alternatives from worst to best.}

\item{\code{rank_p}}{Pre-order matrix specifying identified relations between
the alternatives - values are P+ (a preferred to b),
P- (b preferred to a), I (indifferent), R (incomparable)}

\item{\code{graph}}{processed adajncency matrix into network digram.}

\item{\code{ConcordanceMatrix}}{Concordance matrix is one of two working
relations (concordance and discordance) which are subsequently used
to construct the final dominance relation. For an outranking aSb to
be validated, a sufficient majority of criteria should be in favor of
this assertion.}

\item{\code{CredibilityMatrix}}{matrix assessing the strength of the assertion
that a is at least as good as b.}

\item{\code{finalRankingUnsorted}}{named vector of unsorted order of the
alternatives, added for purposes of sensitivity analysis}

\item{\code{DiscordanceMatrixCriteria}}{Discordance matrix is one of two
working relations (concordance and discordance) which are subsequently
used to construct the final dominance relation. The concept of
discordance is complementary to the one of (concordance and represent
the discomfort experienced in the choosing of alternative a above
alternative b. In comparison with ELECTRE I or II methods discordance
matrix for ELECTRE III is computed separately for every criterion.}

\item{\code{adjancancyMatrix}}{Adjacency Matrix allows to visualize results as
network diagram}

\item{\code{final_ranking}}{final order of the alternatives.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-electre3-new}{\code{electre3$new()}}
\item \href{#method-electre3-compute}{\code{electre3$compute()}}
\item \href{#method-electre3-summary}{\code{electre3$summary()}}
\item \href{#method-electre3-sensitivity}{\code{electre3$sensitivity()}}
\item \href{#method-electre3-clone}{\code{electre3$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre3-new"></a>}}
\if{latex}{\out{\hypertarget{method-electre3-new}{}}}
\subsection{Method \code{new()}}{
public constructor, creates electre3 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre3$new(pm, w, p, q, v, minmaxcriteria = "max", alpha = 0.3, beta = 0.15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{weght vector}

\item{\code{p}}{preference threshold - Vector containing the preference
thresholds constraints defined for each criterion.}

\item{\code{q}}{Indifference threshold - Vector containing the indifference
thresholds constraints defined for each criterion.}

\item{\code{v}}{Veto threshold - Vector containing the veto thresholds
constraints defined for each criterion.}

\item{\code{minmaxcriteria}}{vector of criteria direction to be used during
performance matrix normalization. Can be replaced with single max or
min if all criteria are to be maximized or minimized. Implicitly set
to max.}

\item{\code{alpha}}{alpha and beta coefficients are used in downward and upward
distillation procedure as well as final ranking procedure to construct
orders. Both coef. are used iteratively change thresholds limiting
evaluation of the outranking relations between the alternatives.}

\item{\code{beta}}{see alpha. Preset values of alpha = 0.3 and beta = 0.15 are
set as per Vallee and Zielniewicz (1994). This version of method allows
only to set single value for the purpose.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{PM <- cbind(
  c(-14,129,-10,44,-14),
  c(90,100,50,90,100),
  c(40,0,10,5,20),
  c(40,0,10,5,20),
  c(100,0,100,20,40)
)
rownames(PM) <- c("Project1","Project2","Project3","Project4","Project5")
colnames(PM) <- c( "CR1","CR2","CR3","CR4","CR5")
minmaxcriteria <- 'max'
Q <- c(25,16,0,12,10) #Indifference thresholds
P <- c(50,24,1,24,20) #Preference thresholds
V <- c(100,60,2,48,90) #Veto thresholds
w <- c(1,1,1,1,1) #weights
t <- electre3$new(PM, w, P, Q, V, minmaxcriteria)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre3-compute"></a>}}
\if{latex}{\out{\hypertarget{method-electre3-compute}{}}}
\subsection{Method \code{compute()}}{
computes the ELECTRE III problem. Usually doesn't need to be run manually
 as it is called automatically by class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre3$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre3-summary"></a>}}
\if{latex}{\out{\hypertarget{method-electre3-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the ELECTRE III method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre3$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre3-sensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-electre3-sensitivity}{}}}
\subsection{Method \code{sensitivity()}}{
sensitivity analysis for ELECTRE III object documentation) is performed
 on values of preference (p), indefference (q) and veto (v) thresholds.

Since the thresholds (all of them) are specfied separately for each
 criterion, their sensitivity needs to be also evaluated separately.
 This is being realized in this function by generating separate
 dataframe for each threshold (sens_p, sens_q, sens_v) with following
 structure:

criterium; from; default; to

From and to limits are limist of of solution stability, meaning that
 going under it (from) or over it (to) will produce different result.

Sensitivity is tested from default value (the value used to produce
 result) going down to 0 for from column and up to maximal performance
 in the criterium j (max(pm[j])). If no change in result is detected,
 value "insens." is inserted into dataframe.

Note that the computational part is very similar to Electre IV method
 and thus is realized using \code{\link{sensitivity_e34}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre3$sensitivity(steps = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{steps}}{how many steps should sensitivity testing take. Interval
for testing will be split to steps segments. The higher number of steps
the smaller the step will be, the more granular the testing will be.
Set to 100 by default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns dataframes sens_p, sens_q and sens_v with sensitivity limit for
 the criteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre3-clone"></a>}}
\if{latex}{\out{\hypertarget{method-electre3-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
