% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electre_4_paramCheck.R
\name{Electre_4_paramCheck}
\alias{Electre_4_paramCheck}
\title{function for checking of the parameters fothe Electre 4 method}
\usage{
Electre_4_paramCheck(pm, p, q, v, minmaxcriteria = "max")
}
\arguments{
\item{pm}{performance of the alternatives in criteria. Alternatives are in
rows, criteria in columns.}

\item{p}{vector of preference thresholds (each threshold is set separately
for each criterium)}

\item{q}{vector of indifference thresholds}

\item{v}{vector of veto thresholds}

\item{minmaxcriteria}{vector of optimization direction (min/max) for the
criteria}
}
\description{
Checks parameters for ELECTRE 4 method (see \code{\link{Electre_4}}).
 It should stop processing of the inputs early when problem with
 parameters is detected by producing error message into the console.

When no problem with the parameters is detected, it just ends without
 returning any value, allowing the calling function using it to continue.

The function is also used by \code{\link{Electre_3_paramCheck}},
 \code{\link{electre1s}} and \code{\link{electre_TRI}}.
}
\examples{
PM <- cbind(
  c(-14,129,-10,44,-14),
  c(90,100,50,90,100),
  c(40,0,10,5,20),
  c(40,0,10,5,20),
  c(100,0,100,20,40)
)
rownames(PM) <- c("Project1","Project2","Project3","Project4","Project5")
colnames(PM) <- c( "CR1","CR2","CR3","CR4","CR5")
minmaxcriteria <- "max"
Q <- c(25,16,0,12,10) #Indifference thresholds
P <- c(50,24,1,24,20) #Preference thresholds
V <- c(100,60,2,48,90) #Veto thresholds
Electre_4_paramCheck(PM, P, Q, V, minmaxcriteria)
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{1s}
\keyword{ELECTRE}
\keyword{III}
\keyword{IV}
\keyword{TRI}
\keyword{parameter}
\keyword{validity}
