% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalRanking.R
\name{finalRanking}
\alias{finalRanking}
\title{computes final ranking based on pre-order matrix}
\usage{
finalRanking(alt, rank_P)
}
\arguments{
\item{alt}{vector of alternatives (its names)}

\item{rank_P}{pre-order matrix}
}
\value{
returns data frame with final ranking in sorted and unsorted form:
\itemize{
  \item final_ranking - final ranking of the alternatives in sorted form
  \item finalRankingUnsorted - final ranking of the alternatives in unsorted
 form (useful for sensitivity analyses)
}
}
\description{
Takes pre-order matrix as computed by ELECTRE III (\link{electre3}) and
 ELECTRE IV (\link{electre4}) and computes final ranking from it.
}
\references{
Meyer, P. at al. MCDA package. GitHub: 2021, available from:
 \url{https://github.com/paterijk/MCDA/blob/master/}
}
\seealso{
\link{electre3}

\link{electre4}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{III}
\keyword{IV}
\keyword{final}
\keyword{methods}
\keyword{ranking}
