% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuca.R
\name{fuca}
\alias{fuca}
\title{Faire Un Choix Adégnat}
\description{
translates to "make adequate choice". Method proposed by Fernando et al
 (2011). The best alternative is determined based on ranks aggregation.

First we assign ranks to performance of the alternatives in criteria and we
 apply weights on resulting matrix of these ranks.

Alternative with smallest Ri is top ranked.

\mjsdeqn{R_i = \sum_{j=1}^n (r_{ij} \cdot w_j)}

where rij is ranked performance of alternative in criterium.
}
\examples{

## ------------------------------------------------
## Method `fuca$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- fuca$new(pm, w = w, minmax = minmax)
}
\references{
MENDOZA LUIS FERNANDO, Morales et al. Selecting the best portfolio
 alternative from a hybrid multiobjective GA-MCDM approach for New
 Product Development in the pharmaceutical industry. In: 2011 IEEE
 Symposium on Computational Intelligence in Multicriteria
 Decision-Making (MDCM). 2011, pp. 159–166, available from:
 https://doi.org/10.1109/SMDCM.2011.5949271.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{FUCA}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{w}}{weight vector}

\item{\code{minmax}}{vector of optimization direction for criteria. Use
min/max. If sing min or max is used, function will presume same
direction for all criteria}

\item{\code{rankPM}}{performance matrix expresed in form of ranks}

\item{\code{Ri}}{vector of agreggated weighted ranks}

\item{\code{rank}}{vector of alternative ranks derived from Ri}

\item{\code{result}}{data frame with values of Ri and rank}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fuca-new}{\code{fuca$new()}}
\item \href{#method-fuca-compute}{\code{fuca$compute()}}
\item \href{#method-fuca-summary}{\code{fuca$summary()}}
\item \href{#method-fuca-clone}{\code{fuca$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuca-new"></a>}}
\if{latex}{\out{\hypertarget{method-fuca-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuca$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- fuca$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuca-compute"></a>}}
\if{latex}{\out{\hypertarget{method-fuca-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of FUCA model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuca$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuca-summary"></a>}}
\if{latex}{\out{\hypertarget{method-fuca-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the FUCA count method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuca$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuca-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fuca-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuca$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
