% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyvikor.R
\name{fuzzyvikor}
\alias{fuzzyvikor}
\title{method used to solve multiple criteria decision making}
\description{
The acronym VIKOR stands for: VlseKriterijumska Optimizacija I Kompromisno
 Resenje, in Serbian multicriteria optimization and compromise solution. The
 method has been especially designed to deal with problematic situations
 when the alternatives are characterized by non-commensurable and
 conflicting criteria, for which VIKOR provides compromise solution.
 Methodologically VIKOR is close to another method TOPSIS. Original VIKOR
 uses five steps to derive such compromise solution.

FuzzyVIKOR uses trapezoidal fuzzy number for computation purposes. Other
 then using fuzzy number for computation, the results are basically the
 same: S, R and Q, which can be used to compute compromise solution. Look
 in the VIKOR function (see \link{vikor}) description to better
 understand the computation proces (or the source code).

At present time FuzzyVIKOR function does not have implemented establishment
 of compromise solution.

Mathematically the only thing different of FuzzyVIKOR to norlam VIKOR is
 necessity to compute crisp number from the trapezoidal fuzzy number.
 Following formula is used to do that.

\mjsdeqn{ crisp_a =  -a_1 \cdot a_2 + a_3 \cdot a_4 + \frac{(a_4 - a_3)^2}{3} - \frac{1}{3} \cdot \frac{(a_2 - a_1)^2}{-a_1 - a_2 + a_3 + a_4} }

Where \mjseqn{crisp_a} ... is crisp number for number \mjseqn{a = (a_1, a_2, a_3, a_4)}.
}
\examples{

## ------------------------------------------------
## Method `fuzzyvikor$new`
## ------------------------------------------------

dictionaryW <- rbind(
  c(0, 0, 0.1, 0.2),
  c(0.1, 0.2, 0.2, 0.3),
  c(0.2, 0.3, 0.4, 0.5),
  c(0.4, 0.5, 0.5, 0.6),
  c(0.5, 0.6, 0.7, 0.8),
  c(0.7, 0.8, 0.8, 0.9),
  c(0.8, 0.9, 1, 1)
)
# V very, L low, M medium, H high
rownames(dictionaryW) <- c('VL', 'L', 'ML', 'M', 'MH', 'H', 'VH')
w <- rbind(
  c('H', 'VH', 'VH'),
  c('M', 'H', 'VH'),
  c('M', 'MH', 'ML'),
  c('H', 'VH', 'MH')
)
 rownames(w) <- c('Investment cost', 'Employment needs', 'Social impact',
 'Environmental impact')
dictionaryPM <- rbind(
  c(0, 0, 0.1, 0.2),
  c(0.1, 0.2, 0.2, 0.3),
  c(0.2, 0.3, 0.4, 0.5),
  c(0.4, 0.5, 0.5, 0.6),
  c(0.5, 0.6, 0.7, 0.8),
  c(0.7, 0.8, 0.8, 0.9),
  c(0.8, 0.9, 1, 1)
)
# V very, P poor, M medium, F fair, G good
rownames(dictionaryPM) <- c('VP', 'P', 'MP', 'F', 'MG', 'G', 'VG')
PM <- rbind(
  c('VG', 'G', 'MG'),
  c('MP', 'F', 'F'),
  c('MG', 'MP', 'F'),
  c('MG', 'MG', 'VG'),
  c('VP', 'P', 'G'),
  c('F', 'G', 'G'),
  c('F', 'MG', 'MG'),
  c('F', 'VG', 'G'),
  c('MG', 'MG', 'VG'),
  c('G', 'G', 'VG'),
  c('P', 'VP', 'MP'),
  c('F', 'MP', 'MG'),
  c('P', 'P', 'MP'),
  c('MG', 'VG', 'G'),
  c('MP', 'F', 'F'),
  c('MG', 'VG', 'G'),
  c('G', 'G', 'VG'),
  c('VG', 'MG', 'F'),
  c('G', 'VG', 'G'),
  c('MG', 'F', 'MP'),
  c('MP', 'P', 'P'),
  c('VP', 'F', 'P'),
  c('G', 'MG', 'MG'),
  c('P', 'MP', 'F')
)
alternatives <- c('site 1', 'site 2', 'site 3', 'site 4', 'site 5',
                  'site 6')
 result <- fuzzyvikor$new(PM, dictionaryPM, w, dictionaryW, alternatives)
}
\references{
ALAOUI, Mohamed El. Fuzzy TOPSIS: Logic, Approaches, and Case Studies. Boca
 Raton: CRC Press, 2021. 216 s. ISBN 978-0-367-76748-8.

Papathanasiou, Jason, Ploskas, Nikolaos. Multiple Criteria Decision Aid
 Methods, Examples and Python Implementations. Springer, 173 p.,
 ISBN 978-3-319-91648-4.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{FuzzyVIKOR}
\keyword{VIKOR}
\keyword{fuzzy}
\keyword{set}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{criteria}}{vector of criterias}

\item{\code{alt}}{vector of alternative's names}

\item{\code{w}}{weights (matrix of weights - decision makers in columns and
criteria in rows)}

\item{\code{pm}}{performance matrix with n columns and no. criteria x no. of
alternatives rows. I. e. with 3 criteria and 2 alternatives the rows
are: 1: 1.cri-1.alt., 2: 1.cri.-2.alt, 3: 2.cri-1.alt, 4: 2.cri-2.alt,
...}

\item{\code{v}}{weight for strategy of majority of the criteria (0-1)}

\item{\code{dictionary_pm}}{dictionary of linguistic variables for criteria}

\item{\code{dictionary_w}}{dictionary for wights (single matrix|dataframe)}

\item{\code{S}}{list of alternatives using S-metric}

\item{\code{R}}{list of alternatives using R-metric}

\item{\code{Q}}{list of alternatives using Q-metric}

\item{\code{result}}{dataframe with S, R, Q metrics and its accompanying ranking
Q represents balance between S and R. All metrics are minimized}

\item{\code{compromiseSolution}}{compromise solution based on S, R and Q metrics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fuzzyvikor-new}{\code{fuzzyvikor$new()}}
\item \href{#method-fuzzyvikor-compute}{\code{fuzzyvikor$compute()}}
\item \href{#method-fuzzyvikor-summary}{\code{fuzzyvikor$summary()}}
\item \href{#method-fuzzyvikor-clone}{\code{fuzzyvikor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzyvikor-new"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzyvikor-new}{}}}
\subsection{Method \code{new()}}{
public constructor for fuzzyvikor R6 class. Checks validity of the
 provided parameters. Initializes model and computes it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzyvikor$new(pm, dictionary_pm, w, dictionary_w, alt, v = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix with n columns and no. criteria x no. of
alternatives rows. I. e. with 3 criteria and 2  alternatives the rows
are: 1: 1.cri-1.alt., 2: 1.cri.-2.alt, 3: 2.cri-1.alt,
4: 2.cri-2.alt, ...}

\item{\code{dictionary_pm}}{- dictionary of linguistic variables for criteria
(single matrix)}

\item{\code{w}}{weights (matrix of weights - decision makers in columns and
criteria in rows)}

\item{\code{dictionary_w}}{dictionary for wights (single matrix|dataframe)}

\item{\code{alt}}{vector with names of the alternatives}

\item{\code{v}}{weight for strategy of majority of the criteria (0-1)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dictionaryW <- rbind(
  c(0, 0, 0.1, 0.2),
  c(0.1, 0.2, 0.2, 0.3),
  c(0.2, 0.3, 0.4, 0.5),
  c(0.4, 0.5, 0.5, 0.6),
  c(0.5, 0.6, 0.7, 0.8),
  c(0.7, 0.8, 0.8, 0.9),
  c(0.8, 0.9, 1, 1)
)
# V very, L low, M medium, H high
rownames(dictionaryW) <- c('VL', 'L', 'ML', 'M', 'MH', 'H', 'VH')
w <- rbind(
  c('H', 'VH', 'VH'),
  c('M', 'H', 'VH'),
  c('M', 'MH', 'ML'),
  c('H', 'VH', 'MH')
)
 rownames(w) <- c('Investment cost', 'Employment needs', 'Social impact',
 'Environmental impact')
dictionaryPM <- rbind(
  c(0, 0, 0.1, 0.2),
  c(0.1, 0.2, 0.2, 0.3),
  c(0.2, 0.3, 0.4, 0.5),
  c(0.4, 0.5, 0.5, 0.6),
  c(0.5, 0.6, 0.7, 0.8),
  c(0.7, 0.8, 0.8, 0.9),
  c(0.8, 0.9, 1, 1)
)
# V very, P poor, M medium, F fair, G good
rownames(dictionaryPM) <- c('VP', 'P', 'MP', 'F', 'MG', 'G', 'VG')
PM <- rbind(
  c('VG', 'G', 'MG'),
  c('MP', 'F', 'F'),
  c('MG', 'MP', 'F'),
  c('MG', 'MG', 'VG'),
  c('VP', 'P', 'G'),
  c('F', 'G', 'G'),
  c('F', 'MG', 'MG'),
  c('F', 'VG', 'G'),
  c('MG', 'MG', 'VG'),
  c('G', 'G', 'VG'),
  c('P', 'VP', 'MP'),
  c('F', 'MP', 'MG'),
  c('P', 'P', 'MP'),
  c('MG', 'VG', 'G'),
  c('MP', 'F', 'F'),
  c('MG', 'VG', 'G'),
  c('G', 'G', 'VG'),
  c('VG', 'MG', 'F'),
  c('G', 'VG', 'G'),
  c('MG', 'F', 'MP'),
  c('MP', 'P', 'P'),
  c('VP', 'F', 'P'),
  c('G', 'MG', 'MG'),
  c('P', 'MP', 'F')
)
alternatives <- c('site 1', 'site 2', 'site 3', 'site 4', 'site 5',
                  'site 6')
 result <- fuzzyvikor$new(PM, dictionaryPM, w, dictionaryW, alternatives)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzyvikor-compute"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzyvikor-compute}{}}}
\subsection{Method \code{compute()}}{
computes the Fuzzy VIKOR problem. Usually doesn't need to be run manually
 as it is called automatically by class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzyvikor$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzyvikor-summary"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzyvikor-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the Fuzzy VIKOR method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzyvikor$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzyvikor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzyvikor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzyvikor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
