% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Grey relational analysis}
\description{
One of most well known application of grey theory. It defines situations
 with no information as black, and those with perfect information as white.
 However, neither of these idealized situations ever occurs in real world
 problems. In fact, situations between these extremes, which contain partial
 information, are described as being grey. (from Wikipedia).

Computationally the GRA expects performance matrix to be in numeric form.
 Values of performance of the alternatives in criteria are normalized using
 min-max normalization.

Then deviation sequence k is computed

\mjsdeqn{k = max_i - n_{ij} = 1 - n_{ij}}

Computatioon of k is simplified due to usage of min-max normalization. Using
 this type of normalization leads always to max = 1 and min = 0, n
 represents normalized value.

Then Grey relational coefficient (GRC) is computed as

\mjsdeqn{GRC = \frac{\Delta_{min} + DC \Delta_{max}}{k_{ij} + DC \Delta_{max}} = \frac{0 + DC \cdot 1}{k_{ij} + DC \cdot 1} = \frac{DC}{k_{ij} + DC}}

DC is distinguished coefficient (default value: 0.5). Provided
 simplification of the above equation holds only if min-max normalization is
 used. If other form of normalization is used then min and max values can be
 expected to differ from 0 and 1.

GRC then can be processed into gray relational grade

\mjsdeqn{GRA = \frac{\sum_{i=1}^m GRC_i}{m}}

Where m is number of criteria.
}
\examples{

## ------------------------------------------------
## Method `gra$new`
## ------------------------------------------------

# from: https://www.youtube.com/watch?v=J63Q5IbK-cs
alternatives <- c("car 1", "car 2", "car 3", "car 4", "car 5", "car 6",
 "car 7", "car 8", "car 9", "car 10")
criteria <- c("quality", "condition", "security", "delivery days",
 "fuel consumption", "price")
M <- rbind(
  c(3, 6, 4, 20, 2, 30000),
  c(4, 4, 6, 15, 2.2, 32000),
  c(6, 5, 9, 18, 3, 32100),
  c(5, 6, 3, 23, 2.8, 28000),
  c(4, 8, 7, 30, 1.5, 29000),
  c(8, 3, 6, 35, 1.9, 27000),
  c(7, 2, 5, 33, 1.7, 28500),
  c(3, 8, 3, 34, 1.6, 30500),
  c(8, 4, 8, 40, 2.5, 33000),
  c(9, 3, 7, 34, 2, 29800)
)
rownames(M) <- alternatives
colnames(M) <- criteria
minmax <- c("max", "max", "max", "min", "min", "min")
t <- gra$new(M, minmax)
}
\references{
How to use Grey Relational Analysis to decide? Available from
 \url{https://www.youtube.com/watch?v=J63Q5IbK-cs}.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{GRA}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not normalized) performance matrix}

\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights (sum of weights = 1)}

\item{\code{dc}}{distinguished coefficient (0.5 by default)}

\item{\code{gra}}{results of GRA analysis (average Grey relational coefficient)}

\item{\code{gra_sorted}}{ordered results of GRA analysis (average Grey
relational coefficient). Sorted form best to worst.}

\item{\code{grc}}{Grey relational coefficient}

\item{\code{minmax}}{vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{result}}{datagrame with GRA and ranks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gra-new}{\code{gra$new()}}
\item \href{#method-gra-compute}{\code{gra$compute()}}
\item \href{#method-gra-summary}{\code{gra$summary()}}
\item \href{#method-gra-interaction_diagram}{\code{gra$interaction_diagram()}}
\item \href{#method-gra-clone}{\code{gra$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gra-new"></a>}}
\if{latex}{\out{\hypertarget{method-gra-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gra$new(pm, minmax = "max", dc = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}

\item{\code{dc}}{distinguished coefficient (0.5 by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# from: https://www.youtube.com/watch?v=J63Q5IbK-cs
alternatives <- c("car 1", "car 2", "car 3", "car 4", "car 5", "car 6",
 "car 7", "car 8", "car 9", "car 10")
criteria <- c("quality", "condition", "security", "delivery days",
 "fuel consumption", "price")
M <- rbind(
  c(3, 6, 4, 20, 2, 30000),
  c(4, 4, 6, 15, 2.2, 32000),
  c(6, 5, 9, 18, 3, 32100),
  c(5, 6, 3, 23, 2.8, 28000),
  c(4, 8, 7, 30, 1.5, 29000),
  c(8, 3, 6, 35, 1.9, 27000),
  c(7, 2, 5, 33, 1.7, 28500),
  c(3, 8, 3, 34, 1.6, 30500),
  c(8, 4, 8, 40, 2.5, 33000),
  c(9, 3, 7, 34, 2, 29800)
)
rownames(M) <- alternatives
colnames(M) <- criteria
minmax <- c("max", "max", "max", "min", "min", "min")
t <- gra$new(M, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gra-compute"></a>}}
\if{latex}{\out{\hypertarget{method-gra-compute}{}}}
\subsection{Method \code{compute()}}{
Computed the GRA using params provided in constructor. Usually run
 automatically by constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gra$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gra-summary"></a>}}
\if{latex}{\out{\hypertarget{method-gra-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the GRA method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gra$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gra-interaction_diagram"></a>}}
\if{latex}{\out{\hypertarget{method-gra-interaction_diagram}{}}}
\subsection{Method \code{interaction_diagram()}}{
Plots interaction diagram  between the criteria based on Grey
 relational coefficient (GRC).

Not usefull for visializing GRA results using large ammount of criteria.
 Reasonable maximal number of criteria is around 6. Increasing number
 of criteria increases number of diagrams ploted into constant size plot
 and decrease in readibility as result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gra$interaction_diagram()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gra-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gra-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gra$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
