% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2AdjancancyMatrix.R
\name{Graph2AdjancancyMatrix}
\alias{Graph2AdjancancyMatrix}
\title{Transforms graph (a->b format) to adjacency matrix}
\usage{
Graph2AdjancancyMatrix(G, alt)
}
\arguments{
\item{G}{Graph which should be transformed}

\item{alt}{alternatives names}
}
\value{
adjacency matrix for the graph with 0/1 representing graph.
}
\description{
Function transforms back graph specified by edges to adjacency matrix, with
 1 where edge exists and 0 where it does not.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{adjacency}
\keyword{graph}
\keyword{matrix}
