% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mairca.R
\name{mairca}
\alias{mairca}
\title{Multi-Attribute Ideal-Real Comparative Analysis}
\description{
MAIRCA evaluations is argued in sense of gaps between these ideal
 solutions and the real performance of the alternatives. The better the
 alternative is, the lower the gap should be.

We start the process by formulating performance matrix PM as usual (criteria
 in colums and alternatives in rows). We set the preferences for decision
 maker PA to 1/m, where m in number of alternatives.

Next perform theoretical evaluation by applying weights to PA. To establish
 real evaluation matrix we compare performance of the alternative
 in the criterium with ideal (or antiideal solution) and divide by the
 difference between ideal and anti-ideal solution.

For benefit criteria we do that by:

\mjsdeqn{t_{rij} = t_{pij}(\frac{x_{ij} - x_i^-}{x_i^+ - x_i^-})}

For cost criteria:

\mjsdeqn{t_{rij} = t_{pij}(\frac{x_{ij} - x_i^+}{x_i^- - x_i^+})}

where xij ... elements of performance matrix, xi+ ... ideal solution, xi-
 ... andtiideal solution and tpij are weighted preferences.

We compute gap matrix G by comparing weighted preferences and results of
 real evaluation (tpij - trij). If tpij == trij then gij = 0, if tpij > trij
 then gij = tpij - trij.

We aggregate gap matrix across the criteria to establish final values of
 criteria functions:

\mjsdeqn{Q_i = \sum_{j=1}^n g_{ij}}

This measure is directly usable for ranking purposes. Since we are
 minimizing gaps we are also minimizing Qi.
}
\examples{

## ------------------------------------------------
## Method `mairca$new`
## ------------------------------------------------

cri <- c("C1", "C2", "C3")
alt <- c("A1", "A2", "A3", "A4", "A5")
pm <- cbind(
  c(5, 10, 4, 8, 1),
  c(4, 8, 2, 16, 4),
  c(7, 7, 6, 8, 5)
)
rownames(pm) <- alt
colnames(pm) <- cri
minmax <- c("max", "max", "min")
w <- c(0.25, 0.33, 0.42)
result <- mairca$new(pm, w, minmax)
}
\references{
GIGOVIĆ, Ljubomir, PAMUČAR, Dragan, BAJIĆ, Zoran, MILIĆEVIĆ, Milić. The
 Combination of Expert Judgment and GIS-MAIRCA Analysis for the Selection of
 Sites for Ammunition Depots. Sustainability. 2016, vol. 8, no. 4, available
 from: https://doi.org/10.3390/su8040372, ISSN 2071-1050.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MAICRA}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows and criteria in
columns)}

\item{\code{w}}{weight vector}

\item{\code{minmax}}{vector of optimization directions. Use only values of min
or max. Single min/max will apply chosen direction to all criteria}

\item{\code{gap}}{gap matrix describing gap between the preference a real
performance of the alternative}

\item{\code{Q}}{vector of final values of criteria functions for alternatives}

\item{\code{rank}}{vector of ranks based on Q}

\item{\code{result}}{data frame with gap, Q and rank}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mairca-new}{\code{mairca$new()}}
\item \href{#method-mairca-compute}{\code{mairca$compute()}}
\item \href{#method-mairca-summary}{\code{mairca$summary()}}
\item \href{#method-mairca-clone}{\code{mairca$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mairca-new"></a>}}
\if{latex}{\out{\hypertarget{method-mairca-new}{}}}
\subsection{Method \code{new()}}{
public constructor allowing the user to construct Borta count decision
 analysis problem and compute it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mairca$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion. only
numeric values expected. Rows and columns are expected to be named.}

\item{\code{w}}{weight vector}

\item{\code{minmax}}{value or vector of values 'min' or 'max' specifying
optimization direction for the criterium}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialized R6 class with computed results for MAIRCA
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cri <- c("C1", "C2", "C3")
alt <- c("A1", "A2", "A3", "A4", "A5")
pm <- cbind(
  c(5, 10, 4, 8, 1),
  c(4, 8, 2, 16, 4),
  c(7, 7, 6, 8, 5)
)
rownames(pm) <- alt
colnames(pm) <- cri
minmax <- c("max", "max", "min")
w <- c(0.25, 0.33, 0.42)
result <- mairca$new(pm, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mairca-compute"></a>}}
\if{latex}{\out{\hypertarget{method-mairca-compute}{}}}
\subsection{Method \code{compute()}}{
Computes MAIRCA model based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mairca$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mairca-summary"></a>}}
\if{latex}{\out{\hypertarget{method-mairca-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the MAIRCA count method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mairca$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mairca-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mairca-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mairca$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
