% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mare.R
\name{mare}
\alias{mare}
\title{Multi Attribute Range Evaluation}
\description{
Interesting extension of \link{wsm} method. Internally it uses WSM for its
 recommendations, it allows to express the uncertainties as the range the
 true performance measure is in.

Other methods use fuzzy logic for the purpose, MARE approaches the problem
 by just manually setting the range manually. The method then computes WSM
 3-times: 1) for normal measures, 2) minimal measures and 3) maximal
 measures.
}
\references{
Richard E. Hodgett, Elaine B. Martin, Gary Montague, Mark
Talford (2014). Handling uncertain decisions in whole process design.
Production Planning & Control, Volume 25, Issue 12, 1028-1038.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{pm_min}}{minimal bounds for sample in pm}

\item{\code{pm_max}}{maximal bounds for sample in pm}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{result}}{data frame with results of the method}

\item{\code{resultPlot}}{visualisation of the results with uncertainties}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mare-new}{\code{mare$new()}}
\item \href{#method-mare-compute}{\code{mare$compute()}}
\item \href{#method-mare-plotResult}{\code{mare$plotResult()}}
\item \href{#method-mare-summary}{\code{mare$summary()}}
\item \href{#method-mare-clone}{\code{mare$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mare-new"></a>}}
\if{latex}{\out{\hypertarget{method-mare-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mare$new(pm, pm_min, pm_max, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{pm_min}}{minimal bounds for sample in pm}

\item{\code{pm_max}}{maximal bounds for sample in pm}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mare-compute"></a>}}
\if{latex}{\out{\hypertarget{method-mare-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of MARE model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mare$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mare-plotResult"></a>}}
\if{latex}{\out{\hypertarget{method-mare-plotResult}{}}}
\subsection{Method \code{plotResult()}}{
plot visualisation of the estimate with min and max as error estimates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mare$plotResult()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mare-summary"></a>}}
\if{latex}{\out{\hypertarget{method-mare-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the MARE method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mare$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mare-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mare-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mare$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
