% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcda_method.R
\name{mcda_method}
\alias{mcda_method}
\title{Apply MCDA method to the decision problem}
\usage{
mcda_method(pm, w, minmax = "max", method = "wsm")
}
\arguments{
\item{pm}{performance matrix criteria in columns, alternatives in rows,
numeric values expressed}

\item{w}{weights vector}

\item{minmax}{'min' or 'max' to specify cost or benefit criterion, max is
default value}

\item{method}{specify MCDA method, use constant from table above,
the 'wsm' is default value.}
}
\value{
class with computed decision problem
}
\description{
Helper function for application of the various MCDA methods.

The pasckage implements MCDA methods as R6 classes, which need to be
 initiated using class constructor, usually in form of name$new(parameters).
 the function allow users to abstract a little bit.

The parameters will apply the method specified in method parameter and
 compute the decision problem. Method then returns full fledged decision
 object, which then can be used in same way as if constructor of the method
 had been called directly.

Supported methods:

\tabular{ll}{
       \bold{constant} \tab \bold{method} \cr
       \link{aras} \tab Additive Ratio Assessment \cr
       \link{borda} \tab Borda voting system \cr
       \link{copeland} \tab Copeland's AKA Llull method \cr
       \link{copras} \tab Complex Proportional Assessment \cr
       \link{cradis} \tab Compromise Ranking of Alternatives from Distance to Ideal Solution \cr
       \link{edas} \tab Evaluation Based on Distance from Average Solution \cr
       \link{fuca} \tab Faire Un Choix Adégnat \cr
       \link{mabac} \tab Multi-Attribute Border Approximation are Comparison \cr
       \link{mairca} \tab Multi-Attribute Ideal-Real Comparative Analysis \cr
       \link{marcos} \tab Measurement of Alternatives and Ranking According to Compromise Solution \cr
       \link{maut} \tab Multi-Attribute Utility Theory \cr
       \link{moora} \tab Multi-Objective Optimization on the basis of Ratio Analysis \cr
       \link{moosra} \tab Multi-Objective Optimization on the basis of Simple Ratio Analysis \cr
       \link{msim} \tab Modified Similarity Index \cr
       \link{ocra} \tab Operational Competitiveness  Ratings  Analysis \cr
       \link{piv} \tab Proximity Indexed Value \cr
       \link{probid} \tab Preference Ranking on the Basis of Ideal-Average Distance \cr
       \link{psi} \tab Preference Selection Index \cr
       \link{ram} \tab Root Assesment Method \cr
       \link{rawec} \tab Ranking of Alternatives with Weights of Criterion \cr
       \link{regime} \tab REGIME \cr
       \link{rov} \tab Range of Value Method \cr
       \link{saw} \tab Simple Additive Weight \cr
       \link{sprobid} \tab Simplified Preference Ranking on the Basis of Ideal-Average Distance \cr
       \link{todim} \tab TOmada de Decisao Interativa Multicriterio \cr
       \link{topsis} \tab TOPSIS decision support method \cr
       \link{wpm} \tab Weighted Product Method \cr
       \link{wsm} \tab Weighted Sum method \cr
   }

Please note that this function does not support all MCDA methods implemented
 in the package. See the documentation for full list of supported methods.
}
\examples{
alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
t <- mcda_method(pm = M, w = w, minmax = "max", method = "wsm")
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
