% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcda_pairwise_weights.R
\name{mcda_pairwise_weights}
\alias{mcda_pairwise_weights}
\title{Compute weights based on pairwise comparison}
\usage{
mcda_pairwise_weights(pm, method)
}
\arguments{
\item{pm}{preference matrix n:n, where n is number of criteria}

\item{method}{constant (see the table to identify the method to use)}
}
\value{
list with named weights vector and CR (for AHP only)
}
\description{
Function allows to apply various method which derive weights based on stated
 preferences, when comparing every pair of criteria in decision problem.

At present time the function computes:
\tabular{ll}{
       \bold{constant} \tab \bold{method}\cr
       AHP \tab Analytic Hierarchy Process\cr
       binary \tab Binary pairwise camparison weight estimation\cr
       RANCOM \tab RANking COMparison\cr
   }

\bold{Binary pairwise camparison weight estimation}

Derives weights of the criteria using binary preferences of importance
 of the criteria by comparing each pair of criteria and choosing one more
 important of the pair.

Idea is that the criteria, which are more important are preferred more often
 than those, which are less important.

Weights are computed by counting number of times the criterium has been
 preferred and comparing the value to number of comparisons realized. The
 result is then normalized using vestor normalization to derive weights.

The approach is very similar to other methods such as Fuller's triangle.

\bold{RANCOM - RANking COMparison}

Very similar method to binary pairwise comparison only with slightly
 differently expressed preferences. In RANCOM the preference is 1 if
 f(Ci) < f(Cj), 0 if f(Ci) < f(Cj) and 0.5 if f(Ci) == f(Cj).

The preferences are then summed up on per row basis and weights derived by
 vector normalizing the result.

\bold{AHP - Analytic Hierarchy Process}

AHP is one of most widely used methods for MCDA and weight derivation based
 on stated preferences. 

This function implements only single preference matrix computation, so it
 does not implement hierarchy at all. 

The preferences are stated in interval 1 (equal) to 9 (extremely more inportrant)
 and 1/9 (extremely less important) to 1.

One of interesting properties of AHP is ability to evaluate randomnes of
 stated preferences by comparing its consistency index (CI) to average random
 consistency index (RCI). 

The weights are derived using geometric mean method.

\mjsdeqn{CI = \frac{\lambda_{mmax} - n}{n - 1}}

Lambda max is largest eigenvalue of the preference matrix, n is number of criteria

The ratio (CR - Consistency Ration) of CI/RCI should be under 0.1 to reject
 hyphothesis, that the stated preferences has been stated randomly. Function
 allows to perform this evaluation maximally for 15 criteria in single matrix.
 In case you need more, use hierarchy.
}
\examples{
PM <- rbind(
   c(0.5,1,1,0,1,1),
   c(0,0.5,0,0,1,1),
   c(0,1,0.5,0,1,1),
   c(1,1,1,0.5,1,1),
   c(0,0,0,0,0.5,1),
   c(0,0,0,0,0,0.5)
)
rownames(PM) <- colnames(PM) <- c("C1", "C2", "C4", "C5", "C6", "C8")
t <- mcda_pairwise_weights(PM, method = "RANCOM")
}
\references{
ŠENOVSKÝ, Pavel. Modelling of Decision Procesess. 4th edition, VŠB - TU
 Ostrava, Ostrava 2024, 195 str., available from
 \url{https://fbiweb.vsb.cz/~sen76/data/uploads/skripta/modelling4ed.pdf}
 [cit. 2024-09-15]

WIĘCKOWSKI, Jakub, KIZIELEWICZ, Bartłomiej, SHEKHOVTSOV, Andrii, SAŁABUN, Wojciech.
 RANCOM: A novel approach to identifying criteria relevance based on inaccuracy
 expert judgments. Engineering Applications of Artificial Intelligence. 2023,
 vol. 122, p. 106114, available from: \url{https://doi.org/10.1016/j.engappai.2023.106114},
 ISSN 0952-1976.

Řehák, D., Šenovský, P. Preference risk assessment of electric power critical
 infrastructure. In Chemical Engineering Transactions, Vol. 36, pp. 469-474,
 DOI 10.3303/CET1436079, ISSN: 1974-9791
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
