% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merec.R
\name{merec}
\alias{merec}
\title{METthod based on Removal Effects of Criteria}
\description{
simple method for weight estimation. Works in six steps. Start with (1)
 assembling performance matrix with alternatives in rows and criteria in
 columns.

Next we need to normalize the PM (2). For benefit criteria we use:

\mjsdeqn{n_{ij} = \frac{min_k x_{kj}}{x_{ij}}}

for cost criteria

\mjsdeqn{n_{ij} = \frac{x_{ij}}{max_k x_{kj}}}

Then we compute overal performance of overal performance of the alternatives
 Si (step 3)

\mjsdeqn{S_i = ln(1 + (\frac{1}{m} \sum_j |ln(n_{ij})|)}

In step 4 the calculation of performance alternatives by removig each
 attribute (Sij') is performed

\mjsdeqn{S_{ij}^\prime = ln(1 + (\frac{1}{m} \sum_{k,k \ne j} |ln(n_{ik})|))}

Then a Removal effect (Ej) is calculated (step 5).

\mjsdeqn{E_j = \sum_j |S_{ij}^\prime - S_i|}

Finally the final weights (wj) are coputed in step 6.

\mjsdeqn{w_j = \frac{E_j}{\sum_k E_k}}
}
\examples{

## ------------------------------------------------
## Method `merec$new`
## ------------------------------------------------

#see Bhanale et al. for the reference
alternatives <- c("ASEA-IRB 60/2", "Cincinnati Milacrone T3-726",
 "Cybotech V15 Electric Robot", "Hitachi America Process Robot",
 "Unimation PUMA 500/600", "United States Robots Maker 110",
 "Yaskawa Electric Motoman L3C")
criteria <- c("Load capacity (LC) [kg]", "Repeatability (RE) [mm]",
 "Maximum tip speed (MPS) [mm/s]", "Memory capacity (MC)",
 "Manipulator reach (MR) [mm]")
pm <- data.frame(
 c(60, 6.35, 6.8, 10, 2.5, 4.5, 3),
 c(0.4, 0.15, 0.1, 0.2, 0.1, 0.08, 0.1),
 c(2540, 1016, 1727.2, 1000, 560, 1016, 177),
 c(500, 3000, 1500, 2000, 500, 350, 1000),
 c(990, 1041, 1676, 965, 915, 508, 920)
)
colnames(pm) <- criteria
rownames(pm) <- alternatives
minmax <- c("max", "min", "max", "max", "max")
t <- merec$new(pm, minmax)
}
\references{
Bhangale, P. P., Agrawal, V. P., & Saha, S. K. (2004). Attribute based
 specification, comparison and selection of a robot. Mechanism and Machine
 Theory, 39(12), 1345-1366.

Keshavarz-Ghorabaee, M., Amiri, M., Zavadskas, E. K., Turskis, Z., &
 Antucheviciene, J. (2021). Determination of objective weights using a new
 method based on the removal effects of criteria (MEREC). Symmetry, 13(4),
 525.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MEREC}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not normalized) performance matrix}

\item{\code{pm}}{normalized performance matrix}

\item{\code{minmax}}{criteria optimization direction (min or max). Can be
substituted for single min or max value if all criteria share same
optimization direction.}

\item{\code{w}}{estimated weights}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-merec-new}{\code{merec$new()}}
\item \href{#method-merec-compute}{\code{merec$compute()}}
\item \href{#method-merec-clone}{\code{merec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merec-new"></a>}}
\if{latex}{\out{\hypertarget{method-merec-new}{}}}
\subsection{Method \code{new()}}{
Initializes and computes MEREC model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{merec$new(pm, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{minmax}}{criteria optimization direction (min or max). Can be
substituted for single min or max value if all criteria share same
optimization direction.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
computed MEREC model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#see Bhanale et al. for the reference
alternatives <- c("ASEA-IRB 60/2", "Cincinnati Milacrone T3-726",
 "Cybotech V15 Electric Robot", "Hitachi America Process Robot",
 "Unimation PUMA 500/600", "United States Robots Maker 110",
 "Yaskawa Electric Motoman L3C")
criteria <- c("Load capacity (LC) [kg]", "Repeatability (RE) [mm]",
 "Maximum tip speed (MPS) [mm/s]", "Memory capacity (MC)",
 "Manipulator reach (MR) [mm]")
pm <- data.frame(
 c(60, 6.35, 6.8, 10, 2.5, 4.5, 3),
 c(0.4, 0.15, 0.1, 0.2, 0.1, 0.08, 0.1),
 c(2540, 1016, 1727.2, 1000, 560, 1016, 177),
 c(500, 3000, 1500, 2000, 500, 350, 1000),
 c(990, 1041, 1676, 965, 915, 508, 920)
)
colnames(pm) <- criteria
rownames(pm) <- alternatives
minmax <- c("max", "min", "max", "max", "max")
t <- merec$new(pm, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merec-compute"></a>}}
\if{latex}{\out{\hypertarget{method-merec-compute}{}}}
\subsection{Method \code{compute()}}{
Compute the MEREC using params provided in constructor. Usually run
 automatically by constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{merec$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-merec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{merec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
