% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocra.R
\name{ocra}
\alias{ocra}
\title{Operational Competitiveness  Ratings  Analysis}
\description{
MCDA ranking method based on separate evaluation of the cost and benefit
 criteria.

After formulating performance matrix, preference ratings with respect to
 cost criteria are determined:

\mjsdeqn{\overline{I_i} = \sum_{k=1}^q w_k \frac{max(x_i^k) - x_i^k}{min(x_i^k)}}

Where q is number fo cost criteria, xik is performance score of i-th
 alternative in k-th criterium and w is the weight of criterium.

Then linear preference rating for cost criteria is computed:

\mjsdeqn{\overline{\overline{I_i}} = \overline{I_i} - min(\overline{I_i})}

Same approach is used for benefit criteria:

\mjsdeqn{\overline{O_i} = \sum_{h=1}^b w_h \frac{x_i^h - min(x_i^h)}{min(x_i^h)}}

\mjsdeqn{\overline{\overline{O_i}} = \overline{O_i} - min(\overline{O_i})}

Finally overall performance ratings of competitive alternatives is computed:

\mjsdeqn{P_i = \overline{\overline{I_i}} + \overline{\overline{O_i}} - min(\overline{\overline{I_i}} + \overline{\overline{O_i}})}

The alternative with highest overall performance is best.
}
\examples{

## ------------------------------------------------
## Method `ocra$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- ocra$new(pm, w = w, minmax = minmax)
}
\references{
Miloš Madić, Dušan Petković and Miroslav Radovanović. SELECTION OF
 NON-CONVENTIONAL MACHINING PROCESSES USING THE OCRA METHOD. Serbian Journal
 of Management 10 (1) (2015) 61 - 73, DOI:10.5937/sjm10-6802
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{OCRA}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives (rpws) in criteria (columns)}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{w}}{weight vector}

\item{\code{Pi}}{overall preference rating (higher better)}

\item{\code{rank}}{ranked alternatives, based on Pi}

\item{\code{result}}{data frame with preference rating and ranking}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ocra-new}{\code{ocra$new()}}
\item \href{#method-ocra-compute}{\code{ocra$compute()}}
\item \href{#method-ocra-summary}{\code{ocra$summary()}}
\item \href{#method-ocra-clone}{\code{ocra$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ocra-new"></a>}}
\if{latex}{\out{\hypertarget{method-ocra-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ocra$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- ocra$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ocra-compute"></a>}}
\if{latex}{\out{\hypertarget{method-ocra-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of OCRA model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ocra$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ocra-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ocra-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the OCRA method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ocra$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ocra-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ocra-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ocra$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
