% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_check_marcos.R
\name{param_check_marcos}
\alias{param_check_marcos}
\title{parameter checking form MARCOS function (internal function)}
\usage{
param_check_marcos(pm, w, minmax = "max")
}
\arguments{
\item{pm}{original (not normalized) performance matrix}

\item{w}{weights of the criteria}

\item{minmax}{vector of criteria direction to be used during
performance matrix normalization. Can be replaced with single max or
min if all criteria are to be maximized or minimized. Implicitly set
to max.}
}
\value{
minmax vector (for situations when the minmax parameters is set as
 single min or max value)
}
\description{
function perform validation of the provided parameters. The validation is
 used by both \link{marcos} and \link{cradis} functions.

If all validations yield no error, the function returns adjusted minmax
 vector. If the error is detected then the run of the process is stopped
 with appropriate error message.

Function is intended for internal use only.
}
\keyword{CRADIS}
\keyword{MARCOS}
