% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_check_wsm.R
\name{param_check_wsm}
\alias{param_check_wsm}
\title{parameter check for WSM and WSM-like methods (internal)}
\usage{
param_check_wsm(pm, w, minmax = "max")
}
\description{
The function checks parameters for \link{wsm} function.

Package has implemented some other functions which share these checks, such
 as \link{saw} and the function helps to keep the code DRY.

Function is intended for internal use only.
}
\keyword{SAW}
\keyword{WSM}
