% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piprecia.R
\name{piprecia}
\alias{piprecia}
\title{Pivot Pairwise Relative Criteria Importance Assessment}
\usage{
piprecia(relative_weights, criteria)
}
\arguments{
\item{relative_weights}{dataframe of relative weights of the criteria to
first criterium. Each column has a preference of one decision-maker.}

\item{criteria}{dataframe of criteria names ordered in descending manner.
Each decision maker in separate column}
}
\value{
vector of normalized weights for criteria
}
\description{
New spin on \link{swara} method. PIPRECIA is believed to be more effective
 in complex decision-making situations.

First five steps follow SWARA method. The 6-th step allows for aggregation
 of the weights derived from preferences of individual decision-makers. 

\mjsdeqn{w_j^* = (\prod_{m = 1}^M w_j^m)^{1/M}}

\mjsdeqn{w_j = \frac{w_j^*}{\sum_{j = 1}^m w_j^*}}
}
\examples{
criteria <- data.frame(
  DM1 = c("C1", "C2", "C3", "C4"),
  DM2 = c("C1", "C3", "C2", "C4"),
  DM3 = c("C1", "C3", "C4", "C2")
)
rel_weights <- data.frame(
  DM1 = c(NA, 0.48, 0.35, 0.22),
  DM2 = c(NA, 0.5, 0.4, 0.3),
  DM3 = c(NA, 0.4, 0.3, 0.2)
)
t <- piprecia(rel_weights, criteria)

}
\references{
Chakrakborty, S., Chatterjee, P. Das, P. P. Multi-Criteria Decision Making
 Methods in Manufacturing Environments: Models and Applications. CRC Press,
 Boca Raton, 2024, 450 p., ISBN: 978-1-00337-703-0
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{PIPRECIA}
\keyword{SWARA}
\keyword{weights}
