% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scoreM.R
\name{plot.scoreM}
\alias{plot.scoreM}
\title{Function to plot contribution of the criteria to overall performance of the
 alternatives.}
\usage{
\method{plot}{scoreM}(x, title = "Overal score of the variants")
}
\arguments{
\item{x}{weighted preference matrix - values in matrix used to descibe
strength of the relation.}

\item{title}{title to be used for the graph. "Overall score of the variants"
by default}
}
\value{
stacked bar chart of criteria contributions to overall performance
of the alternative
}
\description{
Takes weighted preference matrix and uses it to plot stacked bar chart
 describing how the criteria (forming the stocks in graph) contribute to the
 overall score of the alternatives. Visualized in form of the ordered bar
 plot.

This is usefull especially for methods such as \link{wsm}, which derives
 results directly from performance matrix.
}
\examples{
KoboGlo <- c(80, 97, 50, 100, 100, 94)
SonyPRST3 <- c(80, 88, 100, 20, 70, 70)
KindlePaperwhite2 <- c(100, 85, 50, 100, 70, 100)
PBTouchLux <- c(80, 90, 100, 100, 70, 94)
BookeenCybookOdyssey <- c(80, 100, 100, 100, 85, 50)
criteria <- c("Display", "weight", "HWButtons", "FrontLight", "batery",
  "price")
PM <- as.data.frame(
  rbind(KoboGlo,
        SonyPRST3,
        KindlePaperwhite2,
        PBTouchLux,
        BookeenCybookOdyssey))
names(PM) <- criteria
w <- c(5, 3, 4, 5, 2, 1)
preferences <- mcda_wsm(PM, w, 'max')
plot.scoreM(preferences$weightedPM)

}
\references{
De Brouwer, Philippe J. S.: "The The Big R-Book: From Data Science to
 Learning Machines and Big Data ", Wiley, 2020, 928 p., ISBN  978-1119632726.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MCDA}
\keyword{WSM}
\keyword{performance}
\keyword{score}
