% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.threshold.3d.R
\name{plot.threshold.3d}
\alias{plot.threshold.3d}
\title{Plots alternatives participation in kernel with changing concordance and
 discordance thresholds}
\usage{
\method{plot}{threshold.3d}(
  df,
  title = "Kernel participation",
  x_label = "concordance",
  y_label = "discordance"
)
}
\arguments{
\item{df}{dataframe with alternatives (0/1) and concordance, discordance
columns}

\item{title}{main title of the graph}

\item{x_label}{label of the x axis representing concordance threshold}

\item{y_label}{label of the y axis representing discordance threshold}
}
\value{
a graph representing participation of the alternative in kernel with
 changing concordance and discordance threshold
}
\description{
Plot visualized participation of the alternatives in the kernel in
 \code{\link{electre1}} function.

Alternatives are organized on z axis. X represents concordance and y axis
 represents discordance threshold. Visually each alternative has its own
 row which allows us to explore how the participation in kernel changes in
 3D.

Look for missing dots for indication that the alternative is not
 participating in the kernel at that level of concordance and discordance
 threshold.

Function is constructed similarly to \code{\link{plot.threshold}}, but
 produces 3D graph. The alternatives participation is indicated only by 0/1.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{concordance}
\keyword{discordance}
\keyword{electre1}
\keyword{electre2}
\keyword{sensitivity}
\keyword{threshold}
