% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_order_matrix.R
\name{pre_order_matrix}
\alias{pre_order_matrix}
\title{function to create preorder matrix}
\usage{
pre_order_matrix(rank_D, rank_A, alt)
}
\arguments{
\item{rank_D}{descending distillation ranking in the form of ordered dataframe}

\item{rank_A}{ascending distillation ranking in the form of ordered dataframe}

\item{alt}{vector of alternatives names}
}
\value{
pre-order matrix
}
\description{
Function takes outputs of descending and ascending distillation pre-order and
 created pre-order matrix describing outranking relation between the
 alternatives.

The relation can be P+ (a outranks b), P- (b outranks a), I (indifference)
 and R (incomparable). Information from preorder matrix can be utilized to
 create adjacency matrix and construct final ranking.

This function is being utilized by \link{electre3} and \link{electre4}.
}
\references{
Meyer, P. at al. MCDA package. GitHub: 2021, available from:
 \url{https://github.com/paterijk/MCDA/blob/master/}

Prombo, M. Package OutrankingTools, CRAN: 2015, available from:
 \url{https://cran.r-project.org/web/packages/OutrankingTools/}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{electre3}
\keyword{electre4}
\keyword{matrix}
\keyword{pre-order}
