% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferenceDegree.R
\name{preferenceDegree}
\alias{preferenceDegree}
\title{computes preference degree for SIR and PROMETHEE functions}
\usage{
preferenceDegree(nalt, ncri, DK, d, qj, pj, sj, alt, cri)
}
\arguments{
\item{nalt}{number of alternatives}

\item{ncri}{number of criteria}

\item{DK}{result of paiweise comparison}

\item{d}{type of function (default, U-shape, V-shape, level), vector - sets
the type for every criterion}

\item{qj}{indifference threshold vector}

\item{pj}{preference threshold vector}

\item{sj}{ntermediate threshold vector}

\item{alt}{vector of alternative's names}

\item{cri}{vector of criteria's names}
}
\value{
preference degree matrix
}
\description{
The function is being used in \code{\link{SIR}} and
 \code{\link{PROMETHEE}} functions to compute preferencee degree.
 degree is then further used by these methods to derive its
 recommendations.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{PROMETHEE}
\keyword{SIR}
\keyword{degree}
\keyword{preference}
