% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probid.R
\name{probid}
\alias{probid}
\title{Preference Ranking on the Basis of Ideal-Average Distance (WIP implementation)}
\description{
Proposed by Wang et al. (2021). The evaluation works on basis of comparison
 of the perfomance of the alternatives to range of reference solutions.
 For example TOPSIS works with single ideal and antiideal solution, PROBID
 works with tiered ideal solutions - forming m tiers of ideal solutions. (m
 is number of alternatives).

Additionaly method compares the performance also to average perfomance.

In computation we first normalize performance matrix using vector
 normalization. Then we apply weights on the normalized performance matrix.

To continue with computation we need to construct tiered ideal solution by
 sorting weighted normalized performance matrix and compute averages.

Euclidean distances of each alternative to each of the ideal solutions as
 well as to average needs to be computed.

Distance do k-th ideal solution:

\mjsdeqn{S_{i(k)} = \sqrt{\sum_{j=1}^n (v_{ij} - v_{(k)j})^2}}

Distance to average solution:

\mjsdeqn{S_{i(avg)} = \sqrt{\sum_{j=1}^n (v_{ij} - \overline{v_j})^2}}

Based on it we can compute overall positive-ideal and negative-ideal
 distances.

For positive-ideal (pid) when m is odd number:

\mjsdeqn{S_{i(pid)} = \sum_{k=1}^{\frac{m+1}{2}}\frac{1}{k}S_{i(k)}}

when m is even number:

\mjsdeqn{S_{i(pid)} = \sum_{k=1}^{m/2}\frac{1}{k}S_{i(k)}}

For negative-ideal (nid) when m is odd number:

\mjsdeqn{S_{i(nid)} = \sum_{k=\frac{m+1}{2}}^{m}\frac{1}{m - k + 1}S_{i(k)}}

when m is even number:

\mjsdeqn{S_{i(nid)} = \sum_{k=\frac{m}{2} + 1}^{m}\frac{1}{m - k + 1}S_{i(k)}}

Ideal and anti-ideal solutions then can be used to compute their ratio:

\mjsdeqn{R_i = \frac{S_{i(pid)}}{S_{i(nid)}}}

which is used to compute performance score P (higher is better):

\mjsdeqn{P_i = \frac{1}{1 + R_i^2} + S_{i(avg)}}
}
\examples{

## ------------------------------------------------
## Method `probid$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- probid$new(pm, w = w, minmax = minmax)
}
\references{
WANG, Zhiyuan, RANGAIAH, Gade Pandu, WANG, Xiaonan. Preference Ranking on
 the Basis of Ideal-Average Distance Method for Multi-Criteria
 Decision-Making. Industrial & Engineering Chemistry Research. 2021, vol.
 60, no. 30, pp. 11216–11230, available from:
 https://doi.org/10.1021/acs.iecr.1c01413, ISSN 0888-5885.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{PROBID}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{normPM}}{normalized performance matrix (using vector normalization)}

\item{\code{w}}{weights vector}

\item{\code{result}}{data frame with distances to ideal solution (si_pid), anti-
ideal solution (si_nid), distances ratio (Ri), performance score (Pi),
rank of the alternatives}

\item{\code{minmax}}{vector of optimization directions to the criteria. Use
max/min only. Single max or min will be interpreted as direction to
optimize all criteria.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-probid-new}{\code{probid$new()}}
\item \href{#method-probid-compute}{\code{probid$compute()}}
\item \href{#method-probid-summary}{\code{probid$summary()}}
\item \href{#method-probid-clone}{\code{probid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probid-new"></a>}}
\if{latex}{\out{\hypertarget{method-probid-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{probid$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- probid$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probid-compute"></a>}}
\if{latex}{\out{\hypertarget{method-probid-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of PROBID model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{probid$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probid-summary"></a>}}
\if{latex}{\out{\hypertarget{method-probid-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the PROBID count method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{probid$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-probid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{probid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
