% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promethee1.R
\name{promethee1}
\alias{promethee1}
\title{method for computations of partial preference using PROMETHEE I method}
\description{
PROMETHEE stands for Preference Ranking Organization METHod for Enrichment
 Evaluation. Promethee I method is intended for establishment of partial
 ranking of the alternatives, by evaluation of positive and negative
 preferences flows in pairwise comparisons of the alternatives.

Method uses general \link{PROMETHEE} function and computes comparison states
 on top of it.

The computation starts with computing positive and negative flows using
 PROMETHEE function. This function is being used in PROMETHEE I, II and III
 methods (see \link{promethee2}, \link{promethee3}).

Since the flow computation is shared among the PROMETHEE family methods, it
 is implemented as separate function in \link{PROMETHEE}. Refer to
 the function documentation for details on computation of flows.

Flows are then used to derive some preferences in alternatives. For example
 if (P+(a) > P+(b) and P-(a) < P-(b)) or (P+(a) == P+(b) and P-(a) < P-(b))
 or (P+(a) > P+(b) and P-(a) == P-(b)) then we can say that a is preferred
 to b: aPb.

If positive and negative flows are same for both alternatives, we can say,
 that we are indifferent (preferentially): aIb, otherwise the alternatives
 are incomparable: aRb.
}
\examples{

## ------------------------------------------------
## Method `promethee1$new`
## ------------------------------------------------

# Example from Fuzzy TOPSIS book (see references)
# ammended error in tab. 4.9, the computation presumes maximization of
# all criteria
PM <- cbind(
  c(80, 65, 83, 40, 52, 94),
  c(90, 58, 60, 80, 72, 96),
  c(600, 200, 400, 1000, 600, 700),
  c(54, 97, 72, 75, 20, 36),
  c(8, 1, 4, 7, 3, 5),
  c(5, 1, 7, 10, 8, 6)
)
colnames(PM) <- c('C1', 'C2', 'C3', 'C4', 'C5', 'C6')
rownames(PM) <- c('A1', 'A2', 'A3', 'A4', 'A5', 'A6')
minmax <- 'max'
shape <- c('U-shape', 'V-shape', 'linear', 'level', 'default',
           'Gaussian')
p <- c(10, 0, 450, 50, 0, 0) #indifference threshold
q <- c(0, 30, 50, 10, 0, 0) #preference threshold
s <- c(0,0,0,0,0,5) #intermediate threshold
w <- c(0.1667, 0.1667, 0.1667, 0.1667, 0.1667, 0.1665)
result <- promethee1$new(PM, shape, w, minmax, q, p, s)
}
\references{
ALAOUI, Mohamed El. Fuzzy TOPSIS: Logic, Approaches, and Case Studies. Boca
 Raton: CRC Press, 2021. 216 s. ISBN 978-0-367-76748-8.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{PROMETHEE}
\keyword{promethee1}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original performance matrix (as used in constructor)}

\item{\code{pm}}{performance matrix with all criteria converted to maximize}

\item{\code{pref_function}}{vector of preference functions for the criteria}

\item{\code{w}}{weights vector}

\item{\code{minmax}}{vector of min/max (set direction of optimization for
criteria)}

\item{\code{i_threshold}}{indefference threshold vector}

\item{\code{p_threshold}}{preference threshold vector}

\item{\code{im_threshold}}{itermediate threshold vector}

\item{\code{positiveFlow}}{vector representing how the alternative is preferred
to other alternatives}

\item{\code{negativeFlow}}{vector representing how alternative is outranked by
other alternatives}

\item{\code{preferenceDegree}}{matrix representing aggregated weighted
preferences of the alternatives across the criteria}

\item{\code{preferenceDegreeUnw}}{list of matrixes with unweighted preferences,
separate matrix for each criterion}

\item{\code{pairweiseComparison}}{list of matrixes measuring nominal
differences in alternatives performance in criterions. Separate
matrixes are constructed for each criterion.}

\item{\code{preferenceMatrix}}{preference matrix with specified P+ (a preferred
to b), P- (b preferred to a), I (indifferent) and R (incomparable) for
every pair of alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-promethee1-new}{\code{promethee1$new()}}
\item \href{#method-promethee1-compute}{\code{promethee1$compute()}}
\item \href{#method-promethee1-summary}{\code{promethee1$summary()}}
\item \href{#method-promethee1-sensitivity}{\code{promethee1$sensitivity()}}
\item \href{#method-promethee1-clone}{\code{promethee1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-promethee1-new"></a>}}
\if{latex}{\out{\hypertarget{method-promethee1-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for PROMETHEE I method. Valides the inputs and
 computes the method based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{promethee1$new(
  pm,
  pref_function,
  w,
  minmax = "max",
  i_threshold = NULL,
  p_threshold = NULL,
  im_threshold = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion. only
numeric values expected. Rows and columns are expected to be named.}

\item{\code{pref_function}}{vector, specifies type of function used to compute
preferences. Need to be set for each criterion. Possible values are:
'default', 'U-shape', 'V-shape', 'level', 'linear', 'Gaussian'. Choice
of function type will decide on what type of threshold (if any) is
required for computation. Each criterion can use different preference
function.}

\item{\code{w}}{vector containing the weights of the criteria. Values need to
0 <= wi <= 1, sum(wi) = 1}

\item{\code{minmax}}{can be set to either value or vector. Value (min or max)
is usable in situation when all criteria are either benefit or cost
(are not mixed). If Criteria orientation is mixed, vector is required
to set criterion orientation right.}

\item{\code{i_threshold}}{vector containing indifference thresholds for
criteria. Not all types of performance functions require it. The
parameter must be used if there is at least one criterion, for which it
is required. Values for all other criteria should be 0 (and will not be
used during computations). Only 'U-shape', 'level', 'linear' functions
need this threshold.}

\item{\code{p_threshold}}{vector containing preference thresholds for criteria.
Not all types of performance functions require it. The parameter must
be used if there is at least one criterion, for which it is required.
Values for all other criteria should be 0 (and will not be used during
computations). Only 'V-shape', 'level', 'linear' functions need this
threshold.}

\item{\code{im_threshold}}{vector containing intermediate thresholds for
criteria. only Gaussian type performance functions require this type of
threshold. If preference and indifference thresholds are present, the
PROMETHEE function will try to 'gues' intermediate threshold as value
right in the middle between these thresholds.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example from Fuzzy TOPSIS book (see references)
# ammended error in tab. 4.9, the computation presumes maximization of
# all criteria
PM <- cbind(
  c(80, 65, 83, 40, 52, 94),
  c(90, 58, 60, 80, 72, 96),
  c(600, 200, 400, 1000, 600, 700),
  c(54, 97, 72, 75, 20, 36),
  c(8, 1, 4, 7, 3, 5),
  c(5, 1, 7, 10, 8, 6)
)
colnames(PM) <- c('C1', 'C2', 'C3', 'C4', 'C5', 'C6')
rownames(PM) <- c('A1', 'A2', 'A3', 'A4', 'A5', 'A6')
minmax <- 'max'
shape <- c('U-shape', 'V-shape', 'linear', 'level', 'default',
           'Gaussian')
p <- c(10, 0, 450, 50, 0, 0) #indifference threshold
q <- c(0, 30, 50, 10, 0, 0) #preference threshold
s <- c(0,0,0,0,0,5) #intermediate threshold
w <- c(0.1667, 0.1667, 0.1667, 0.1667, 0.1667, 0.1665)
result <- promethee1$new(PM, shape, w, minmax, q, p, s)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-promethee1-compute"></a>}}
\if{latex}{\out{\hypertarget{method-promethee1-compute}{}}}
\subsection{Method \code{compute()}}{
computes PROMETHEE I model based on parameters set in properties of the
 class. It is usually not neccessary to call this funtion manually as it
 is called automatically by the class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{promethee1$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-promethee1-summary"></a>}}
\if{latex}{\out{\hypertarget{method-promethee1-summary}{}}}
\subsection{Method \code{summary()}}{
genrates basic information about the computed model and prints it to the
 console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{promethee1$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-promethee1-sensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-promethee1-sensitivity}{}}}
\subsection{Method \code{sensitivity()}}{
test sensitivity of the model to changes in the thresholds.

Provides sens_i (for indifference threshold), sens_p (for preference
 threshold) and sens_im (for intermediate treshold) dataframes in
 structure:

\itemize{
  \item criterium
  \item from - lower bound of sensitivity
  \item default - value computed by original model
  \item to - upper bound of sensitivity
  \item function - type of preference function for criterium
}

Dataframes do have a numeric value with lower/upper bound of
 sensitivity identifying last value of the threshold for which the
 result still doesn't change.

If no such value is identified "insens." is provided.

Also note that PROMETHEE function is complex, because of existence of
 "preference" functions, which are stated separately for each criterium
 and each of these has a different requirements on types of thresholds
 it uses. For example level and linear functions use both preference and
 indifference thresholds (but not intermediate). V-shape function uses
 preference threshold only, U-shape uses indifference threshold only.

Gaussian function uses intermediate threshold only, but if preference
 andindifference thresholds are provided, the sensitivity is being
 tested on in interval between these two.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{promethee1$sensitivity(step = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{number of steps to divide threshold testing interval}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframes sens_i, sens_p and sens_im with sensitivity limit for
 the criteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-promethee1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-promethee1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{promethee1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
