% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promethee_param_check.R
\name{promethee_param_check}
\alias{promethee_param_check}
\title{internal function to check parameters for PROMETHEE and SIR functions}
\usage{
promethee_param_check(
  pm,
  preference_function,
  w,
  indifference_treshold = NULL,
  prefference_threshold = NULL,
  intermediate_threshold = NULL
)
}
\arguments{
\item{pm}{Matrix or data frame containing the performance table. Each row
corresponds to an alternative, and each column to a criterion. only numeric
values expercted. Rows and columns are expected to be named.}

\item{preference_function}{vector, specifies type of function used to
compute preferences. Need to be set for each criterion. Possible values
are: 'default', 'U-shape', 'V-shape', 'level', 'linear', 'Gaussian'. Choice
of function type will decide on what type of threshold (if any) is required
for computation. Each criterion can use different preference function.}

\item{w}{vector containing the weights of the criteria. Values need to
0 <= wi <= 1, sum(wi) = 1}

\item{indifference_treshold}{vector containing indifference threshods for
criteria. Not all types of performance functions require it. The parameter
must be used if there is at least one criterion, for which it is required.
Values for all other criteria should be 0 (and will not be used during
computations). Only 'U-shape', 'level', 'linear' functions need this
type of threshold.}

\item{prefference_threshold}{vector containing prefference threshods for
criteria. Not all types of performance functions require it. The parameter
must be used if there is at least one criterion, for which it is required.
Values for all other criteria should be 0 (and will not be used during
computations). Only 'V-shape', 'level', 'linear' functions need this
threshold.}

\item{intermediate_threshold}{vector containing intermetiate thresholds for
criteria. only Gaussian type performance functions rewuire this type of
threshold. If prefference and indifference thresholds are present, the
PROMETHEE function will try to 'gues' intermediate threshold as value right
in the middle between these thresholds.}
}
\value{
The function returns recomputed value of intermediate threshold if
 no error in parameter checking logic is detected.
}
\description{
Performs parameter checking for \link{PROMETHEE} and \link{sir} functions.
 It especially focuses on preference, indifference and intermediate
 thresholds as these various preference functions deal with these in
 different way.

Function tries to derive intermediate threshold for Gausian preference
 function (if it is missing) from prefererence and indifference thresholds.

Function then returns ammended vector of intermediate thresholds.

This function is intended for package's internal use only.
}
\references{
ALAOUI, Mohamed El. Fuzzy TOPSIS: Logic, Approaches, and Case Studies. Boca
 Raton: CRC Press, 2021. 216 s. ISBN 978-0-367-76748-8.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{PROMETHEE}
\keyword{SIR}
