% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Preference Selection Index}
\description{
Decision support method in family of weighted sum methods such as \link{wsm}
 and many other. The difference is, that the method does not take weights as
 parameter but derives them from provided performance information in manner
 which is similar to EWM (see \link{mcda_objective_weights}).

The process is following. First the performance matrix (alternatives in
 rows and criteria in colums is provided). These values are normalized. For
 benefit criteria we use:

\mjsdeqn{R_{ij} = \frac{x_{ij}}{max_j x_{ij}}}

and for cost criteria we use:

\mjsdeqn{R_{ij} = \frac{min_j x_{ij}}{x_{ij}}}

next we compute performance variation value (PV) computation by comparing
 normalized performance value and mean performance in the criterium.

\mjsdeqn{PV = \sum_{j=1}^m (R_{ij} - mean(R_j))^2}

We estimate overall performance

\mjsdeqn{\Psi = 1 - PV_j}

Weight vector can be then computed as

\mjsdeqn{w = \frac{\Psi}{\sum \Psi}}

Overall performance selection index (PSI) can be computed using normal WSM
 approach

\mjsdeqn{PSI = \sum_{i = 1}^n R_{ij} \cdot w_i}

Ranking of the alternatives is directly derivable from value of PSI.
}
\examples{

## ------------------------------------------------
## Method `psi$new`
## ------------------------------------------------

#from https://www.youtube.com/watch?v=jlxHKBkai_w
alternatives <- c("car 1", "car 2", "car 3", "car 4", "car 5", "car 6",
 "car 7", "car 8", "car 9", "car 10")
criteria <- c("quality", "condition", "security", "delivery days",
 "fuel consumption", "price")
M <- rbind(
  c(3, 6, 4, 20, 2, 30000),
  c(4, 4, 6, 15, 2.2, 32000),
  c(6, 5, 9, 18, 3, 32100),
  c(5, 6, 3, 23, 2.8, 28000),
  c(4, 8, 7, 30, 1.5, 29000),
  c(8, 3, 6, 35, 1.9, 27000),
  c(7, 2, 5, 33, 1.7, 28500),
  c(3, 8, 3, 34, 1.6, 30500),
  c(8, 4, 8, 40, 2.5, 33000),
  c(9, 3, 7, 34, 2, 29800)
)
rownames(M) <- alternatives
colnames(M) <- criteria
minmax <- c("max", "max", "max", "min", "min", "min")
t <- psi$new(M, minmax)
}
\references{
PSI. How to use Preference Selection Index "PSI" to select the best
 alternatives? Available from
 \url{https://www.youtube.com/watch?v=jlxHKBkai_w}.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{EWM}
\keyword{PSI}
\keyword{WSM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original performance matrix (not normalized)}

\item{\code{pm}}{normalized performance matrix (alternatives in rows, criteria
in columns)}

\item{\code{minmax}}{vector with optimization direction of the criteria (min or
max), can be substituted by single min or max if optimization direction
for all criteria is the same.}

\item{\code{w}}{computed weight vector for criteria}

\item{\code{PSI}}{weighted sum percentages}

\item{\code{scoreM}}{graph visualizing contribution of the criteria to overall
PSI}

\item{\code{result}}{dataframe with WSM output (and ranks)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-psi-new}{\code{psi$new()}}
\item \href{#method-psi-compute}{\code{psi$compute()}}
\item \href{#method-psi-summary}{\code{psi$summary()}}
\item \href{#method-psi-clone}{\code{psi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-psi-new"></a>}}
\if{latex}{\out{\hypertarget{method-psi-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psi$new(pm, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class included computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#from https://www.youtube.com/watch?v=jlxHKBkai_w
alternatives <- c("car 1", "car 2", "car 3", "car 4", "car 5", "car 6",
 "car 7", "car 8", "car 9", "car 10")
criteria <- c("quality", "condition", "security", "delivery days",
 "fuel consumption", "price")
M <- rbind(
  c(3, 6, 4, 20, 2, 30000),
  c(4, 4, 6, 15, 2.2, 32000),
  c(6, 5, 9, 18, 3, 32100),
  c(5, 6, 3, 23, 2.8, 28000),
  c(4, 8, 7, 30, 1.5, 29000),
  c(8, 3, 6, 35, 1.9, 27000),
  c(7, 2, 5, 33, 1.7, 28500),
  c(3, 8, 3, 34, 1.6, 30500),
  c(8, 4, 8, 40, 2.5, 33000),
  c(9, 3, 7, 34, 2, 29800)
)
rownames(M) <- alternatives
colnames(M) <- criteria
minmax <- c("max", "max", "max", "min", "min", "min")
t <- psi$new(M, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-psi-compute"></a>}}
\if{latex}{\out{\hypertarget{method-psi-compute}{}}}
\subsection{Method \code{compute()}}{
Computed the PSI using params provided in constructor. Usually run
 automatically by constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psi$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-psi-summary"></a>}}
\if{latex}{\out{\hypertarget{method-psi-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the PSI method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psi$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-psi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-psi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{psi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
