% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_method.R
\name{r_method}
\alias{r_method}
\title{R-method}
\description{
Simple decision support method based on evaluation of rands assigned to the
 criteria (ranking importance of criteria) and performance of alternatives
 in the criteria.

For both weights are computed:

\mjsdeqn{w_j = \frac{1/\sum_{k=1}^j(1/r_k)}{\sum_{j=1}^n(1/r_k)}}

where wj is the weight of the j-th alternative or criterium, rk is the rank
 assigned to k-th alternative or criterium.

The resulting two sets of weights are aggregated basically using \link{wsm}
 method, where criteria weights serve as weights and performance weights
 serve as pervormance measure.

Rasulting score can be used for ranking purposes, with highest value being
 the best.
}
\references{
Shankar Chakraborty, Prasenjit Chatterjee, Partha Protim Das. R METHOD. In
 Multi-Criteria Decision-Making Methods in Manufacturing Environments. Apple
 Academic Press: New York, 2023, pp. 337-342, ISBN 9781003377030.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rm}}{ranking matrix consisting of performance ranks evaluation of
alternatives in criteria}

\item{\code{criteria_ranks}}{ranks of criteria describing their importance to the
decision}

\item{\code{w}}{weights vector (based on criteria_ranks)}

\item{\code{wrm}}{weights data frame derived form ranking matrix}

\item{\code{result}}{data frame with composite score and assigned ranks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-r_method-new}{\code{r_method$new()}}
\item \href{#method-r_method-compute}{\code{r_method$compute()}}
\item \href{#method-r_method-summary}{\code{r_method$summary()}}
\item \href{#method-r_method-clone}{\code{r_method$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r_method-new"></a>}}
\if{latex}{\out{\hypertarget{method-r_method-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r_method$new(rm, criteria_ranks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rm}}{normalized performance matrix}

\item{\code{criteria_ranks}}{anks of criteria describing their importance to the
decision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r_method-compute"></a>}}
\if{latex}{\out{\hypertarget{method-r_method-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of SECA model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r_method$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r_method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-r_method-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the R method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r_method$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r_method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-r_method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{r_method$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
