% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafsi.R
\name{rafsi}
\alias{rafsi}
\title{Ranking of Alternatives Through Functional of Criterion Sub-Intervals into a Single Interval}
\description{
As in other methods we start with performance matrix. We use it to derive
 ideal (ai) and antiidedeal (an) values.

Then we map each of the elements of performance matrix into correcponding
 criteria intervals Cj in <an, ai> for beneficial critera and Cj in <ai, an>
 for cost criteria.

We work with 2k fixed points. So worst performance in criterium will always
 be mapped to n1 and the best will be mapped to n2k. With n1 = 1 and n2k can
 be seen as how many times better is the best against the aitiideal. In is
 recommended for n2k to be at least 6 or more.

The PM needs to be mapped to the intervals:

\mjsdeqn{f_i (x) = \frac{n2k - n_1}{a_{Ij} - a_{Nj}} x + \frac{a_{Ij}n_1 - a_{Nj}n_{2k}}{a_{Ij} - a_{Nj}}}

The result needs to be normalized. For benefit sriteria we use:

\mjsdeqn{\hat{s_{ij}} = \frac{s_{ij}}{2A}}

and for cost criteria:

\mjsdeqn{\hat{s_{ij}} = \frac{H}{2s_{ij}}}

where A is arithmetic and H is harmonic mean:

\mjsdeqn{A = \frac{n_1 + n_{2k}}{2}}

\mjsdeqn{H = \frac{2}{\frac{1}{n_1} + \frac{1}{n_{2k}}}}

To the resulting matrix we apply the weights and sum across the criteria to
 get measure Vi, which we sort in descending maner.
}
\references{
Shankar Chakraborty, Prasenjit Chatterjee, Partha Protim Das. Ranking of
 Alternatives through Functional Mapping of Criterion Sub-Intervals into
 a Single Interval (RAFSI) Method. In Multi-Criteria Decision-Making
 Methods in Manufacturing Environments. Apple Academic Press: New York, 2023,
 pp. 317-323, ISBN 9781003377030.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{n2k}}{number of intervals, recommended to be 6 or more (integer)}

\item{\code{result}}{results dataframe with value and ranking of the alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rafsi-new}{\code{rafsi$new()}}
\item \href{#method-rafsi-compute}{\code{rafsi$compute()}}
\item \href{#method-rafsi-summary}{\code{rafsi$summary()}}
\item \href{#method-rafsi-clone}{\code{rafsi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rafsi-new"></a>}}
\if{latex}{\out{\hypertarget{method-rafsi-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rafsi$new(pm, w, minmax = "max", n2k = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{n2k}}{number of intervals, recommended to be 6 or more (integer)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rafsi-compute"></a>}}
\if{latex}{\out{\hypertarget{method-rafsi-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of RAFSI model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rafsi$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rafsi-summary"></a>}}
\if{latex}{\out{\hypertarget{method-rafsi-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the RAFSI method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rafsi$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rafsi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rafsi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rafsi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
