% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rancom.R
\name{rancom}
\alias{rancom}
\title{RANking COMparison}
\usage{
rancom(pm)
}
\value{
list:
\itemize{
   \item preferenceGraph - preference relation between criteria expressed as
 a graph
   \item w - derived weights (sum(w) = 1)
}
}
\description{
Very similar method to binary pairwise comparison (see \link{binary_pairwise_comp})
 only with slightly differently expressed preferences. In RANCOM the
 preference is 1 if f(Ci) < f(Cj), 0 if f(Ci) < f(Cj) and 0.5 if
 f(Ci) == f(Cj).

The preferences are them summed up on per row basis and weights derived by
 vector normalizing the result.
}
\examples{
PM <- rbind(
   c(0,1,1,0,1,1),
   c(0,0,0,0,1,1),
   c(0,1,0,0,1,1),
   c(1,1,1,0,1,1),
   c(0,0,0,0,0,1),
   c(0,0,0,0,0,0)
)
rownames(PM) <- colnames(PM) <- c("C1", "C2", "C4", "C5", "C6", "C8")
t <- rancom(PM)
}
\references{
WIĘCKOWSKI, Jakub, KIZIELEWICZ, Bartłomiej, SHEKHOVTSOV, Andrii, SAŁABUN, Wojciech.
 RANCOM: A novel approach to identifying criteria relevance based on inaccuracy
 expert judgments. Engineering Applications of Artificial Intelligence. 2023,
 vol. 122, p. 106114, available from: \url{https://doi.org/10.1016/j.engappai.2023.106114},
 ISSN 0952-1976.
}
