% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawec.R
\name{rawec}
\alias{rawec}
\title{Ranking of Alternatives with Weights of Criterion}
\description{
Proposed by Puška et al. (2024). The procedure is as follows. First we
 formulate performance matrix. Next we double normalize it. For benefit
 criteria:

\mjsdeqn{n_{ij} =\frac{x_{ij}}{x_{jmax}}}

\mjsdeqn{n_{ij}^{'} =\frac{x_{jmin}}{x_{ij}}}

For cost criteria:

\mjsdeqn{n_{ij} =\frac{x_{jmin}}{x_{ij}}}

\mjsdeqn{n_{ij}^{'} =\frac{x_{ij}}{x_{jmax}}}

In other words we transfor all criteria to benefit criteria (nij) and cost
 criteria (n'ij). This is unique feature of RAWEC method.

In step 3 deviations from criterion weight are computed:

\mjsdeqn{v_{ij} =\sum_{j=1}^n w_j(1-n_{ij})}

\mjsdeqn{v_{ij}^{'} =\sum_{j=1}^n w_j(1-n_{ij}^{'})}

Finally we calculate value of RAWEC Qi by:

\mjsdeqn{Q_i =\frac{v_{ij}^{'} - v_{ij}}{v_{ij}^{'} + v_{ij}}}

The higher the better. This value can be translated into the ranks for
 alternatives.
}
\examples{

## ------------------------------------------------
## Method `rawec$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- rawec$new(pm, w = w, minmax = minmax)
}
\references{
Puška, A., Štilić, A., Pamučar, D., Božanić, D., & Nedeljković, M. (2024).
 Introducing a Novel multi-criteria Ranking of Alternatives with Weights of
 Criterion (RAWEC) model. MethodsX, 102628.
 https://doi.org/10.1016/j.mex.2024.102628
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{RAWEC}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{normPM1}}{normalized performance matrix (to benefit)}

\item{\code{normPM2}}{normalized performance matrix (to cost)}

\item{\code{w}}{weights vector}

\item{\code{Qi}}{vector of RAWEC values}

\item{\code{rank}}{vector of ranks for alternatives based on Qi}

\item{\code{minmax}}{vector of optimization directions to the criteria. Use
max/min only. Single max or min will be interpreted as direction to
optimize all criteria.}

\item{\code{result}}{dataframe with Q1 and ranks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rawec-new}{\code{rawec$new()}}
\item \href{#method-rawec-compute}{\code{rawec$compute()}}
\item \href{#method-rawec-summary}{\code{rawec$summary()}}
\item \href{#method-rawec-clone}{\code{rawec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rawec-new"></a>}}
\if{latex}{\out{\hypertarget{method-rawec-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawec$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- rawec$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rawec-compute"></a>}}
\if{latex}{\out{\hypertarget{method-rawec-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of RAWEC model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawec$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rawec-summary"></a>}}
\if{latex}{\out{\hypertarget{method-rawec-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the RAWEC count method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawec$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rawec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rawec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rawec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
