% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regime.R
\name{regime}
\alias{regime}
\title{REGIME}
\description{
Simple MADM method for decision support in situation. Proposed by Hinloopen
 et al in 1983.

As in other methods we start we establishing performance of the alternatives
 in criteria. This implementation presumes that the numerical scales are
 used.

First we compute superiority index. We do so by pairwise comparison of the
 alternatives in their respective criteria, identifying such criteria, where
 alternative 1 is superior to alternative 2.

 Next we compute superiority identifier for the pairwise comparisons by
 summing weights of the criteria we identified in superior index. (We sum
 the weights of the criteria in which alternative 1 is superior to
 alternative 2).

We construct impact analysis matrix in which we rank all alternatives
 separately for each criterium.

Based on the ranking we construct REGIME matrix of parwise comparisons of
 the alternatives in criteria. We use -1 value if aleternative 1 is lower
 rank then alternative 2, 0 if they are equal and +1 if alternative 1 has
 better rank then alternative 2.

Based on information provided by REGIME matrix guide index is constructed.
 WE apply weights to criteria and sum across the criteria to derive
 agregated comparison matrix - in REGIME method caled guiding index.

Modern interpretation would be to see this matrix as adjacancy matrix
 representing comparisons among the alternatives.

Based on this information we can derive approximate order of the
 alternatives. Basically alternatives which have been preferred more times are
 better than these preferred lower number of times.

Positive guiding index indicates that alternative 1 is better then
 alternative 2 (and vice versa).

This information can be used to visualize the preference direction, although
 Hinloopen did not use this interpretation when he developed the method.
}
\examples{

## ------------------------------------------------
## Method `regime$new`
## ------------------------------------------------

alternatives <- c("A1", "A2", "A3")
criteria <- c("C1", "C2", "C3", "C4", "C5")
pm <- rbind(
  c(3, 2, 4, 24000, 4),
  c(1.2, 3, 2, 25000, 3),
  c(1.5, 4, 1, 32000, 1)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.1, 0.175, 0.25, 0.35, 0.125)
minmax <- c("min", "max", "max", "max", "min")
t <- regime$new(pm, w, minmax)
}
\references{
Hinloopen, E., Nijkamp, P., Rietveld, P. (1983). The Regime Method: A New
 Multicriteria Technique. In: Hansen, P. (eds) Essays and Surveys on
 Multiple Criteria Decision Making. Lecture Notes in Economics and
 Mathematical Systems, vol 209. Springer, Berlin, Heidelberg.
 \url{https://doi.org/10.1007/978-3-642-46473-7_13}

Alinezhad, A., Khalili, J. New Methods and Applications in Multiple
 Attribute Decision Making (MADM). Springer Nature Switzerland AG, 2019,
 233 p., ISBN 978-3-030-15009-9
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MADM}
\keyword{REGIME}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}

\item{\code{w}}{normalized weights of the criteria}

\item{\code{gi}}{guide index - comparison matrix of the alternatives agregated
accros the criteria}

\item{\code{finalRank}}{- final ranking of the alternatives}

\item{\code{graph}}{- visualization of the guide index using network diagram}

\item{\code{result}}{dataframe with gi and rank}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-regime-new}{\code{regime$new()}}
\item \href{#method-regime-compute}{\code{regime$compute()}}
\item \href{#method-regime-summary}{\code{regime$summary()}}
\item \href{#method-regime-clone}{\code{regime$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regime-new"></a>}}
\if{latex}{\out{\hypertarget{method-regime-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{regime$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{normalized weight vector for the criteria}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c("A1", "A2", "A3")
criteria <- c("C1", "C2", "C3", "C4", "C5")
pm <- rbind(
  c(3, 2, 4, 24000, 4),
  c(1.2, 3, 2, 25000, 3),
  c(1.5, 4, 1, 32000, 1)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.1, 0.175, 0.25, 0.35, 0.125)
minmax <- c("min", "max", "max", "max", "min")
t <- regime$new(pm, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regime-compute"></a>}}
\if{latex}{\out{\hypertarget{method-regime-compute}{}}}
\subsection{Method \code{compute()}}{
computes REGIME model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{regime$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regime-summary"></a>}}
\if{latex}{\out{\hypertarget{method-regime-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the REGIME method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{regime$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-regime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{regime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
