% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rim_norm.R
\name{rim_norm}
\alias{rim_norm}
\title{Reference Ideal Model Normalization}
\usage{
rim_norm(tonorm, A, B, C, D)
}
\arguments{
\item{tonorm}{numbers vector to normalize}

\item{A}{lower limit of the range}

\item{B}{upper limit of the range}

\item{C}{lower limit of reference ideal}

\item{D}{upper limit of reference ideal}
}
\value{
normaluzed vector of numbers
}
\description{
Normalization of values used in multi-criteria decision analysis method RIM.

The normalization is based on distance from reference ideal. Consider
 choosing of new emploee with one criterium being age. Supose that we
 consider possible age range [A, B] to [23, 65] years, with ideal reference
 [C, D] in [30, 35].

Considering the above we can say that A < C and D < B (reference interval is
 inside of possible range of values).

The normalization of x is done by computing:

\mjsdeqn{d_{min}(X, [C, D]) = min( |x - C|, |x - D|)}

and three possible scenarios. In first the scenario the value to normalize
 is in reference ideal inerval [C, D]:

\mjsdeqn{f(x, [A, B], [C, D]) = 1}

In second scenario, the to normalize value is in interval [A, C]:

\mjsdeqn{f(x, [A, B], [C, D]) = 1 - \frac{d_{min}(x, [C, D])}{|A - C|}}

Finally the value to normalize can be in the interval [D, B]:

\mjsdeqn{f(x, [A, B], [C, D]) = 1- \frac{d_{min}(x, [C, D])}{|D - B|}}

In other words the normalization approach focusses on establishing how far
 is the value from reference ideal.
}
\examples{
tonorm <- c(25, 30, 35, 40, 50)
A <- 23
B <- 60
C <- 30
D <- 35
t <- rim_norm(tonorm, A, B, C, D)

}
\references{
CABLES, E., LAMATA, M. T., VERDEGAY, J. L. RIM-reference ideal method in
 multicriteria decision making. Information Sciences. 2016, Vol. 337–338,
 pp. 1–10, available from: https://doi.org/10.1016/j.ins.2015.12.011,
 ISSN 0020-0255.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{RIM}
