% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roms.R
\name{roms}
\alias{roms}
\title{Rank Ordering Methods (ROMs)}
\usage{
roms(ranks, p = 1, method = "Rank Sum")
}
\arguments{
\item{ranks}{vector of ranks for the criteria}

\item{p}{rank exponent, must be p >= 0, only used in Rank Exponent Method}

\item{method}{constant identifying method in ROMs, which should be applied.
Currently: Rank Sum, Rank Exponent, Rank Reciprocal are supported.}
}
\description{
Group of methods for deriving weights based on known rank of the criteria.

The process is generally same for all the methods. First we asign ranks to
 the criteria, then we compute the wweights using method's appropriate
 equation.

Currently following methods are supported (bold are constant names to use
 as parameter of the function):

\itemize{
  \item \bold{Rank Sum} Method
  \item \bold{Rank Exponent} Method
  \item \bold{Rank Reciprocal} Method
}

\bold{Rank Sum Method}

\mjsdeqn{w_j = \frac{n - r_j + 1}{\sum_{k=1}^n (n - r_k + 1)}}

\bold{Rank Exponent Method}

\mjsdeqn{w_j = \frac{(n - r_j + 1)^p}{\sum_{k=1}^n (n - r_k + 1)^p}}

\bold{Rank Reciprocal Method}

\mjsdeqn{w_j = \frac{1 / r_j}{\sum_{k=1}^n (1 / r_k)}}
}
\examples{
r <- c(3, 4, 5, 1, 6, 2)
result <- roms(r, method = "Rank Sum")
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
