% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_p12.R
\name{sensitivity_p12}
\alias{sensitivity_p12}
\title{test sensitivity of the PROMETHEE I and II models to changes in the
 thresholds.}
\usage{
sensitivity_p12(object, step = 100)
}
\arguments{
\item{object}{PROMETHEE I, II, III or SIR model we want to test sensitivity
for}

\item{step}{number of steps to divide threshold testing interval (default:
100)}
}
\value{
dataframes sens_i, sens_p and sens_im with sensitivity limit for
 the criteria
}
\description{
Function tests sensitivity of \link{promethee1} and \link{promethee2} models
 to changes of thresholds.

Computes sens_i (for indifference threshold), sens_p (for preference
 threshold) and sens_im (for intermediate treshold) dataframes in
 structure:

\itemize{
  \item criterium
  \item from - lower bound of sensitivity
  \item default - value computed by original model
  \item to - upper bound of sensitivity
  \item function - type of preference function for criterium
}

Dataframes do have a numeric value with lower/upper bound of
 sensitivity identifying last value of the threshold for which the
 result still doesn't change.

If no such value is identified "insens." is provided.

Also note that PROMETHEE function is complex, because of existence of
 "preference" functions, which are stated separately for each criterium
 and each of these has a different requirements on types of thresholds
 it uses. For example level and linear functions use both preference and
 indifference thresholds (but not intermediate). V-shape function uses
 preference threshold only, U-shape uses indifference threshold only.

Gaussian function uses intermediate threshold only, but if preference
 andindifference thresholds are provided, the sensitivity is being
 tested on in interval between these two.

Function is intended for internal use only

Also note that right now the sensitivity testing for the various models
 presumes, that if positive and negative flows remain same, so does the
 final ranking. That means that sensitivity of both methods is tested
 through optic of the PROMETHEE I method. If above mentioned presumption
 would not hold, the PROMETHEE II (or III) would need to be tested based
 on either net flow or from it derived ranking.
}
\keyword{PROMETHEE,}
\keyword{SIR}
\keyword{promethee1,}
\keyword{promethee2}
\keyword{promethee3}
\keyword{sensitivity}
