% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprobid.R
\name{sprobid}
\alias{sprobid}
\title{Simplified Preference Ranking on the Basis of Ideal-Average Distance (WIP
 implementation)}
\description{
Simplified version of \link{probid}. The method has first 4 steps same as
 PROBID. In step 5 evaluates first and last quarted of ideal solution using
 Si_pid (Si positive ideal) and Si_nid (Si negative ideal).

For positive solution ideal whem m >= 4:

\mjsdeqn{S_{i(pid)} = \sum_{k=1}^{m/4} \frac{1}{k}(S_{i(k)})}

when 0 < m < 4:

\mjsdeqn{S_{i(pid)} = S_{i1}}

For negative ideal solution when m >=4:

\mjsdeqn{S_{i(nid)} = \sum_{k=m+1-(m/4)}^{m} \frac{1}{m - k + 1}(S_{i(k)})}

when 0 < m < 4

\mjsdeqn{S_{i(nid)} = S_{im}}

Then the ratio of two Si measures is computed performance score P:

\mjsdeqn{P_i = \frac{S_{i(nid)}}{S_{i(pid)}}}

Pi measure is directly usable for ranking of the alternatives (higher =
 better)
}
\examples{

## ------------------------------------------------
## Method `sprobid$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- sprobid$new(pm, w = w, minmax = minmax)
}
\references{
Baydaş, M., Kavacik, M. Wang, Z. Interpreting the Determinants of
 Sensitivity in MCDM Methods with a New Perspective: An Application on E -
 Scooter Selection with the PROBID Method. Spectrum of Engineering and
 Management Sciences, Vol. 2, No. 1, pp. 17-35, available from:
 https://doi.org/10.31181/sems2120242b, ISSN 3009-3309
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{sPROBID}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{normPM}}{normalized performance matrix (using vector normalization)}

\item{\code{w}}{weights vector}

\item{\code{result}}{data frame with distances to ideal solution (si_pid), anti-
ideal solution (si_nid), distances ratio (Ri), performance score (Pi),
rank of the alternatives}

\item{\code{minmax}}{vector of optimization directions to the criteria. Use
max/min only. Single max or min will be interpreted as direction to
optimize all criteria.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sprobid-new}{\code{sprobid$new()}}
\item \href{#method-sprobid-compute}{\code{sprobid$compute()}}
\item \href{#method-sprobid-summary}{\code{sprobid$summary()}}
\item \href{#method-sprobid-clone}{\code{sprobid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sprobid-new"></a>}}
\if{latex}{\out{\hypertarget{method-sprobid-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sprobid$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- sprobid$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sprobid-compute"></a>}}
\if{latex}{\out{\hypertarget{method-sprobid-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of PROBID model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sprobid$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sprobid-summary"></a>}}
\if{latex}{\out{\hypertarget{method-sprobid-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the sPROBID count method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sprobid$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sprobid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sprobid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sprobid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
