% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aras.R
\name{summary.aras}
\alias{summary.aras}
\title{Summary method for aras objects}
\usage{
\method{summary}{aras}(obj)
}
\arguments{
\item{obj}{aras object to process summary for}
}
\description{
Summary method for aras objects
}
