% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aroman.R
\name{summary.aroman}
\alias{summary.aroman}
\title{Summary method for aroman}
\usage{
\method{summary}{aroman}(obj)
}
\arguments{
\item{obj}{aroman object to process summary for}
}
\description{
Summary method for aroman
}
