% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.balanced_spotis.R
\name{summary.balanced_spotis}
\alias{summary.balanced_spotis}
\title{Summary method for balanced_spotis objects}
\usage{
\method{summary}{balanced_spotis}(obj)
}
\arguments{
\item{obj}{balanced_spotis object to process summary for}
}
\description{
Summary method for balanced_spotis objects
}
