% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.borda.R
\name{summary.borda}
\alias{summary.borda}
\title{Summary method for borda objects}
\usage{
\method{summary}{borda}(obj)
}
\arguments{
\item{obj}{borda object to process summary for}
}
\description{
Summary method for borda objects
}
