% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cocoso.R
\name{summary.cocoso}
\alias{summary.cocoso}
\title{Summary method for cocoso objects}
\usage{
\method{summary}{cocoso}(obj)
}
\arguments{
\item{obj}{cocoso object to process summary for}
}
\description{
Summary method for cocoso objects
}
