% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.codas.R
\name{summary.codas}
\alias{summary.codas}
\title{Summary method for codas objects}
\usage{
\method{summary}{codas}(obj)
}
\arguments{
\item{obj}{codas object to process summary for}
}
\description{
Summary method for codas objects
}
