% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.copeland.R
\name{summary.copeland}
\alias{summary.copeland}
\title{Summary method for copeland objects}
\usage{
\method{summary}{copeland}(obj)
}
\arguments{
\item{obj}{copeland object to process summary for}
}
\description{
Summary method for copeland objects
}
