% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.copras.R
\name{summary.copras}
\alias{summary.copras}
\title{Summary method for copras objects}
\usage{
\method{summary}{copras}(obj)
}
\arguments{
\item{obj}{copras object to process summary for}
}
\description{
Summary method for copras objects
}
