% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cradis.R
\name{summary.cradis}
\alias{summary.cradis}
\title{Summary method for cradis objects}
\usage{
\method{summary}{cradis}(obj)
}
\arguments{
\item{obj}{cradis object to process summary for}
}
\description{
Summary method for cradis objects
}
