% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.edas.R
\name{summary.edas}
\alias{summary.edas}
\title{Summary method for edas objects}
\usage{
\method{summary}{edas}(obj)
}
\arguments{
\item{obj}{edas object to process summary for}
}
\description{
Summary method for edas objects
}
