% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.electre1.R
\name{summary.electre1}
\alias{summary.electre1}
\title{Summary method for electre1 objects}
\usage{
\method{summary}{electre1}(obj)
}
\arguments{
\item{obj}{electre1 object to process summary for}
}
\description{
Summary method for electre1 objects
}
