% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.electre1s.R
\name{summary.electre1s}
\alias{summary.electre1s}
\title{Summary method for electre1s objects}
\usage{
\method{summary}{electre1s}(obj)
}
\arguments{
\item{obj}{electre1s object to process summary for}
}
\description{
Summary method for electre1s objects
}
