% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.electre2.R
\name{summary.electre2}
\alias{summary.electre2}
\title{Summary method for electre2 objects}
\usage{
\method{summary}{electre2}(obj)
}
\arguments{
\item{obj}{electre2 object to process summary for}
}
\description{
Summary method for electre2 objects
}
