% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.electre3.R
\name{summary.electre3}
\alias{summary.electre3}
\title{Summary method for electre3 objects}
\usage{
\method{summary}{electre3}(obj)
}
\arguments{
\item{obj}{electre3 object to process summary for}
}
\description{
Summary method for electre3 objects
}
