% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.electre4.R
\name{summary.electre4}
\alias{summary.electre4}
\title{Summary method for electre4 objects}
\usage{
\method{summary}{electre4}(obj)
}
\arguments{
\item{obj}{electre4 object to process summary for}
}
\description{
Summary method for electre4 objects
}
