% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.electretri.R
\name{summary.electretri}
\alias{summary.electretri}
\title{Summary method for electretri objects}
\usage{
\method{summary}{electretri}(obj)
}
\arguments{
\item{obj}{electretri object to process summary for}
}
\description{
Summary method for electretri objects
}
