% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ervd.R
\name{summary.ervd}
\alias{summary.ervd}
\title{Summary method for ervd objects}
\usage{
\method{summary}{ervd}(obj)
}
\arguments{
\item{obj}{ervd object to process summary for}
}
\description{
Summary method for ervd objects
}
