% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.evamix.R
\name{summary.evamix}
\alias{summary.evamix}
\title{Summary method for evamix objects}
\usage{
\method{summary}{evamix}(obj)
}
\arguments{
\item{obj}{evamix object to process summary for}
}
\description{
Summary method for evamix objects
}
