% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.extended_aroman.R
\name{summary.extended_aroman}
\alias{summary.extended_aroman}
\title{Summary method for extended_aroman objects}
\usage{
\method{summary}{extended_aroman}(obj)
}
\arguments{
\item{obj}{extended_aroman object to process summary for}
}
\description{
Summary method for extended_aroman objects
}
