% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fuca.R
\name{summary.fuca}
\alias{summary.fuca}
\title{Summary method for fuca objects}
\usage{
\method{summary}{fuca}(obj)
}
\arguments{
\item{obj}{fuca object to process summary for}
}
\description{
Summary method for fuca objects
}
