% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fuzzytopsis.R
\name{summary.fuzzytopsis}
\alias{summary.fuzzytopsis}
\title{Summary method for fuzzytopsis objects}
\usage{
\method{summary}{fuzzytopsis}(obj)
}
\arguments{
\item{obj}{fuzzytopsis object to process summary for}
}
\description{
Summary method for fuzzytopsis objects
}
