% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fuzzyvikor.R
\name{summary.fuzzyvikor}
\alias{summary.fuzzyvikor}
\title{Summary method for fuzzyvikor objects}
\usage{
\method{summary}{fuzzyvikor}(obj)
}
\arguments{
\item{obj}{fuzzyvikor object to process summary for}
}
\description{
Summary method for fuzzyvikor objects
}
