% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gra.R
\name{summary.gra}
\alias{summary.gra}
\title{Summary method for gra objects}
\usage{
\method{summary}{gra}(obj)
}
\arguments{
\item{obj}{gra object to process summary for}
}
\description{
Summary method for gra objects
}
