% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mabac.R
\name{summary.mabac}
\alias{summary.mabac}
\title{Summary method for mabac objects}
\usage{
\method{summary}{mabac}(obj)
}
\arguments{
\item{obj}{mabac object to process summary for}
}
\description{
Summary method for mabac objects
}
