% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.macbeth.R
\name{summary.macbeth}
\alias{summary.macbeth}
\title{Summary method for macbeth objects}
\usage{
\method{summary}{macbeth}(obj)
}
\arguments{
\item{obj}{macbeth object to process summary for}
}
\description{
Summary method for macbeth objects
}
