% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mairca.R
\name{summary.mairca}
\alias{summary.mairca}
\title{Summary method for MAIRCA objects}
\usage{
\method{summary}{mairca}(obj)
}
\arguments{
\item{obj}{MAIRCA object to process summary for}
}
\description{
Summary method for MAIRCA objects
}
