% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.marcos.R
\name{summary.marcos}
\alias{summary.marcos}
\title{Summary method for marcos objects}
\usage{
\method{summary}{marcos}(obj)
}
\arguments{
\item{obj}{marcos object to process summary for}
}
\description{
Summary method for marcos objects
}
