% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mare.R
\name{summary.mare}
\alias{summary.mare}
\title{Summary method for mare objects}
\usage{
\method{summary}{mare}(obj)
}
\arguments{
\item{obj}{mare object to process summary for}
}
\description{
Summary method for mare objects
}
