% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.maut.R
\name{summary.maut}
\alias{summary.maut}
\title{Summary method for maut objects}
\usage{
\method{summary}{maut}(obj)
}
\arguments{
\item{obj}{maut object to process summary for}
}
\description{
Summary method for maut objects
}
