% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.melchior.R
\name{summary.melchior}
\alias{summary.melchior}
\title{Summary method for melchior objects}
\usage{
\method{summary}{melchior}(obj)
}
\arguments{
\item{obj}{melchior object to process summary for}
}
\description{
Summary method for melchior objects
}
