% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.melichor.R
\name{summary.melichor}
\alias{summary.melichor}
\title{Summary method for melichor objects}
\usage{
\method{summary}{melichor}(obj)
}
\arguments{
\item{obj}{melichor object to process summary for}
}
\description{
Summary method for melichor objects
}
