% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.merec.R
\name{summary.merec}
\alias{summary.merec}
\title{Summary method for merec objects}
\usage{
\method{summary}{merec}(obj)
}
\arguments{
\item{obj}{merec object to process summary for}
}
\description{
Summary method for merec objects
}
