% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.moora.R
\name{summary.moora}
\alias{summary.moora}
\title{Summary method for moora objects}
\usage{
\method{summary}{moora}(obj)
}
\arguments{
\item{obj}{moora object to process summary for}
}
\description{
Summary method for moora objects
}
