% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.moosra.R
\name{summary.moosra}
\alias{summary.moosra}
\title{Summary method for moosra objects}
\usage{
\method{summary}{moosra}(obj)
}
\arguments{
\item{obj}{moosra object to process summary for}
}
\description{
Summary method for moosra objects
}
