% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.msim.R
\name{summary.msim}
\alias{summary.msim}
\title{Summary method for MSIM objects}
\usage{
\method{summary}{msim}(obj)
}
\arguments{
\item{obj}{MSIM object to process summary for}
}
\description{
Summary method for MSIM objects
}
