% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ocra.R
\name{summary.ocra}
\alias{summary.ocra}
\title{Summary method for ocra objects}
\usage{
\method{summary}{ocra}(obj)
}
\arguments{
\item{obj}{ocra object to process summary for}
}
\description{
Summary method for ocra objects
}
