% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.oreste.R
\name{summary.oreste}
\alias{summary.oreste}
\title{Summary method for oreste objects}
\usage{
\method{summary}{oreste}(obj)
}
\arguments{
\item{obj}{oreste object to process summary for}
}
\description{
Summary method for oreste objects
}
