% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pammsem.R
\name{summary.pamssem}
\alias{summary.pamssem}
\title{Summary method for pamssem objects}
\usage{
\method{summary}{pamssem}(obj)
}
\arguments{
\item{obj}{pamssem object to process summary for}
}
\description{
Summary method for pamssem objects
}
