% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.piv.R
\name{summary.piv}
\alias{summary.piv}
\title{Summary method for piv objects}
\usage{
\method{summary}{piv}(obj)
}
\arguments{
\item{obj}{piv object to process summary for}
}
\description{
Summary method for piv objects
}
