% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.probid.R
\name{summary.probid}
\alias{summary.probid}
\title{Summary method for probid objects}
\usage{
\method{summary}{probid}(obj)
}
\arguments{
\item{obj}{probid object to process summary for}
}
\description{
Summary method for probid objects
}
