% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.promethee1.R
\name{summary.promethee1}
\alias{summary.promethee1}
\title{Summary method for promethee1 objects}
\usage{
\method{summary}{promethee1}(obj)
}
\arguments{
\item{obj}{promethee1 object to process summary for}
}
\description{
Summary method for promethee1 objects
}
