% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.promethee2.R
\name{summary.promethee2}
\alias{summary.promethee2}
\title{Summary method for promethee2 objects}
\usage{
\method{summary}{promethee2}(obj)
}
\arguments{
\item{obj}{promethee2 object to process summary for}
}
\description{
Summary method for promethee2 objects
}
