% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.promethee3.R
\name{summary.promethee3}
\alias{summary.promethee3}
\title{Summary method for promethee3 objects}
\usage{
\method{summary}{promethee3}(obj)
}
\arguments{
\item{obj}{promethee3 object to process summary for}
}
\description{
Summary method for promethee3 objects
}
