% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.psi.R
\name{summary.psi}
\alias{summary.psi}
\title{Summary method for psi objects}
\usage{
\method{summary}{psi}(obj)
}
\arguments{
\item{obj}{psi object to process summary for}
}
\description{
Summary method for psi objects
}
