% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qualiflex.R
\name{summary.qualiflex}
\alias{summary.qualiflex}
\title{Summary method for qualiflex objects}
\usage{
\method{summary}{qualiflex}(obj)
}
\arguments{
\item{obj}{qualiflex object to process summary for}
}
\description{
Summary method for qualiflex objects
}
