% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r_method.R
\name{summary.r_method}
\alias{summary.r_method}
\title{Summary method for r_method objects}
\usage{
\method{summary}{r_method}(obj)
}
\arguments{
\item{obj}{r_method object to process summary for}
}
\description{
Summary method for r_method objects
}
