% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rafsi.R
\name{summary.rafsi}
\alias{summary.rafsi}
\title{Summary method for rafsi objects}
\usage{
\method{summary}{rafsi}(obj)
}
\arguments{
\item{obj}{aras object to process summary for}
}
\description{
Summary method for rafsi objects
}
