% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ram.R
\name{summary.ram}
\alias{summary.ram}
\title{Summary method for ram objects}
\usage{
\method{summary}{ram}(obj)
}
\arguments{
\item{obj}{ram object to process summary for}
}
\description{
Summary method for ram objects
}
