% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rawec.R
\name{summary.rawec}
\alias{summary.rawec}
\title{Summary method for rawec objects}
\usage{
\method{summary}{rawec}(obj)
}
\arguments{
\item{obj}{rawec object to process summary for}
}
\description{
Summary method for rawec objects
}
