% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regime.R
\name{summary.regime}
\alias{summary.regime}
\title{Summary method for regime objects}
\usage{
\method{summary}{regime}(obj)
}
\arguments{
\item{obj}{regime object to process summary for}
}
\description{
Summary method for regime objects
}
