% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rim.R
\name{summary.rim}
\alias{summary.rim}
\title{Summary method for rim objects}
\usage{
\method{summary}{rim}(obj)
}
\arguments{
\item{obj}{rim object to process summary for}
}
\description{
Summary method for rim objects
}
