% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rov.R
\name{summary.rov}
\alias{summary.rov}
\title{Summary method for rov objects}
\usage{
\method{summary}{rov}(obj)
}
\arguments{
\item{obj}{rov object to process summary for}
}
\description{
Summary method for rov objects
}
