% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.saw.R
\name{summary.saw}
\alias{summary.saw}
\title{Summary method for saw objects}
\usage{
\method{summary}{saw}(obj)
}
\arguments{
\item{obj}{saw object to process summary for}
}
\description{
Summary method for saw objects
}
