% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seca.R
\name{summary.seca}
\alias{summary.seca}
\title{Summary method for seca objects}
\usage{
\method{summary}{seca}(obj)
}
\arguments{
\item{obj}{seca object to process summary for}
}
\description{
Summary method for seca objects
}
