% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sir.R
\name{summary.sir}
\alias{summary.sir}
\title{Summary method for sir objects}
\usage{
\method{summary}{sir}(obj)
}
\arguments{
\item{obj}{sir object to process summary for}
}
\description{
Summary method for sir objects
}
