% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.smart.R
\name{summary.smart}
\alias{summary.smart}
\title{Summary method for smart objects}
\usage{
\method{summary}{smart}(obj)
}
\arguments{
\item{obj}{smart object to process summary for}
}
\description{
Summary method for smart objects
}
