% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spotis.R
\name{summary.spotis}
\alias{summary.spotis}
\title{Summary method for aras objects}
\usage{
\method{summary}{spotis}(obj)
}
\arguments{
\item{obj}{aras object to process summary for}
}
\description{
Summary method for aras objects
}
