% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sprobid.R
\name{summary.sprobid}
\alias{summary.sprobid}
\title{Summary method for sprobid objects}
\usage{
\method{summary}{sprobid}(obj)
}
\arguments{
\item{obj}{sprobid object to process summary for}
}
\description{
Summary method for sprobid objects
}
