% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.todim.R
\name{summary.todim}
\alias{summary.todim}
\title{Summary method for todim objects}
\usage{
\method{summary}{todim}(obj)
}
\arguments{
\item{obj}{todim object to process summary for}
}
\description{
Summary method for todim objects
}
