% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.topsis.R
\name{summary.topsis}
\alias{summary.topsis}
\title{Summary method for topsis objects}
\usage{
\method{summary}{topsis}(obj)
}
\arguments{
\item{obj}{topsis object to process summary for}
}
\description{
Summary method for topsis objects
}
