% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.vikor.R
\name{summary.vikor}
\alias{summary.vikor}
\title{Summary method for vikor objects}
\usage{
\method{summary}{vikor}(obj)
}
\arguments{
\item{obj}{vikor object to process summary for}
}
\description{
Summary method for vikor objects
}
