% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.waspas.R
\name{summary.waspas}
\alias{summary.waspas}
\title{Summary method for waspas objects}
\usage{
\method{summary}{waspas}(obj)
}
\arguments{
\item{obj}{waspas object to process summary for}
}
\description{
Summary method for waspas objects
}
