% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.wisp.R
\name{summary.wisp}
\alias{summary.wisp}
\title{Summary method for wisp objects}
\usage{
\method{summary}{wisp}(obj)
}
\arguments{
\item{obj}{wisp object to process summary for}
}
\description{
Summary method for wisp objects
}
