% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.wpm.R
\name{summary.wpm}
\alias{summary.wpm}
\title{Summary method for wpm objects}
\usage{
\method{summary}{wpm}(obj)
}
\arguments{
\item{obj}{wpm object to process summary for}
}
\description{
Summary method for wpm objects
}
