% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.wsm.R
\name{summary.wsm}
\alias{summary.wsm}
\title{Summary method for wsm objects}
\usage{
\method{summary}{wsm}(obj)
}
\arguments{
\item{obj}{wsm object to process summary for}
}
\description{
Summary method for wsm objects
}
