% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todim.R
\name{todim}
\alias{todim}
\title{TOmada de Decisao Interativa Multicriterio}
\description{
Multicriteria decision analysis method based on diminance measure.
 Computation starts with normalizing the performance matrix (aleternatives
 in rows and criteria in columns) using linear normalization. Then relative
 weights are computed.

\mjsdeqn{w_{cr} = \frac{w_c}{max w}}

Then dominance degree of  the alternatives is computed using

\mjsdeqn{\delta (A_i, A_j) = \sum_{c=1}^m \Theta_c(A_i, A_j) \quad \forall(i,j)}

\mjsdeqn{\Theta (A_i, A_j) = \left\lbrace\begin{array}{ll} \sqrt{\frac{w_{cr}(P_{ic} - P_{jc})}{\sum_{c=1}^m w_{cr}}} & \text{ if }(P_{ic} - P_{jc}) > 0 \cr 0 & \text{ if }(P_{ic} - P_{jc}) = 0 \cr \frac{-1}{\theta}\sqrt{\frac{(\sum_{c=1}^m w_{cr})(P_{ic} - P_{jc})}{w_{cr}}} & \text{ if }(P_{ic} - P_{jc}) < 0\end{array}\right.}

Theta symbol is loss damping factor. Different values lead to different
 shapes of the prospect theoretical value function.

Finally overall dominance degree for alternatives is computed. This measure
 is directly usable to establish ranking of the alternatives.

\mjsdeqn{\xi_i = \frac{\sum_{j=1}^m \delta(A_i, A_j) - min \sum_{j=1}^m \delta(A_i, A_j)}{max \sum_{j=1}^m \delta(A_i, A_j) - min \sum_{j=1}^m \delta(A_i, A_j)}}
}
\examples{

## ------------------------------------------------
## Method `todim$new`
## ------------------------------------------------

# from: https://www.youtube.com/watch?v=VetfNIHLyn0
alternatives <- c("AI 2024-T6", "AI 5052-O", "SS 301 FH", "SS 310-3AH",
 "Ti-6AI-4V", "Inconel 718", "70Cu-30Zn")
criteria <- c("Toughness index", "Yield Strength", "Young's Modulus",
 "Density", "Thermal expansion", "Thermal conductivity", "Specific Heat")
pm <- cbind(
 c(75.5, 95, 770, 187, 179, 239, 273),
 c(420, 91, 1365, 1120, 875, 1190, 200),
 c(74.2, 70, 189, 210, 112, 217, 112),
 c(2.8, 2.68, 7.9, 7.9, 4.43, 8.51, 8.53),
 c(21.4, 22.1, 16.9, 14.4, 9.4, 11.5, 19.9),
 c(0.37, 0.33, 0.04, 0.03, 0.016, 0.31, 0.29),
 c(0.16, 0.16, 0.08, 0.08, 0.09, 0.07, 0.06)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.28, 0.14, 0.05, 0.24, 0.19, 0.05, 0.05)
minmax <- c("max", "max", "max", "min", "min", "min", "min")
t <- todim$new(pm, w, minmax)
}
\references{
SOLÍS, Irvin David Bonilla, PÉREZ-DOMÍNGUEZ, Luis Asunción, DELGADO, Rosa
 Patricia Ramírez, DÍAZ, Marling Carolina Cordero. Web Application
 Development for TODIM Method Automation and Alternatives Evaluation. Data
 and Metadata. 2025, roč. 4, s. 492–492, dostupné z:
 https://doi.org/10.56294/dm2025492, ISSN 2953-4917.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{TODIM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (unnormalized) performance matrix of
alternatives (in rows) in criteria (columns)}

\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{weight vector}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria.}

\item{\code{results}}{overall dominance degree alternative (best = max)}

\item{\code{results_sorted}}{by dominance degree alternative (best = max)}

\item{\code{result}}{score and rank dataframe}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-todim-new}{\code{todim$new()}}
\item \href{#method-todim-compute}{\code{todim$compute()}}
\item \href{#method-todim-summary}{\code{todim$summary()}}
\item \href{#method-todim-clone}{\code{todim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-todim-new"></a>}}
\if{latex}{\out{\hypertarget{method-todim-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{todim$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{w}}{weight vector (sum of weights = 1)}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# from: https://www.youtube.com/watch?v=VetfNIHLyn0
alternatives <- c("AI 2024-T6", "AI 5052-O", "SS 301 FH", "SS 310-3AH",
 "Ti-6AI-4V", "Inconel 718", "70Cu-30Zn")
criteria <- c("Toughness index", "Yield Strength", "Young's Modulus",
 "Density", "Thermal expansion", "Thermal conductivity", "Specific Heat")
pm <- cbind(
 c(75.5, 95, 770, 187, 179, 239, 273),
 c(420, 91, 1365, 1120, 875, 1190, 200),
 c(74.2, 70, 189, 210, 112, 217, 112),
 c(2.8, 2.68, 7.9, 7.9, 4.43, 8.51, 8.53),
 c(21.4, 22.1, 16.9, 14.4, 9.4, 11.5, 19.9),
 c(0.37, 0.33, 0.04, 0.03, 0.016, 0.31, 0.29),
 c(0.16, 0.16, 0.08, 0.08, 0.09, 0.07, 0.06)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.28, 0.14, 0.05, 0.24, 0.19, 0.05, 0.05)
minmax <- c("max", "max", "max", "min", "min", "min", "min")
t <- todim$new(pm, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-todim-compute"></a>}}
\if{latex}{\out{\hypertarget{method-todim-compute}{}}}
\subsection{Method \code{compute()}}{
computes TODIM model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{todim$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-todim-summary"></a>}}
\if{latex}{\out{\hypertarget{method-todim-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the TODIM method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{todim$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-todim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-todim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{todim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
