% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topsis_ideal.R
\name{topsis_ideal}
\alias{topsis_ideal}
\title{function to compute ideal and antiedeal solution for TOPSIS and SIR
 (internal use only)}
\usage{
topsis_ideal(pm, inferiority = FALSE)
}
\arguments{
\item{pm}{normalized performance matrix}

\item{inferiority}{FALSE by defaul to compute ideals as max from pm, TRUE
will minimize ideal and maximize anitiideal}
}
\value{
Returns list with following structure:
\itemize{
  \item a_ideal - Positive ideal soluation A*
  \item a_anti - Anti-ideal solution A-
  \item d_ideal - Closenes to ideal variant
  \item d_anti - Closeness to anti-ideal variant
  \item closenes - Relative closeness of the alternatives to ideal solution
}
}
\description{
Function to compute ideal and antiideal solution to normalized perfomance
 matrix.

The function is utilized by \link{sir} and \link{topsis} methods for
 this purpose. It incorporates steps 3 - 5 of the TOPSIS computation.
 Consult \link{topsis} function's documentation for implementation details.
}
\references{
ALAOUI, Mohamed El. Fuzzy TOPSIS: Logic, Approaches, and Case Studies. Boca
 Raton: CRC Press, 2021. 216 s. ISBN 978-0-367-76748-8.

Papathanasiou, Jason, Ploskas, Nikolaos. Multiple Criteria Decision Aid
 Methods, Examples and Python Implementations. Springer, 173 p., ISBN
 978-3-319-91648-4.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{SIR}
\keyword{TOSPIS}
\keyword{anti-ideal}
\keyword{ideal}
