% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_invalid_val}
\alias{validation$validate_invalid_val}
\title{Validate that the matrix or dataframe has only valid values}
\arguments{
\item{m}{matrix or dataframe to check}

\item{valid_val}{vector of valid values to check}

\item{msg}{Identification of what are we checking (i.e. performance matrix)}
}
\description{
Vector of valid values must contain full set of valid values to check
 against. If other value is detected the error will be generated and
 computation run will be stoped.
}
\examples{
PM <- rbind(
   c(0,1,1,0,1,1),
   c(0,0,0,0,1,1),
   c(0,1,0,0,1,1),
   c(1,1,1,0,1,1),
   c(0,0,0,0,0,1),
   c(0,0,0,0,0,0)
)
rownames(PM) <- colnames(PM) <- c("C1", "C2", "C4", "C5", "C6", "C8")
validation$validate_invalid_val(PM, c(0, 1), "Preference matrix")
}
