% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_matrix_numeric}
\alias{validation$validate_matrix_numeric}
\title{Validates that the matrix or dataframe has only numeric elements}
\arguments{
\item{pm}{matrix or dataframe to check}
}
\description{
Validates that the matrix or dataframe has only numeric elements
}
\examples{
alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB', 'KRK',
  'MAA', 'EZE')
  M <- rbind(
    c(0, 0, 0, 1, 1, 0, 0, 0, 1, 0),
    c(0, 0, 1, 1, 1, 1, 0, 0, 1, 0),
    c(0, 0, 0, 1, 1, 0, 0, 0, 1, 0),
     c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
     c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
     c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0),
     c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
     c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
     c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
     c(0, 0, 0, 0, 1, 1, 0, 0, 0, 0)
   )
   rownames(M) <- alternatives
   colnames(M) <- alternatives
validation$validate_matrix_numeric(M)
}
