% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_minmax}
\alias{validation$validate_minmax}
\title{Validate min-max vector (internal use only)}
\arguments{
\item{minmax}{vector of min or max values signifying optimization direction
of the criteria}

\item{ncri}{number of criteria in decision problem}
}
\value{
validated vector of min/max for criteria
}
\description{
Minmax parameter should use only the "min" or "max" values. Appart of
 testing this the function also deals with the parameter set as single min
 or max (all criteria in same optimization direction) and creates full
 vector from this information.

Also test if number of criteria in minmax is the same as no. of criteria in
 ncri.
}
