% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_no_elements_vs_cri}
\alias{validation$validate_no_elements_vs_cri}
\title{Validates that the number of elements in the vector is same as number of
 criteria}
\arguments{
\item{vect}{vector to check}

\item{ncri}{number of criteria}

\item{msg}{identification of what are we checking to use in error message}

\item{test_numeric}{set to TRUE to test vect as numeric vector}
}
\description{
Validates that the number of elements in the vector is same as number of
 criteria
}
