% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_pm_01_symetry}
\alias{validation$validate_pm_01_symetry}
\title{Validates that to 1 in rows there is symetrical 0 in columns (and vice
 versa)}
\arguments{
\item{pm}{performance matrix}
}
\description{
The preference matrix for binary preference comparison is symetrical if
 a is found more important then b (represented by 1) then b must be less
 important than a (represented by 0). This function checks that this is the
 case
}
