% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_pm_0_or_1}
\alias{validation$validate_pm_0_or_1}
\title{Validate, that the metrix has only 0/1 in it}
\arguments{
\item{pm}{performance matrix}
}
\description{
Check whether there are other elements then 0 or 1 in the matrix.
}
