% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_pm_fuzzy}
\alias{validation$validate_pm_fuzzy}
\title{Validate consistency of performance matrix specified using fuzzy numbers}
\arguments{
\item{pm}{performance matrix}

\item{dict_pm}{dictionary for performance matrix}

\item{w}{weight matrix}

\item{dict_w}{dictionary for weight matrix}

\item{fn}{type of fuzzy number (i.e. 3 for triangular, 4 for trapeziodal)}

\item{alt}{vector of alternatives names}
}
\description{
Validate consistency of performance matrix specified using fuzzy numbers
}
