% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_pm_rows_columns_same}
\alias{validation$validate_pm_rows_columns_same}
\title{Validate that the number of rows and columns is same}
\arguments{
\item{pm}{performance matrix}
}
\description{
Validate that the number of rows and columns is same
}
