% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_scalar_numeric}
\alias{validation$validate_scalar_numeric}
\title{Validate that scalar is number}
\arguments{
\item{s}{the value to be checked}

\item{msg}{identification of what failed to check (i.e. "Discrimination
threshold") must be a number.}
}
\description{
Validate that scalar is number
}
\examples{
psi <- 0.5
validation$validate_scalar_numeric(psi, "coefficient of determination")
}
