% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_scalar_same}
\alias{validation$validate_scalar_same}
\title{Validates that provided params val1 and 2 are same}
\arguments{
\item{val1}{first value to check}

\item{val2}{second value to check}

\item{msg}{identification of what we are checking}
}
\description{
Validates that provided params val1 and 2 are same
}
\examples{
validation$validate_scalar_same(5, 5, "message")
}
