% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_value_in_interval}
\alias{validation$validate_value_in_interval}
\title{Validates that provided val is in <from;to> interval.}
\arguments{
\item{val}{numeric value to check}

\item{from}{lower bound of the interval to check}

\item{to}{upper bound of interval to check}

\item{msg}{identification of what we are checking}
}
\description{
Validates that provided val is in <from;to> interval.
}
\examples{
validation$validate_value_in_interval(10, 0, 100, "value")
}
