% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_vector_progression}
\alias{validation$validate_vector_progression}
\title{Validate the numeric progresion in provided vector}
\arguments{
\item{vect}{numeric vector with values to check}
}
\description{
Presumes that the elements in the vector are ordered ascending. This is
 usefull for various threshold checks.
}
\examples{
c_minus <- 0.65
c_zero <- 0.75
c_plus <- 0.85
d_minus <- 0.25
d_plus <- 0.5
thres <- c(d_minus, d_plus, c_minus, c_zero, c_plus)
validation$validate_vector_progression(thres)
}
