% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_vector_progression2}
\alias{validation$validate_vector_progression2}
\title{Validate that vect1 < vect2 (element wise)}
\arguments{
\item{vect1}{first vector to compare (smaller)}

\item{vect2}{second vector to comapare (larger)}

\item{msg}{string specifying what are we comparing}
}
\description{
Perform check by comparint elements in vect1 with vect2 to ensure that
 every vect1_i < vect2_i, where i are the elements in the vector.

This function is a variant of validate_vector_progression() function but it
 works with 2 vectors instead of progression of scalars.

The function is intended for validation consistency of various shresholds.
}
