% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_w}
\alias{validation$validate_w}
\title{validate weight vector}
\arguments{
\item{w}{weight vector}

\item{ncri}{number of criteria}
}
\description{
Validates that the weights vector is numeric and number of elements in it
 is equal to number of criteria.
}
