% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_w_sum_eq_1}
\alias{validation$validate_w_sum_eq_1}
\title{Validate that sum of weights is equal to 1}
\arguments{
\item{w}{weight vector}
}
\description{
Validate that sum of weights is equal to 1
}
