% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validation_vector_in_interval}
\alias{validation$validation_vector_in_interval}
\title{Validates tahl all values in the vactor are in specified bounds}
\arguments{
\item{vect}{numeric vector to check}

\item{from}{lower bound of the checking interval}

\item{to}{upper bound for the checking interval}

\item{msg}{identification of object we are checking (i.e. indifference
threshold)}
}
\description{
Validates tahl all values in the vactor are in specified bounds
}
