% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_compare.R
\name{vector_compare}
\alias{vector_compare}
\title{comparest two vectors element wise}
\usage{
vector_compare(vector1, vector2)
}
\arguments{
\item{vector1}{first vector to compare}

\item{vector2}{second vector to compare}
}
\value{
TRUE if all elements of the vector are same, otherwise
returns FALSE
}
\description{
`vector_compare` compares two vectors on per element basis. The function
 is intended to be internaly used by various sensitivity evaluation
 functions, which presumes, that provided vectors do have same structure,
 are ordered is same way (or more specifically are not ordered).
}
\examples{
a <- b <- c(1, 2, 3, 4, 5)
c <- c(2, 2, 3, 4, 5)
d <- c(1, 2, 3)
vector_compare(a, b) # TRUE
vector_compare(a, c) # FALSE
vector_compare(a, d) # error
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{II}
\keyword{III}
\keyword{comparison}
\keyword{vector}
