% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIKORIndexes.R
\name{VIKORIndexes}
\alias{VIKORIndexes}
\title{subroutine to compute S & R index in VIKOR and FuzzyVIKOR methods}
\usage{
VIKORIndexes(car, bw_perf, cw, v = NULL)
}
\arguments{
\item{car}{array with the performances (crisp alternative ratings in fuzzy
variant). Alternatives are in rows, criteria in columns}

\item{cw}{vector containing the weights of the criteria.}

\item{v}{weight for strategy of majority of the criteria (0-1)}

\item{bw_pref}{matrix  with the best, worst performances and differences
between them. Has columns (best, worst, difference) and rows for criteria}
}
\value{
The function returns a list structured as follows:
\itemize{
   \item S - ordered list of alternatives using S-metric
   \item R -ordered list of alternatives using R-metric
   \item Q - ordered list of alternatives using Q-metric
   \item compromiseSolution - list of alternatives forming compromise
 solution (based on Q, S, R metrics)
}
}
\description{
The acronym VIKOR stands for: VlseKriterijumska Optimizacija I Kompromisno
 Resenje, in Serbian multicriteria optimization and compromise solution.

The method presents generalized solution for computation of S and R indexes
 using wights, performance matrix (crisp performance matrix if using fuzzy
 numbers) and information on best and worst values in criteria.
}
\references{
ALAOUI, Mohamed El. Fuzzy TOPSIS: Logic, Approaches, and Case Studies. Boca
 Raton: CRC Press, 2021. 216 s. ISBN 978-0-367-76748-8.

Papathanasiou, Jason, Ploskas, Nikolaos. Multiple Criteria Decision Aid
 Methods, Examples and Python Implementations. Springer, 173 p., ISBN
 978-3-319-91648-4.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{FuzzyVIKOR}
\keyword{VIKOR}
