% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm.R
\name{wpm}
\alias{wpm}
\title{Weighted Product Model AKA MEW}
\description{
One of original (simpliest) models in MCDA. Works similarly to \link{wsm},
 but instead of summing weighted performance of the alternatives it is based
 on product of performances of alternatives to power of weights.

Mathematically this is expressed as:

\mjsdeqn{P(A_K) = \prod_{j=1}^{n} a_{Kj}^{w_j}}

Where a is performance of alternative K in sriterium j, n is number of
 criteria and w are weights. Computed value of P is directly comparable.

Note that the method is also known under different name: MEW Multiplicative
 Exponent Weighting.
}
\examples{

## ------------------------------------------------
## Method `wpm$new`
## ------------------------------------------------

PM <- rbind(
 c(25, 20, 15, 30),
 c(10, 30, 20, 30),
 c(30, 10, 30, 10)
)
rownames(PM) <- c("A1", "A2", "A3")
colnames(PM) <- c("C1", "C2", "C3", "C4")
w <- c(0.2, 0.15, 0.4, 0.25)
t <- wpm$new(PM, w)
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MEW}
\keyword{WPM}
\keyword{WSM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{normalized performance matrix alternatives in rows, criteria
in columns}

\item{\code{w}}{weights, sum of weights = 1}

\item{\code{p}}{final weighted product of the alternative}

\item{\code{result}}{dataframe with weighted product of alternatives (p) and
ranks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wpm-new}{\code{wpm$new()}}
\item \href{#method-wpm-compute}{\code{wpm$compute()}}
\item \href{#method-wpm-summary}{\code{wpm$summary()}}
\item \href{#method-wpm-clone}{\code{wpm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wpm-new"></a>}}
\if{latex}{\out{\hypertarget{method-wpm-new}{}}}
\subsection{Method \code{new()}}{
validates input parameter, runs the computation of the WPM model and
 returns WPM object with the inputs and results of the model.

WPM does not have any means of normalization, so it is up to the
 analytic to provide the normalized performance matrix. This also
 presumes that the cost/benefit criteria have already been resolved
 outside the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wpm$new(pm, w)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix alternatives in rows, criteria
in columns}

\item{\code{w}}{weights, sum of weights = 1}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of WPM class with results
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{PM <- rbind(
 c(25, 20, 15, 30),
 c(10, 30, 20, 30),
 c(30, 10, 30, 10)
)
rownames(PM) <- c("A1", "A2", "A3")
colnames(PM) <- c("C1", "C2", "C3", "C4")
w <- c(0.2, 0.15, 0.4, 0.25)
t <- wpm$new(PM, w)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wpm-compute"></a>}}
\if{latex}{\out{\hypertarget{method-wpm-compute}{}}}
\subsection{Method \code{compute()}}{
Computes the WPM model based on the properties of the class. Usually it
 does not need to be run manually as this function is run by the class's
 constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wpm$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wpm-summary"></a>}}
\if{latex}{\out{\hypertarget{method-wpm-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the WPM method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wpm$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wpm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wpm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wpm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
