% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpNode.R
\name{ahpNode}
\alias{ahpNode}
\title{class for creation of AHP hierarchies}
\description{
class for creation of AHP hierarchies

class for creation of AHP hierarchies
}
\examples{
pm <- rbind(
  c(1, 5, 7, 5, 9, 9),
  c(0.2, 1, 5, 5, 7, 7),
  c(0.1429, 0.2, 1, 0.3333, 3, 3),
  c(0.2, 0.2, 3, 1, 5, 5),
  c(0.1111, 0.1429, 0.3333, 0.2, 1, 0.3333),
  c(0.1111, 0.1429, 0.3333, 0.2, 3, 1)
)
rownames(pm) <- colnames(pm) <- c("citizens",	"CI (elements)",
  "public infrastrucutre (trasport)",	"PI (technical infr.)",
  "PI (civil amenities)", "environment")
t <- ahpNode$new(pm = pm, leaf = FALSE, parent_w = 1)

citizens <- rbind(
  c(1, 7, 8),
  c(0.14, 1, 3),
  c(0.125, 0.33, 1)
)
rownames(citizens) <- colnames(citizens) <- c("no. of cit.",
  "characterstic of development",	"public space")
t$child_nodes[[1]] <- ahpNode$new(citizens, leaf = TRUE,
  parent_w = t$scaled_weights[1], name = "citizens")

pi_transport <- rbind(
  c(1, 0.33, 3, 7),
  c(3, 1, 5, 7),
  c(0.33, 0.2, 1, 3),
  c(0.14, 0.14, 0.33, 1)
)
rownames(pi_transport) <- colnames(pi_transport) <- c("rail	road",
  "airports", "water", "ways")
t$child_nodes[[3]] <- ahpNode$new(pi_transport, leaf = TRUE,
  parent_w = t$scaled_weights[3], name = "public infrastrucutre (trasport)")

pi_tech <- rbind(
  c(1, 0.33, 0.33, 5),
  c(3, 1, 3, 5),
  c(3, 0.33, 1, 5),
  c(0.2, 0.2, 0.2, 1)
)
rownames(pi_tech) <- colnames(pi_tech) <- c("power line",
  "public water supply", "sewer watters", "gas pipeline")
t$child_nodes[[4]] <- ahpNode$new(pi_tech, leaf = TRUE,
  parent_w = t$scaled_weights[4], name = "PI (technical infr.)")

pi_civil <- rbind(
  c(1, 3),
  c(0.33, 1)
)
rownames(pi_civil) <- colnames(pi_civil) <- c("important assets",
  "cultural heritage")
t$child_nodes[[5]] <- ahpNode$new(pi_civil, leaf = TRUE,
  parent_w = t$scaled_weights[5], name = "PI (civil amenities)")

env1 <- rbind(
  c(1, 5),
  c(0.2, 1)
)
rownames(env1) <- colnames(env1) <- c("biotic", "bonity of soil")
t$child_nodes[[6]] <- ahpNode$new(pi_civil, leaf = TRUE, parent_w = t$scaled_weights[6], "environment")
t$textTree()
t$plotTree()
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{preference matrix}

\item{\code{leaf}}{TRUE if the node is leaf (has no child)}

\item{\code{parent_w}}{weight of the parent node that is being distributed
into weights of the criteria}

\item{\code{child_nodes}}{list of children nodes (of ahpNode type)}

\item{\code{scaled_weights}}{rescaled AHP priority vector to the size of
parent_w (scaled_weights sums up to parent_w)}

\item{\code{CR}}{consistency ratio}

\item{\code{name}}{name of the node in the hierarchy}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ahpNode-new}{\code{ahpNode$new()}}
\item \href{#method-ahpNode-plotTree}{\code{ahpNode$plotTree()}}
\item \href{#method-ahpNode-textTree}{\code{ahpNode$textTree()}}
\item \href{#method-ahpNode-to_data_tree}{\code{ahpNode$to_data_tree()}}
\item \href{#method-ahpNode-clone}{\code{ahpNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ahpNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ahpNode-new}{}}}
\subsection{Method \code{new()}}{
Creates node in AHP hierarchy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ahpNode$new(pm, leaf = TRUE, parent_w = 1, name = "Root")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{preference matrix}

\item{\code{leaf}}{TRUE if the node is leaf (has no child)}

\item{\code{parent_w}}{weight of the parent node that is being distributed
into weights of the criteria}

\item{\code{name}}{name of the node in the hierarchy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ahpNode-plotTree"></a>}}
\if{latex}{\out{\hypertarget{method-ahpNode-plotTree}{}}}
\subsection{Method \code{plotTree()}}{
plots the tree from this node downwar in the hierarchy. Call from root
 node of the hierarchy to plot whole tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ahpNode$plotTree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ahpNode-textTree"></a>}}
\if{latex}{\out{\hypertarget{method-ahpNode-textTree}{}}}
\subsection{Method \code{textTree()}}{
prints tree in text form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ahpNode$textTree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ahpNode-to_data_tree"></a>}}
\if{latex}{\out{\hypertarget{method-ahpNode-to_data_tree}{}}}
\subsection{Method \code{to_data_tree()}}{
convert R6 structure to data.tree object for easy processing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ahpNode$to_data_tree(node_obj = self)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_obj}}{ahpNode object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ahpNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ahpNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ahpNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
