% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aroman.R
\name{aroman}
\alias{aroman}
\title{Alternative Ranking Order Method Accounting for Two-Step Normalization}
\description{
The approach uses linear (tij) and vector (tij*) normalization of
 performance measures. Based of these averaged normalization is computed:

\mjsdeqn{t_{ij}^{norm} = \frac{\beta t_{ij} + (1 - \beta)t_{ij}^*}{2}}

Where bata is weighting factor in interval 0-1, 0.5 by default.

To t_norm weights are appied (t^). Next we separately aggreates for cost
 criteria (Li) and benefit criteria (Ai).

As last step final ranking Ri is computed:

\mjsdeqn{R_i = L_i^\lambda + A_i^{1 - \lambda}}

Lambda parameter represents coefficient degree of the criteria type. It is
 0.5 by default, but any number 0-1 can be assigned. As rule of thumb the
 ratio between cost and benefit criteria is good starting point.
}
\examples{

## ------------------------------------------------
## Method `aroman$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- aroman$new(pm, w = w, minmax = minmax)
}
\references{
S. Bošković, L. Švadlenka, S. Jovčić, M. Dobrodolac, V. Simić and
 N. Bacanin. "An Alternative Ranking Order Method Accounting for Two-Step
 Normalization (AROMAN)—A Case Study of the Electric Vehicle Selection
 Problem," in IEEE Access, vol. 11, pp. 39496-39507, 2023,
 doi: 10.1109/ACCESS.2023.3265818.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{AROMAN}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{w}}{weight vector}

\item{\code{beta}}{value 0-1 representing weighting factor between 2 types of
normalization (default = 0.5)}

\item{\code{lambda}}{values 0-1, coefficient degree of the criteria type (0.5
by default)}

\item{\code{result}}{dataframe with cost (Li) and Bbenefit (Ai) normalized
aggregated value, final result (Ri) and rank of alternatives.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aroman-new}{\code{aroman$new()}}
\item \href{#method-aroman-compute}{\code{aroman$compute()}}
\item \href{#method-aroman-summary}{\code{aroman$summary()}}
\item \href{#method-aroman-clone}{\code{aroman$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aroman-new"></a>}}
\if{latex}{\out{\hypertarget{method-aroman-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aroman$new(pm, w, minmax = "max", beta = 0.5, lambda = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{beta}}{value 0-1 representing weighting factor between 2 types of
normalization (default = 0.5)}

\item{\code{lambda}}{values 0-1, coefficient degree of the criteria type (0.5
by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- aroman$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aroman-compute"></a>}}
\if{latex}{\out{\hypertarget{method-aroman-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of AROMAN model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aroman$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aroman-summary"></a>}}
\if{latex}{\out{\hypertarget{method-aroman-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the AROMAN method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aroman$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aroman-clone"></a>}}
\if{latex}{\out{\hypertarget{method-aroman-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aroman$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
