% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borda.R
\name{borda}
\alias{borda}
\title{Borda voting system (AKA Borda count/order of merit)}
\description{
Decision making method derived from voting system developed by
 Jean-Charles de Borda in 1770. The method assigns the points based on the
 rank.

On ballot last candidate gets 1 point, 2nd worse 2 points, ... Points for
 candidates are then aggregated across all balots, which gives us results
 for the elections.

In context of multi-criteria decision making the approach can be applied by
 assigning points based on ranking performance of the alternatives in the
 criteria and then suming the points for alternative across the criteria
 leading to measure directly usable to rank the alternatives.
}
\examples{

## ------------------------------------------------
## Method `borda$new`
## ------------------------------------------------

cri <- c("C1", "C2", "C3")
alt <- c("A1", "A2", "A3", "A4", "A5")
pm <- cbind(
  c(5, 10, 4, 8, 1),
  c(4, 8, 2, 16, 4),
  c(7, 7, 6, 8, 5)
)
rownames(pm) <- alt
colnames(pm) <- cri
minmax <- c("max", "max", "min")
result <- borda$new(pm, minmax)
}
\references{
Borda Count. Wikipedia. available from \url{https://en.wikipedia.org/wiki/Borda_count}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{Borta}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{points}}{points based on rank of the criteria in the alternatives}

\item{\code{result}}{matrix with points and assigned ranks for the alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-borda-new}{\code{borda$new()}}
\item \href{#method-borda-compute}{\code{borda$compute()}}
\item \href{#method-borda-summary}{\code{borda$summary()}}
\item \href{#method-borda-clone}{\code{borda$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-borda-new"></a>}}
\if{latex}{\out{\hypertarget{method-borda-new}{}}}
\subsection{Method \code{new()}}{
public constructor allowing the user to construct Borta count decision
 analysis problem and compute it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{borda$new(pm, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion. only
numeric values expercted. Rows and columns are expected to be named.}

\item{\code{minmax}}{value or vector of values 'min' or 'max' specifying
optimization direction for the criterium}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialized R6 class with computed results for Borda count
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cri <- c("C1", "C2", "C3")
alt <- c("A1", "A2", "A3", "A4", "A5")
pm <- cbind(
  c(5, 10, 4, 8, 1),
  c(4, 8, 2, 16, 4),
  c(7, 7, 6, 8, 5)
)
rownames(pm) <- alt
colnames(pm) <- cri
minmax <- c("max", "max", "min")
result <- borda$new(pm, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-borda-compute"></a>}}
\if{latex}{\out{\hypertarget{method-borda-compute}{}}}
\subsection{Method \code{compute()}}{
Computes Borda model based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{borda$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-borda-summary"></a>}}
\if{latex}{\out{\hypertarget{method-borda-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the Borda count method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{borda$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-borda-clone"></a>}}
\if{latex}{\out{\hypertarget{method-borda-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{borda$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
