% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm.R
\name{bwm}
\alias{bwm}
\title{Best worst method}
\usage{
bwm(a_best, a_worst)
}
\arguments{
\item{a_best}{vector of preferences of best criterium to others (preference
1-9)}

\item{a_worst}{vector of preferences of all other criteri to worst
criterium (preference 1-9)}
}
\value{
list:
\itemize{
   \item weights - derived weights
   \item xi - inconsistence indicator
   \item CR - consistency ratio
   \item status - status of LP problem solution
}
}
\description{
Best-worse method (BWM) uses similar approach to weight estimation as AHP,
 but simplifies requrements on pairwise comparisons. While in AHP
 combination of all criteria importance is required, BWM compares only best
 to others and worse criterium to other, all other weights can be derived by
 solving linear programing optimization problem.

Similarly to AHP Consistency ratio can be computed to identify how
 consistent these preferences are, although the interpretation is different.

\itemize{
  \item CR <= 0.1 - exceptional consistency (rare in real-world studies)
  \item CR in (0.1; 0.25) - good consistency, usually acceptable in studies
  \item CR > 0.3 - high level of preferences inconsistence, reevaluate
 preferences
}
}
\references{
REZAEI, Jafar. Best-worst multi-criteria decision-making method. Omega.
 2015, vol. 53, s. 49–57, available from:
 https://doi.org/10.1016/j.omega.2014.11.009, ISSN 0305-0483.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{BWM}
\keyword{weights}
