% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoso.R
\name{cocoso}
\alias{cocoso}
\title{Combined Compromise solution}
\description{
Combined Compromise solution

Combined Compromise solution
}
\details{
#@description
MCDA analysis method based on finding of compomise solution for \link{wsm}
 and \link{wpm} models. The compromise is found in following manner.

Analysis starts with formulation of performance matrix of alternatives
 (rows) in criteria (columns). Values of performance matrix are then
 normalized using min-max method.

In step 3 and 4 of analysis weighted sum (Si) and weighted products (Pi) are
 computed.

\mjsdeqn{S_i = \sum_{j=1}^n w_j r_{ij}}

\mjsdeqn{P_i = \sum_{j=1}^n r_{ij}^{w_j}}

Method proposes three measures for establishing ranking order called ki a)
 to c). Each measure can be used separately to establish ranking of the
 alternatives.

\mjsdeqn{k_{ia} = \frac{P_i + S_i}{\sum_{i=1}^m (P_i + S_i)}}

\mjsdeqn{k_{ib} = \frac{S_i}{min_i S_i} + \frac{P_i}{min_i P_i}}

\mjsdeqn{k_{ic} = \frac{\lambda S_i + (1 - \lambda) P_i}{\lambda max_i S_i + (1 - \lambda) max_i P_i}}

with

\mjsdeqn{0 \le \lambda \le 1}

Lambda in measure c serves as a weight for summary and product part of the
 measure. When lambda = 1, then these two componets will have same impact on
 result.

Lastly final ranking is established

\mjsdeqn{k_i = (k_{ia} k_{ib} k_{ic})^{1/3} + (k_{ia} k_{ib} k_{ic}) / 3}
}
\examples{

## ------------------------------------------------
## Method `cocoso$new`
## ------------------------------------------------

# from: https://www.youtube.com/watch?v=izbd0h0-fSs
alternatives <- c("AI 2024-T6", "AI 5052-O", "SS 301 FH", "SS 310-3AH",
 "Ti-6AI-4V", "Inconel 718", "70Cu-30Zn")
criteria <- c("Toughness index", "Yield Strength", "Young's Modulus",
 "Density", "Thermal expansion", "Thermal conductivity", "Specific Heat")
pm <- cbind(
 c(75.5, 95, 770, 187, 179, 239, 273),
 c(420, 91, 1365, 1120, 875, 1190, 200),
 c(74.2, 70, 189, 210, 112, 217, 112),
 c(2.8, 2.68, 7.9, 7.9, 4.43, 8.51, 8.53),
 c(21.4, 22.1, 16.9, 14.4, 9.4, 11.5, 19.9),
 c(0.37, 0.33, 0.04, 0.03, 0.016, 0.31, 0.29),
 c(0.16, 0.16, 0.08, 0.08, 0.09, 0.07, 0.06)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.28, 0.14, 0.05, 0.24, 0.19, 0.05, 0.05)
minmax <- c("max", "max", "max", "min", "min", "min", "min")
t <- cocoso$new(pm, w, minmax)
}
\references{
A Combined Compromise Solution. 2022. Available from
 \url{https://www.youtube.com/watch?v=izbd0h0-fSs}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{CoCoSo}
\keyword{WPM}
\keyword{WSM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (unnormalized) performance matrix of
alternatives (in rows) in criteria (columns)}

\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{weight vector}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria.}

\item{\code{lambda}}{a constant in interval <0;1> to establish influence of
weighted sum and weighted product on the result}

\item{\code{results}}{provides dataframe with ki a to c partial rankings and ki
representing final (compromise) solution. (largest value best)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cocoso-new}{\code{cocoso$new()}}
\item \href{#method-cocoso-compute}{\code{cocoso$compute()}}
\item \href{#method-cocoso-summary}{\code{cocoso$summary()}}
\item \href{#method-cocoso-clone}{\code{cocoso$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cocoso-new"></a>}}
\if{latex}{\out{\hypertarget{method-cocoso-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cocoso$new(pm, w, minmax = "max", lambda = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{w}}{weight vector (sum of weights = 1)}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}

\item{\code{lambda}}{a constant in interval <0;1> to establish influence of
weighted sum and weighted product on the result. Implicitly set to 0.5
to represent equalt influence of both components}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# from: https://www.youtube.com/watch?v=izbd0h0-fSs
alternatives <- c("AI 2024-T6", "AI 5052-O", "SS 301 FH", "SS 310-3AH",
 "Ti-6AI-4V", "Inconel 718", "70Cu-30Zn")
criteria <- c("Toughness index", "Yield Strength", "Young's Modulus",
 "Density", "Thermal expansion", "Thermal conductivity", "Specific Heat")
pm <- cbind(
 c(75.5, 95, 770, 187, 179, 239, 273),
 c(420, 91, 1365, 1120, 875, 1190, 200),
 c(74.2, 70, 189, 210, 112, 217, 112),
 c(2.8, 2.68, 7.9, 7.9, 4.43, 8.51, 8.53),
 c(21.4, 22.1, 16.9, 14.4, 9.4, 11.5, 19.9),
 c(0.37, 0.33, 0.04, 0.03, 0.016, 0.31, 0.29),
 c(0.16, 0.16, 0.08, 0.08, 0.09, 0.07, 0.06)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.28, 0.14, 0.05, 0.24, 0.19, 0.05, 0.05)
minmax <- c("max", "max", "max", "min", "min", "min", "min")
t <- cocoso$new(pm, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cocoso-compute"></a>}}
\if{latex}{\out{\hypertarget{method-cocoso-compute}{}}}
\subsection{Method \code{compute()}}{
computes CoCoSo model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cocoso$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cocoso-summary"></a>}}
\if{latex}{\out{\hypertarget{method-cocoso-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the CoCoSo method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cocoso$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cocoso-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cocoso-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cocoso$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
