% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codas.R
\name{codas}
\alias{codas}
\title{COmbinative Distance-based ASsessment}
\description{
Method uses Euclidean and taxicab distances to anti-ideal (negative-ideal)
 solution.

The performance of alternatives (rows) in criteria (columns) is normalized
 using "to best" normalization (see \link{mcda_norm}). Then the weights are
 applied.

\mjsdeqn{r_{ij} = w_j n_{ij}}

where w are weights of criteria, n are normalized values of performance
 matrix. We use this information to identify negative-ideal (anti-ideal)
 solution

\mjsdeqn{ns_j = min_i r_{ij}}

Next we compute Eucludean and taxicab distances of the alternatives to this
 negative-ideal solution.

\mjsdeqn{E_i = \sqrt{\sum_{j=1}^m (r_{ij} - ns_j)^2}}

\mjsdeqn{T_i = \sum_{j=1}^m | r_{ij} - ns_j |}

Then we compute relative assesment matrix

\mjsdeqn{h_{ik} = (E_i - E_k) + \psi \times (T_i - T_k)}

Where psi is 1 if difference of eucledian distances is greater, than set
 threshold: |Ei - Ek| >= tau, otherwise the psi = 0

Finaly we calculate assessment score, which is a measure that can be used
 directly to establish ranking of the alternatives. (Highest value 'best').

\mjsdeqn{H_i = \sum_{k=1}^n h_{ik}}
}
\examples{

## ------------------------------------------------
## Method `codas$new`
## ------------------------------------------------

# adjusted from reference article
pm <- rbind(
 c(60, 0.4, 2540, 500, 990),
 c(6.35, 0.15, 1016, 3000, 1041),
 c(6.8, 0.10, 1727.2, 1500, 1676),
 c(10, 0.2, 1000, 2000, 965),
 c(2.5, 0.10, 560, 500, 915),
 c(4.5, 0.08, 1016, 350, 508),
 c(3, 0.1, 1778, 1000, 920)
)
colnames(pm) <- c("Load cap.", "Max. tip speed", 
  "Repeatability", "Memory capacity", "Manipulator reach")
rownames(pm) <- c(paste0("A", 1:7))
w <- c(0.036, 0.192, 0.326, 0.326, 0.120)
minmax <- c("max", "min", "max", "max", "max")
t <- codas$new(pm, w, minmax = minmax, tau = 0.02)
summary(t)
}
\references{
GHORABAEE, M.K. et al. A NEW COMBINATIVE DISTANCE-BASED ASSESSMENT (CODAS)
METHOD FOR MULTI - CRITERIA DECISION - MAKING. Economic Computation and
 Economic Cybernetics Studies and Research, Vol. 50, no. 3, pp. 25-44.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{CODAS}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (unnormalized) performance matrix of
alternatives (in rows) in criteria (columns)}

\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{weight vector}

\item{\code{tau}}{coefficient of determination}

\item{\code{score}}{assesment score (best = highest)}

\item{\code{score_sorted}}{assesment score sorted from best to worst}

\item{\code{result}}{databrame with score and rank}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-codas-new}{\code{codas$new()}}
\item \href{#method-codas-compute}{\code{codas$compute()}}
\item \href{#method-codas-summary}{\code{codas$summary()}}
\item \href{#method-codas-clone}{\code{codas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-new"></a>}}
\if{latex}{\out{\hypertarget{method-codas-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$new(pm, w, minmax = "max", tau = 0.02)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{w}}{weight vector (sum of weights = 1)}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}

\item{\code{tau}}{parameter of threshold, set to 0.02 by default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# adjusted from reference article
pm <- rbind(
 c(60, 0.4, 2540, 500, 990),
 c(6.35, 0.15, 1016, 3000, 1041),
 c(6.8, 0.10, 1727.2, 1500, 1676),
 c(10, 0.2, 1000, 2000, 965),
 c(2.5, 0.10, 560, 500, 915),
 c(4.5, 0.08, 1016, 350, 508),
 c(3, 0.1, 1778, 1000, 920)
)
colnames(pm) <- c("Load cap.", "Max. tip speed", 
  "Repeatability", "Memory capacity", "Manipulator reach")
rownames(pm) <- c(paste0("A", 1:7))
w <- c(0.036, 0.192, 0.326, 0.326, 0.120)
minmax <- c("max", "min", "max", "max", "max")
t <- codas$new(pm, w, minmax = minmax, tau = 0.02)
summary(t)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-compute"></a>}}
\if{latex}{\out{\hypertarget{method-codas-compute}{}}}
\subsection{Method \code{compute()}}{
computes CODAS model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-summary"></a>}}
\if{latex}{\out{\hypertarget{method-codas-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the CODAS method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-clone"></a>}}
\if{latex}{\out{\hypertarget{method-codas-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
