% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codas.R
\name{codas}
\alias{codas}
\title{COmbinative Distance-based ASsessment}
\description{
Method uses Euclidean and taxicab distances to anti-ideal (negative-ideal)
 solution.

The performance of alternatives (rows) in criteria (columns) is normalized
 using "to best" normalization (see \link{mcda_norm}). Then the weights are
 applied.

\mjsdeqn{r_{ij} = w_j n_{ij}}

where w are weights of criteria, n are normalized values of performance
 matrix. We use this information to identify negative-ideal (anti-ideal)
 solution

\mjsdeqn{ns_j = min_i r_{ij}}

Next we compute Eucludean and taxicab distances of the alternatives to this
 negative-ideal solution.

\mjsdeqn{E_i = \sqrt{\sum_{j=1}^m (r_{ij} - ns_j)^2}}

\mjsdeqn{T_i = \sum_{j=1}^m | r_{ij} - ns_j |}

Then we compute relative assesment matrix

\mjsdeqn{h_{ik} = (E_i - E_k) + \psi(E_i - E_k) \times (T_i - T_k)}

Finaly we calculate assessment score, which is a measure that can be used
 directly to establish ranking of the alternatives. (Highest value 'best').

\mjsdeqn{H_i = \sum_{k=1}^n h_{ik}}
}
\examples{

## ------------------------------------------------
## Method `codas$new`
## ------------------------------------------------

# from: https://www.youtube.com/watch?v=VetfNIHLyn0
alternatives <- c("AI 2024-T6", "AI 5052-O", "SS 301 FH", "SS 310-3AH",
 "Ti-6AI-4V", "Inconel 718", "70Cu-30Zn")
criteria <- c("Toughness index", "Yield Strength", "Young's Modulus",
 "Density", "Thermal expansion", "Thermal conductivity", "Specific Heat")
pm <- cbind(
 c(75.5, 95, 770, 187, 179, 239, 273),
 c(420, 91, 1365, 1120, 875, 1190, 200),
 c(74.2, 70, 189, 210, 112, 217, 112),
 c(2.8, 2.68, 7.9, 7.9, 4.43, 8.51, 8.53),
 c(21.4, 22.1, 16.9, 14.4, 9.4, 11.5, 19.9),
 c(0.37, 0.33, 0.04, 0.03, 0.016, 0.31, 0.29),
 c(0.16, 0.16, 0.08, 0.08, 0.09, 0.07, 0.06)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.28, 0.14, 0.05, 0.24, 0.19, 0.05, 0.05)
minmax <- c("max", "max", "max", "min", "min", "min", "min")
t <- codas$new(pm, w, minmax)
}
\references{
Vitarka. CODAS - Combinative Distance-based Assessment #MADM #Optimization
 #MaterialSelection #MCDM. Available from
\url{https://www.youtube.com/watch?v=d0W8wBxPkhY}

Manshadi B.D., Mahmudi, H., Abedian, A., Manmudi, R. A novel method for
 materials selection in mechanical design: combination of non-linear
 linearization and a modified digital logic method. Mater Des 2007, 28:
 8-15
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{CODAS}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (unnormalized) performance matrix of
alternatives (in rows) in criteria (columns)}

\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{weight vector}

\item{\code{psi}}{coefficient of determination}

\item{\code{score}}{assesment score (best = highest)}

\item{\code{score_sorted}}{assesment score sorted from best to worst}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-codas-new}{\code{codas$new()}}
\item \href{#method-codas-compute}{\code{codas$compute()}}
\item \href{#method-codas-summary}{\code{codas$summary()}}
\item \href{#method-codas-clone}{\code{codas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-new"></a>}}
\if{latex}{\out{\hypertarget{method-codas-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$new(pm, w, minmax = "max", psi = 0.02)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{w}}{weight vector (sum of weights = 1)}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}

\item{\code{psi}}{coefficient of determination, set to 0.02 by default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# from: https://www.youtube.com/watch?v=VetfNIHLyn0
alternatives <- c("AI 2024-T6", "AI 5052-O", "SS 301 FH", "SS 310-3AH",
 "Ti-6AI-4V", "Inconel 718", "70Cu-30Zn")
criteria <- c("Toughness index", "Yield Strength", "Young's Modulus",
 "Density", "Thermal expansion", "Thermal conductivity", "Specific Heat")
pm <- cbind(
 c(75.5, 95, 770, 187, 179, 239, 273),
 c(420, 91, 1365, 1120, 875, 1190, 200),
 c(74.2, 70, 189, 210, 112, 217, 112),
 c(2.8, 2.68, 7.9, 7.9, 4.43, 8.51, 8.53),
 c(21.4, 22.1, 16.9, 14.4, 9.4, 11.5, 19.9),
 c(0.37, 0.33, 0.04, 0.03, 0.016, 0.31, 0.29),
 c(0.16, 0.16, 0.08, 0.08, 0.09, 0.07, 0.06)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
w <- c(0.28, 0.14, 0.05, 0.24, 0.19, 0.05, 0.05)
minmax <- c("max", "max", "max", "min", "min", "min", "min")
t <- codas$new(pm, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-compute"></a>}}
\if{latex}{\out{\hypertarget{method-codas-compute}{}}}
\subsection{Method \code{compute()}}{
computes CODAS model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-summary"></a>}}
\if{latex}{\out{\hypertarget{method-codas-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the CODAS method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codas-clone"></a>}}
\if{latex}{\out{\hypertarget{method-codas-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{codas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
