% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cradis.R
\name{cradis}
\alias{cradis}
\title{compromise ranking of alternatives from distance to ideal solution}
\description{
Approach to ranking problem integrating parts of \link{aras}, \link{marcos}
 and \link{topsis} methods.

Similarly to ARAS method it starts with decision problem formulation using
 performance metrix with n criteria in columns and m alternatives in rows
 (step 1). In step 2) The performance in criteria is then normalized using:

\mjsdeqn{r_{ij} = \frac{x_{ij}}{x_{max_j}}}

for benefit criteria and for cost criteria

\mjsdeqn{r_{ij} = \frac{x_{min_j}}{x_{ij}}}

In step 3) the weights are applied on normalized matrix.

\mjsdeqn{v_{ij} = w_{j} r_{ij}}

In step 4) determining ideal and antiideal solution of the problem as
 miximal (ideal) and minimal (anti-ideal) values of criteria.

\mjsdeqn{t_{i} = max(v_{ij})}

\mjsdeqn{t_{ai} = min(v_{ij})}

In step 5) deviations from these is computed as difference between the
 normalized weighted value and the ideal or antiideal solution:

\mjsdeqn{d^+ = t_i - v_{ij}}

\mjsdeqn{d^- = v_{ij} - t_ai}

Then the grades s are computed by summing up these differences (step 6).

\mjsdeqn{s_i^+ = \sum_{j=1}^n d^+}

\mjsdeqn{s_i^- = \sum_{j=1}^n d^-}

Then the utility function is constructed as a ratio between the optimum and
 actual value (step 7).

\mjsdeqn{K_i^+ = \frac{s_0^+}{s_i^+}}

\mjsdeqn{K_i^- = \frac{s_i^-}{s_0^-}}

Finaly (step 8) the ranking of the alternatives ise derived from the average
 of utility functions.

\mjsdeqn{Q_i = \frac{K_i^+ + K_i^-}{2}}

First 4 steps of the method are closely mimicking ARAS approach, remainder
 is philosophically closer to MARCOS or TOPSIS methods. Mathematically the
 approach to the problem is closer to MARCOS.
}
\examples{

## ------------------------------------------------
## Method `cradis$new`
## ------------------------------------------------

# https://doi.org/10.18485/aeletters.2023.8.1.4
w <- c(0.257, 0.129, 0.214, 0.196, 0.204)
pm <- rbind(
  c(2, 110, 3, 2, 3),
  c(5, 100, 5, 3, 3),
  c(3, 90, 4, 5, 2),
  c(10, 80, 3, 4, 4),
  c(4, 85, 2, 4, 5),
  c(8, 80, 3, 4, 4),
  c(5, 95, 2, 4, 3)
)
colnames(pm) <- c(paste0("C", 1:5))
rownames(pm) <- c(paste0("S", 1:7))
t <- cradis$new(pm = pm, w = w, minmax = "max")
summary(t)
}
\references{
Puška, A., Stević, Ž., Panučar, D.: Evaluation and selection of healtcare
 waste incinerators using extended sustainability criteria and multicriteria
 analysis methods. Environment, Development and Sustainability, vol. 24, pp.
 11195-11225 (2022), DOI: 10.1007/s10668-021-01902-2

Ha, Le Dang. SELECTION OF SUITABLE DATA NORMALIZATION METHOD TO COMBINE WITH
 THE CRADIS METHOD FOR MAKING MULTI - CRITERIA DECISION. Applied Engineering
 Letters, Vol. 8, No. 1, pp. 24-35 (2023), DOI: 10.18485/aeletters.2023.8.1.4
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ARAS}
\keyword{CRADIS}
\keyword{MARCOS}
\keyword{TOPSIS}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original state of the performance matrix (before
transforming criteria)}

\item{\code{pm}}{transformed performance matrix, all criteria are maximized}

\item{\code{w}}{weight vector}

\item{\code{minmax}}{directio of optimization, can be replaced by single min
or max if all criteria have same optimization direction.}

\item{\code{q}}{sorted average utility order descending from best to worst}

\item{\code{result}}{data frame with average utility and ranking}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cradis-new}{\code{cradis$new()}}
\item \href{#method-cradis-compute}{\code{cradis$compute()}}
\item \href{#method-cradis-summary}{\code{cradis$summary()}}
\item \href{#method-cradis-clone}{\code{cradis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cradis-new"></a>}}
\if{latex}{\out{\hypertarget{method-cradis-new}{}}}
\subsection{Method \code{new()}}{
public constructor of the function. Validates inputs and computes the
 model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cradis$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows and criteria in
columns)}

\item{\code{w}}{weights vector}

\item{\code{minmax}}{vector of optimization direction (min/max, max is default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
inicialized and computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# https://doi.org/10.18485/aeletters.2023.8.1.4
w <- c(0.257, 0.129, 0.214, 0.196, 0.204)
pm <- rbind(
  c(2, 110, 3, 2, 3),
  c(5, 100, 5, 3, 3),
  c(3, 90, 4, 5, 2),
  c(10, 80, 3, 4, 4),
  c(4, 85, 2, 4, 5),
  c(8, 80, 3, 4, 4),
  c(5, 95, 2, 4, 3)
)
colnames(pm) <- c(paste0("C", 1:5))
rownames(pm) <- c(paste0("S", 1:7))
t <- cradis$new(pm = pm, w = w, minmax = "max")
summary(t)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cradis-compute"></a>}}
\if{latex}{\out{\hypertarget{method-cradis-compute}{}}}
\subsection{Method \code{compute()}}{
computes the model bas of class properties. Usually we do not run the
 computation manually (it is run from class' constructor).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cradis$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cradis-summary"></a>}}
\if{latex}{\out{\hypertarget{method-cradis-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the CRADIS method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cradis$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cradis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cradis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cradis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
