% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dematel.R
\name{dematel}
\alias{dematel}
\title{Decision Making Trial and Evaluation Laboratory (DEMATEL) method for weight
 establishment}
\usage{
dematel(relations)
}
\arguments{
\item{relations}{list of matrixes with stated influences between the
criteria. Each matrix represents oppinion of single expert. Number of
criteria in the matrix must be same in all of these oppinions and in
same structure (order of criteria).}
}
\value{
returns list in following structure
\itemize{
  \item D - relevance - influence from
  \item R - relevance - influence to
  \item prominence
  \item vector of weights for the criteria
}
}
\description{
DEMATEL is a method to derive weight from the information on relation
 (influence) between the criteria specified by the experts on problem
 domain. It can be seen as an alternative to pair-wise comparison methods
 which derive the weights from stated preferences of one criterium when
 comparing with the ether criterium.

Or approaches such as \link{EWM} which derive weights based on entropy of
 performances of alternatives in the criteria.

The procedure of DEMATEL is:

step 1) [average] direct relationship matrix generation

We start with formulating square matrix n:n \mjseqn{A_k} to state the
 influence between the criteria. Dimentsion n nepresents number of criteria,
 A is the matrix of relationships between the criteria and k is number of
 experts we source the information from.

Information from experts is usually gathered using questionaries/oppinions.

In this matrix \mjseqn{a_{ij}(k)} is the influence degree of criterium
 \mjseqn{E_i} on criterium \mjseqn{E_j} determined by
 \mjseqn{K^{th}} expert.

Influence is established on following scale:
\itemize{
 \item 0 - no influence
 \item 1 - low influence
 \item 2 - middle influence
 \item 3 - high influence
 \item 4 - very high influence
}

\deqn{
  A_k = \begin{bmatrix} 0 & ... & a_{1n}(k) \\
  \vdots & \ddots & \vdots \\
  a_{n1}(k) & ... & 0\\
\end{bmatrix}
}

If we have more experts, we will also have more than one oppinions and we
 will need to integrate these into single matrix Z.

\mjsdeqn{Z = \frac{\sum_{k=1}^m A_k}{m}}

step 2) normalizing of the direct realtion matrix

Next the average direct relation matrix needs to be normalized using:

\mjsdeqn{S = max(\sum_{j=1}^n z_{ij}, \sum_{i=1}^n z_{ij})}

\mjsdeqn{X = \frac{Z}{S}}

step 3) creating of Total Relation Matrix

\mjsdeqn{T = \sum_{q=1}^\infty X^i = \frac{X}{1-X}}

step 4) assessing the relevance and prominence

We compute sumation of rows and columns:

\mjsdeqn{D = [D_i]_{n x 1} = [\sum_{j=1}^n t_{ij}]_{n x 1}}

\mjsdeqn{R = [R_j]_{1 x n} = [\sum_{i=1}^n t_{ij}]_{1 x n}}

Where D represents total effect both direct and indirect influences given by
 criterium i to other criteria j. Similarly $R_j$ represents total effort
 (direct and indirect) received by criterium j from other criterias.

\mjsdeqn{\sum D + R}

Is called "prominence" and proves degree of importance role of criterion i
 in system and also gives an index that shows total effect both given and
 received by criterion.

step 5) weights derivation

From values of influence propagated from (D) and to (R) we derive weights

\mjsdeqn{w_{ia} = \sqrt{(D_i + R_j)^2 + (D_i - R_j)^2}}

\mjsdeqn{w_i = \frac{w_{ia}}{\sum_{i=1}^n w_{ia}}}
}
\examples{
op1 <- cbind(
  c(0, 3, 3, 3, 4),
  c(2, 0, 3, 0, 0),
  c(4, 2, 0, 4, 2),
  c(3, 0, 0, 0, 0),
  c(4, 0, 0, 2, 0)
)
rownames(op1) <- colnames(op1) <- c("C1", "C2", "C3", "C4", "C5")
opinion <- list(op1)
t <- dematel(opinion)

}
\references{
Taherdoost, H., Madanchian, M., 2023. Understanding Applications and Best
 Practices of DEMATEL: A Method for Prioritizing Key Factors in
 Multi-Criteria Decision-Making. Journal of Management Science &
 Engineering Research. 6(2): 17-23. DOI:
 \url{https://doi.org/10.30564/jmser.v6i2.5634}
}
\keyword{DEMATEL}
