% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas.R
\name{edas}
\alias{edas}
\title{Evaluation Based on Distance from Average Solution}
\description{
Another method loosely based on \link{wsm}. This method computes positive
 and negative distance of the performance to the average performance value.
 Then computes weighted sums of these averages and then computes average of
 these two metrics. Mathematically it looks like this:

step 1) compute averages of the performance in the criteria (AVj)

step 2) calculate positive distance from the average. For benefit criteria

\mjsdeqn{PDA_{ij} = max(0; \frac{x_{ij}-AV_j}{AV_j})}

and for cost criteria

\mjsdeqn{PDA_{ij} = max(0; \frac{AV_j - x_{ij}}{AV_j})}

Similarly we compute negative distance to the average performance. Formally
 we switch the statements, so for benefit criteria (step 3)

\mjsdeqn{NDA_{ij} = max(0; \frac{AV_j - x_{ij}}{AV_j})}

and for cost criteria

\mjsdeqn{NDA_{ij} = max(0; \frac{x_{ij}-AV_j}{AV_j})}

These two measures can be used to produce weighted positive (SPI) and
 negative (SNI) sums (steps 4 and 5)

\mjsdeqn{SPI = \sum_{j=1}^m PDA_{ij} \cdot w_j}

\mjsdeqn{SNI = \sum_{j=1}^m NDA_{ij} \cdot w_j}

Computed values of the SPI and SNI are normalized against its maximal values
 (step 6)

\mjsdeqn{NSPI = \frac{SPI}{max(SPI)}}

\mjsdeqn{NSNI = 1 - \frac{SNI}{max(SNI)}}

step 7) compute average of NSPI and NSNI

\mjsdeqn{ASI = \frac{NSPI + NSNI}{2}}

Value of ASI is directly usable for ranking of alternatives (largest value
 is best).
}
\examples{

## ------------------------------------------------
## Method `edas$new`
## ------------------------------------------------

#from https://www.youtube.com/watch?v=Kwq_ra6b6eU
alternatives <- c("car 1", "car 2", "car 3", "car 4", "car 5", "car 6",
 "car 7", "car 8", "car 9", "car 10")
criteria <- c("quality", "condition", "security", "delivery days",
 "fuel consumption", "price")
M <- rbind(
  c(3, 6, 4, 20, 2, 30000),
  c(4, 4, 6, 15, 2.2, 32000),
  c(6, 5, 9, 18, 3, 32100),
  c(5, 6, 3, 23, 2.8, 28000),
  c(4, 8, 7, 30, 1.5, 29000),
  c(8, 3, 6, 35, 1.9, 27000),
  c(7, 2, 5, 33, 1.7, 28500),
  c(3, 8, 3, 34, 1.6, 30500),
  c(8, 4, 8, 40, 2.5, 33000),
  c(9, 3, 7, 34, 2, 29800)
)
rownames(M) <- alternatives
colnames(M) <- criteria
minmax <- c("max", "max", "max", "min", "min", "min")
w <- c(0.1, 0.2, 0.1, 0.2, 0.1, 0.3)
t <- edas$new(M, w, minmax)
}
\references{
Use of Evaluation Based on the Distance from Average Solution 'EDAS" in
 decision making process. \url{https://www.youtube.com/watch?v=Kwq_ra6b6eU}.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{EDAS}
\keyword{WSM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{minmax}}{vector with optimization direction of the criteria (min or
max), can be substituted by single min or max if optimization direction
for all criteria is the same.}

\item{\code{w}}{weight vector for criteria}

\item{\code{spi_scoreM}}{graphical representation of criteria contribution to
weighted sum of positive distances}

\item{\code{sni_scoreM}}{graphical representation of criteria contribution to
weighted sum of negative distances}

\item{\code{result}}{Table with computed weighted positive (SPI),
negative (SNI) sums, normalized positive (NSPI), negative (NSNI) sums
and average (ASI) of NSNI and NSPI}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-edas-new}{\code{edas$new()}}
\item \href{#method-edas-compute}{\code{edas$compute()}}
\item \href{#method-edas-summary}{\code{edas$summary()}}
\item \href{#method-edas-clone}{\code{edas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edas-new"></a>}}
\if{latex}{\out{\hypertarget{method-edas-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{edas$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns).}

\item{\code{w}}{weight vector for criteria}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max).
Can use single min/max if optimization direction of all criteria is
same.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#from https://www.youtube.com/watch?v=Kwq_ra6b6eU
alternatives <- c("car 1", "car 2", "car 3", "car 4", "car 5", "car 6",
 "car 7", "car 8", "car 9", "car 10")
criteria <- c("quality", "condition", "security", "delivery days",
 "fuel consumption", "price")
M <- rbind(
  c(3, 6, 4, 20, 2, 30000),
  c(4, 4, 6, 15, 2.2, 32000),
  c(6, 5, 9, 18, 3, 32100),
  c(5, 6, 3, 23, 2.8, 28000),
  c(4, 8, 7, 30, 1.5, 29000),
  c(8, 3, 6, 35, 1.9, 27000),
  c(7, 2, 5, 33, 1.7, 28500),
  c(3, 8, 3, 34, 1.6, 30500),
  c(8, 4, 8, 40, 2.5, 33000),
  c(9, 3, 7, 34, 2, 29800)
)
rownames(M) <- alternatives
colnames(M) <- criteria
minmax <- c("max", "max", "max", "min", "min", "min")
w <- c(0.1, 0.2, 0.1, 0.2, 0.1, 0.3)
t <- edas$new(M, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edas-compute"></a>}}
\if{latex}{\out{\hypertarget{method-edas-compute}{}}}
\subsection{Method \code{compute()}}{
Computes the EDAS using params provided in constructor. Usually run
 automatically by constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{edas$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edas-summary"></a>}}
\if{latex}{\out{\hypertarget{method-edas-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the EDAS method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{edas$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edas-clone"></a>}}
\if{latex}{\out{\hypertarget{method-edas-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{edas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
