% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELECTRE1_Kernel.R
\name{ELECTRE1_Kernel}
\alias{ELECTRE1_Kernel}
\title{computes so called kernell of the decision for ELECTRE I, II and 1S methods}
\usage{
ELECTRE1_Kernel(am)
}
\arguments{
\item{am}{adjacancy matrix or credibility matrix (ELECTRE 1S)}
}
\value{
list of:
\itemize{
  \item graph - graphical representation of domination of one alternative
 over another
  \item dominated - vector of alternatives identified as dominated
  \item kernel - oposite to dominated vector. Consist for alternatives not
 dominated by other alternatives, forming kernel of the solution.
}
}
\description{
Computes kernell of the solution as the set of alternatives which are not
 dominated by any other alternative. Other alternatives then for example can
 be excluded from decision making as they are clearly suboptimal.

The results are presented in graphical form as network diagram with flows
 representing domination relation and two vectors with alternatives in
 kernel and dominated alternatives.

The function is beiung utilized for \link{electre1}, \link{electre2} and
 \link{electre1s} methods.
}
\references{
Balamurali, M.: pyDecisions - A Python Library of management decision making
 techniques. Avilable on-line from
 \url{https://github.com/Valdecy/pyDecisions}

Rogers, Martin and Myastre, Lucien-Yves. ELECTRE and Decision Support:
 Methods and Applications in Engineering and Infrastructure investment.
 Springer 2000, 208 p., ISBN 978-1-4757-5057-7

Prombo, M. Package OutrankingTools, CRAN: 2015, available from:
 \url{https://cran.r-project.org/web/packages/OutrankingTools/}
}
\seealso{
\link{electre1}

\link{electre2}

\link{electre1s}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{1S}
\keyword{ELECTRE}
\keyword{I}
\keyword{II}
\keyword{kernel}
\keyword{methods}
