% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electre4.R
\name{electre4}
\alias{electre4}
\title{ELECTRE IV method for ranking alternatives}
\description{
ELECTRE IV is alternative ranking approach based on ELECTRE III method (see
 \code{\link{electre3}}), but without using any weighting of the criteria.
 Since weighting is required for concordance and discordance matrixes, these
 are not available in ELECTRE IV. Instead method uses more complex system
 of outranking relations:

\itemize{
  \item mp(b,a) - number of criteria for which option b is strictly
 preferred to a
  \item mq(b,a) - number of criteria for which option b is weakly preferred
 to a
  \item mj(b,a) - number of criteria for which option b is judged
 indifferent to a
  \item mo(b,a) = mo(a,b) - number of criteria on which options a and b
 perform identically
}

These are used as a base for computation of credibility matrix.Reamining
 computations are same as for ELECTRE III methods including construction of
 descending and ascending pre-order and deriving final outranking relation.

Please refer further to \code{\link{electre3}} for more in detailed
 information on computational aparatus.
}
\examples{

## ------------------------------------------------
## Method `electre4$new`
## ------------------------------------------------

PM <- cbind(
  c(-14,129,-10,44,-14),
  c(90,100,50,90,100),
  c(40,0,10,5,20),
  c(40,0,10,5,20),
  c(100,0,100,20,40)
)
rownames(PM) <- c("Project1","Project2","Project3","Project4","Project5")
colnames(PM) <- c( "CR1","CR2","CR3","CR4","CR5")
minmaxcriteria <- "max"
Q <- c(25,16,0,12,10) #Indifference thresholds
P <- c(50,24,1,24,20) #Preference thresholds
V <- c(100,60,2,48,90) #Veto thresholds
t <- electre4:new(PM, P, Q, V, minmaxcriteria)
}
\references{
Balamurali, M.: pyDecisions - A Python Library of management decision making
 techniques. Avilable on-line from
 \url{https://github.com/Valdecy/pyDecisions}

Rogers, Martin and Myastre, Lucien-Yves. ELECTRE and Decision Support:
 Methods and Applications in Engineering and Infrastructure investment.
 Springer 2000, 208 p., ISBN 978-1-4757-5057-7

Roy B. : "The outranking approach and the foundations of ELECTRE methods",
 Theory and Decision 31, 1991, 49-73.

Vallée, D.; Zielniewicz, P. 1994. ELECTRE III-IV, version 3.x, Aspects
 Méthodologiques (tome 1), Guide d’utilisation (tome 2). Document du LAMSADE
 85 et 85bis, Université Paris Dauphine

Meyer, P. at al. MCDA package. GitHub: 2021, available from:
 \url{https://github.com/paterijk/MCDA/blob/master/}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{Discrimination}
\keyword{ELECTRE}
\keyword{III}
\keyword{IV}
\keyword{Outranking}
\keyword{analysis}
\keyword{approaches}
\keyword{credibility}
\keyword{matrix}
\keyword{modelling}
\keyword{multicriteria}
\keyword{preference}
\keyword{thresholds}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{performance matrix as inputed into function}

\item{\code{pm}}{adjusted performance matrix will all criteria transformed
in max direction}

\item{\code{p}}{preference threshold}

\item{\code{q}}{indeifference threshold}

\item{\code{v}}{veto threshold}

\item{\code{CredibilityMatrix}}{matrix assessing the strength of the assertion
that “a is at least as good as b"}

\item{\code{rank_d}}{descending distillation ranking - final partial preorder,
orders the alternatives from the best to the worst}

\item{\code{minmaxcriteria}}{vector of criteria optimization direction, either
min or max. Can be replaced with single min or max if all criteria are
optimized in the same direction.}

\item{\code{rank_a}}{ascending distillation ranking - final partial preorder
orders the alternatives from worst to best}

\item{\code{rank_p}}{re-order matrix specifying identified relations between
the alternatives - values are P+ (a prefered to b), P-
(b prefered to a), I (indifferent), R (incomparable)}

\item{\code{adjancancyMatrix}}{Adjancency Matrix of outranking relation between
the alternatives allows to visualize results as network diagram}

\item{\code{graph}}{processed adajncency matrix into network digram}

\item{\code{final_ranking}}{sorted final order of the alternatives}

\item{\code{finalRankingUnsorted}}{amed vector of unsorted order of the
alternatives, added for purposes of sensitivity analysis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-electre4-new}{\code{electre4$new()}}
\item \href{#method-electre4-compute}{\code{electre4$compute()}}
\item \href{#method-electre4-summary}{\code{electre4$summary()}}
\item \href{#method-electre4-sensitivity}{\code{electre4$sensitivity()}}
\item \href{#method-electre4-clone}{\code{electre4$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre4-new"></a>}}
\if{latex}{\out{\hypertarget{method-electre4-new}{}}}
\subsection{Method \code{new()}}{
public constructor, creates electre2 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre4$new(pm, p, q, v, minmaxcriteria = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{p}}{preference threshold - Vector containing the preference
thresholds. Constraints are defined separately for each criterion.}

\item{\code{q}}{vector of indifference thresholds}

\item{\code{v}}{vector of veto thresholds}

\item{\code{minmaxcriteria}}{vector of criteria direction to be used during
performance matrix normalization. Can be replaced with single max or
min if all criteria are to be maximized or minimized. Implicitly set
to max.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{PM <- cbind(
  c(-14,129,-10,44,-14),
  c(90,100,50,90,100),
  c(40,0,10,5,20),
  c(40,0,10,5,20),
  c(100,0,100,20,40)
)
rownames(PM) <- c("Project1","Project2","Project3","Project4","Project5")
colnames(PM) <- c( "CR1","CR2","CR3","CR4","CR5")
minmaxcriteria <- "max"
Q <- c(25,16,0,12,10) #Indifference thresholds
P <- c(50,24,1,24,20) #Preference thresholds
V <- c(100,60,2,48,90) #Veto thresholds
t <- electre4:new(PM, P, Q, V, minmaxcriteria)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre4-compute"></a>}}
\if{latex}{\out{\hypertarget{method-electre4-compute}{}}}
\subsection{Method \code{compute()}}{
computes the ELECTRE IV problem. Usually doesn't need to be run manually
 as it is called automatically by class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre4$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre4-summary"></a>}}
\if{latex}{\out{\hypertarget{method-electre4-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the ELECTRE IV method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre4$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre4-sensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-electre4-sensitivity}{}}}
\subsection{Method \code{sensitivity()}}{
sensitivity analysis for ELECTRE IV object is performed on values of
 preference (p), indefference (q) and veto (v) thresholds.

Since the thresholds (all of them) are specfied separately for each
 criterion, their sensitivity needs to be also evaluated seprately.
 This is being realized in this function by generating separate
 dataframe for each threshold (sens_p, sens_q, sens_v) with following
 structure:

criterium; from; default; to

From and to limits are limist of of solution stability, meaning that
 going under it (from) or over it (to) will produce different result.

Sensitivity is tested from default value (the value used to produce
 result) going down to 0 for from column and up to maximal performance
 in the criterium j (max(pm[j])). If no change in result is detected,
 value "insens." is inserted into dataframe.

Note that the computational part is very similar to Electre III method
 and thus is realized using \code{\link{sensitivity_e34}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre4$sensitivity(steps = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{steps}}{how many steps should sensitivity testing take. Interval
for testing will be split to steps segments. The higher number of steps
the smaller the step will be, the more granular the testing will be.
Set to 100 by default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns dataframes sens_p, sens_q and sens_v with sensitivity limit for
 the criteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-electre4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-electre4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{electre4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
