% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electre_3_paramCheck.R
\name{Electre_3_paramCheck}
\alias{Electre_3_paramCheck}
\title{function for checking of the parameters for the Electre 3 method}
\usage{
Electre_3_paramCheck(
  pm,
  w,
  p,
  q,
  v,
  minmaxcriteria = "max",
  alpha = 0.3,
  beta = 0.15
)
}
\arguments{
\item{pm}{performance of the alternatives in criteria. Alternatives are in
rows, criteria in columns.}

\item{w}{weights of the criteria in decision problem}

\item{p}{vector of preference thresholds (each threshold is set separately
for each criterium)}

\item{q}{vector of indiference thresholds}

\item{v}{vector of veto thresholds}

\item{alpha}{first discrimination threshold (0.3 by default)}

\item{beta}{second discrimination threshold (0.15 by default)}

\item{minmaxcriterie}{vector of optimization direction (min/max) for the
criteria}
}
\description{
Checks parameters for ELECTRE 3 method used by \code{\link{electre3}} R6
 class. It should stop processing of the inputs early when problem with
 parameters is detected by producing error message into the console.

When no problem with the parameters is detected, it just ends without
 returning any value, allowing the calling function using it to continue.

Note that as the Electre 3 method has lot of checks same as Electre 4 method
 that function uses also \code{\link{Electre_4_paramCheck}} for these common
 validity checks.
}
\examples{
PM <- cbind(
  c(-14,129,-10,44,-14),
  c(90,100,50,90,100),
  c(40,0,10,5,20),
  c(40,0,10,5,20),
  c(100,0,100,20,40)
)
rownames(PM) <- c("Project1","Project2","Project3","Project4","Project5")
colnames(PM) <- c( "CR1","CR2","CR3","CR4","CR5")
minmaxcriteria <- "max"
Q <- c(25,16,0,12,10) #Indifference thresholds
P <- c(50,24,1,24,20) #Preference thresholds
V <- c(100,60,2,48,90) #Veto thresholds
w <- c(1,1,1,1,1) #weights
Electre_3_paramCheck(PM, w, P, Q, V, minmaxcriteria)
}
\seealso{
\code{\link{electre3}}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{III}
\keyword{IV}
\keyword{parameter}
\keyword{validity}
