% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electre_asc_dist.R
\name{Electre_asc_dist}
\alias{Electre_asc_dist}
\title{algorithm for ascending distilation}
\usage{
Electre_asc_dist(sm)
}
\arguments{
\item{sm}{confidence matrix}
}
\value{
list of alternatives in ranks from worst-to best.
}
\description{
Algorithm to establish partial preorder by the means of ascending
 distilation. Preorder is achieved by distilling alternatives using
 progresively lower cut-off thresholds.

This aproach is complementary to descending distilation process which
 creates second pre-order.

The algorithm is used in Electre III (\code{\link{electre3}}) and Electre IV
 (\code{\link{electre4}}) methods.
}
\references{
Meyer, P. at al. MCDA package. GitHub: 2021, available from:
 \url{https://github.com/paterijk/MCDA/blob/master/}
}
\seealso{
\code{\link{Electre_desc_dist}}

\code{\link{electre3}}

\code{\link{electre4}}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{III}
\keyword{IV}
