% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electre_desc_dist.R
\name{Electre_desc_dist}
\alias{Electre_desc_dist}
\title{algrithm for descending distilation}
\usage{
Electre_desc_dist(sm)
}
\arguments{
\item{sm}{confidence matrix}
}
\description{
Algorithm to establish partial preorder by the means of descending
 distilation. Preorder is achieved by distilling alternatives using
 progresively lower cutoff thresholds.

This aproach is complementary to ascending distilation process
 \code{\link{Electre_asc_dist}} which creates second preorder.

The algorithm is used in \code{\link{electre3}} and
\code{\link{electre4}} methods.
}
\references{
Meyer, P. at al. MCDA package. GitHub: 2021, available from:
 \url{https://github.com/paterijk/MCDA/blob/master/}
}
\seealso{
\code{\link{Electre_asc_dist}}

\code{\link{electre3}}

\code{\link{electre4}}
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{Electre}
\keyword{III}
\keyword{IV}
\keyword{descending}
\keyword{distillation}
\keyword{methods}
\keyword{pre-order}
