% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ervd.R
\name{ervd}
\alias{ervd}
\title{Election Based on Relatived Value Distance}
\description{
A evolution of TOPSIS method to provide more stable results. The method
 introduces concept of reference points for the criteria the values below
 this point are considered losses, values over it gains.

The procedure starts with vertor normalization of the performance matrix
 (max). To the same scale the reference points are transformed:

\mjsdeqn{\varphi_j = \frac{\mu_{ij}}{\sum_{j=1}^m d_{ij}}}

Then the value af the alternative to criteria. For benefit function:

\mjsdeqn{v_{ij} = (r_{ij} - \varphi_j)^\alpha}

for \mjseqn{r_{ij} > \varphi_j} otherwise

\mjsdeqn{v_{ij} = -\lambda(\varphi_j - r_{ij})^\alpha}

for cost criteria

\mjsdeqn{v_{ij} = (\varphi_j - r_{ij})^\alpha}

for \mjseqn{r_{ij} < \varphi_j}, otherwise

\mjsdeqn{v_{ij} = -\lambda(r_{ij} - \varphi_j)^\alpha}

This concept was introduced by Tvertsky and Kahneman, with
 \mjseqn{\lambda = 2.25} and \mjseqn{\alpha = 0.88} from the empirical
 studies. Generally \mjseqn{\lambda \in <2; 2.5>}.

Then ideal (A+ ... PIS - positive ideal solution) and antiideal (A- ... NIS
 - negative ideal solutions) are identified as max or min in the vij and
 separation measures for PIS and NIS are compited:

\mjsdeqn{S_i^+ = \sum_{j=1}^n w_j \cdot |v_{ij} - v_j^+|}

\mjsdeqn{S_i^- = \sum_{j=1}^n w_j \cdot |v_{ij} - v_j^-|}

Finaly relative closeness index for each cleseness of each alternative to
 ideal solution. Last two steps are same as in TOPSIS:

\mjsdeqn{\phi_i = \frac{S_i^-}{S_i^+ + S_i^-}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{weights vector}

\item{\code{ref_points}}{reference points for the criteria}

\item{\code{minmax}}{vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{lambda}}{alternation factor of the losses, 2.25 by default, should be in <2;2.5>}

\item{\code{alpha}}{diminishing sensitivity parameter, 0.88 by default}

\item{\code{result}}{data frame with results}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ervd-new}{\code{ervd$new()}}
\item \href{#method-ervd-compute}{\code{ervd$compute()}}
\item \href{#method-ervd-summary}{\code{ervd$summary()}}
\item \href{#method-ervd-clone}{\code{ervd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ervd-new"></a>}}
\if{latex}{\out{\hypertarget{method-ervd-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ervd$new(pm, w, minmax = "max", ref_points, lambda = 2.25, alpha = 0.88)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{ref_points}}{reference points for the criteria}

\item{\code{lambda}}{alternation factor of the losses, 2.25 by default, should be in <2;2.5>}

\item{\code{alpha}}{diminishing sensitivity parameter, 0.88 by default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ervd-compute"></a>}}
\if{latex}{\out{\hypertarget{method-ervd-compute}{}}}
\subsection{Method \code{compute()}}{
computes ERVD model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ervd$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ervd-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ervd-summary}{}}}
\subsection{Method \code{summary()}}{
Creates summary for ERVD model and sends it to console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ervd$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ervd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ervd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ervd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
