% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_aroman.R
\name{extended_aroman}
\alias{extended_aroman}
\title{Alternative Ranking Order Method Accounting for Two-Step Normalization}
\description{
Extension of \link{aroman} method. Starts with same steps as AROMAN method
 up to computation if Li and Ai and continues with raising these to lambda:

\mjsdeqn{L_i^{'} = L_i^\lambda}

and

\mjsdeqn{A_i^{'} = A_i^\lambda}

And finally we compute Ri:

\mjsdeqn{R_i = e^{A_i^{'} - L_i^{'}}}
}
\examples{

## ------------------------------------------------
## Method `extended_aroman$new`
## ------------------------------------------------

pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- extended_aroman$new(pm, w = w, minmax = minmax)
}
\references{
Bošković, S., Švadlenka, L., Dobrodolac, M., Jovčić, S., & Zanne, M. (2023).
 An Extended AROMAN Method for Cargo Bike Delivery Concept Selection.
 Decision Making Advances, 1(1), 1–9. https://doi.org/10.31181/v120231
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{extendedAROMAN}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{w}}{weight vector}

\item{\code{beta}}{value 0-1 representing weighting factor between 2 types of
normalization (default = 0.5)}

\item{\code{lambda}}{values 0-1, coefficient degree of the criteria type (0.5
by default)}

\item{\code{result}}{dataframe with cost (Li) and Bbenefit (Ai) normalized
aggregated value, final result (Ri) and rank of alternatives.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-extended_aroman-new}{\code{extended_aroman$new()}}
\item \href{#method-extended_aroman-compute}{\code{extended_aroman$compute()}}
\item \href{#method-extended_aroman-summary}{\code{extended_aroman$summary()}}
\item \href{#method-extended_aroman-clone}{\code{extended_aroman$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extended_aroman-new"></a>}}
\if{latex}{\out{\hypertarget{method-extended_aroman-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{extended_aroman$new(pm, w, minmax = "max", beta = 0.5, lambda = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{beta}}{value 0-1 representing weighting factor between 2 types of
normalization (default = 0.5)}

\item{\code{lambda}}{values 0-1, coefficient degree of the criteria type (0.5
by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  C1 = c(0.185, 0.317, 0.555, 0.731, 0.948),
  C2 = c(2.33, 1.08, 6.45, 8.88, 7.39),
  C3 = c(454, 298, 174, 849, 517)
)
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
w <- c(0.25, 0.4, 0.35)
minmax <- c("max", "min", "max")
res <- extended_aroman$new(pm, w = w, minmax = minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extended_aroman-compute"></a>}}
\if{latex}{\out{\hypertarget{method-extended_aroman-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of extended AROMAN model based on class properties.
 Usually is not run manually as constructor calls this method
 automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{extended_aroman$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extended_aroman-summary"></a>}}
\if{latex}{\out{\hypertarget{method-extended_aroman-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the extended AROMAN method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{extended_aroman$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extended_aroman-clone"></a>}}
\if{latex}{\out{\hypertarget{method-extended_aroman-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{extended_aroman$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
