% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzytopsis.R
\name{fuzzytopsis}
\alias{fuzzytopsis}
\title{TOPSIS method used to solve multiple criteria decision making based on fuzzy
 sets}
\description{
The acronym TOPSIS stands for: Technique of Order Preference Similarity to
 the Ideal Solution. As name suggests TOPSIS provides its guidance based on
 evaluation of the similarity to both ideal and anti-ideal variant of the
 solution.

Original method uses 5 steps for the procedure. In first step the procedure
 normalizes values in performance matrix nad applies weights to it in step
 2. In step 3 ideal variant \mjseqn{A^*} and anti-ideal variant \mjseqn{A^-}
 is computed as maximums and minimums of the criteria in performance matrix.

In step 4 distance to ideal \mjseqn{D^*} and anti-ideal variant \mjseqn{D^-}
 is computed and in step 5 used to compute closenes criterium (CC).

Criterium CC is then directly usable to rank alternatives. CC is always in
 interval of 0-1, the closer the value is to 1, the closer it is to ideal
 variant.

The approach described above is same as for TOPSIS method
 (see \link{topsis}), thou in case of FuzzyTOPSIS triangular fuzzy
 numbers are used to describe the values in both criteria weights and
 performance matrix.

Note: in present version of the function only benefit criteria are being
 supported.

Mathematically main difference to TOPSIS is requirement to create crips
 number from triangular fuzzy number using following formula to compute
 perceived distance between the alternatives:

\mjsdeqn{ dist = \sqrt{\frac{(a_1 - b_1)^2  + (a_2 - b_2)^2 + (a_3 - b_3)^2}{3}}}

Where dist ... distance between alternatives a and b, both a and b are
 specifed in the form of triangle \mjseqn{a = (a_1, a_2, a_3)} and
 \mjseqn{b = (b_1, b_2, b_3)}.

Technically we compute the distance as square root of average square of
 differences between the alternatives.
}
\examples{

## ------------------------------------------------
## Method `fuzzytopsis$new`
## ------------------------------------------------

dictionaryW <- rbind(
 c(0, 0, 0.1),
 c(0, 0.1, 0.3),
 c(0.1, 0.3, 0.5),
 c(0.3, 0.5, 0.7),
 c(0.5, 0.7, 0.9),
 c(0.7, 0.9, 1),
 c(0.9, 1, 1)
)
# V = very, L = low, M = medium, H = high
rownames(dictionaryW) <- c('VL', 'L', 'ML', 'M', 'MH', 'H', 'VH')
w <- rbind(
 c('H', 'VH', 'VH'),
 c('M', 'H', 'VH'),
 c('M', 'MH', 'ML'),
 c('H', 'VH', 'MH')
)
rownames(w) <- c('Investment cost', 'Employment needs', 'Social impact',
 'Environmental impact')
dictionaryPM <- rbind(
 c(0,0,1),
 c(0,1,3),
 c(1,3,5),
 c(3,5,7),
 c(5,7,9),
 c(7,9,10),
 c(9,10,10)
)
#V very, P poor, M medium, F fair, G good
rownames(dictionaryPM) <- c('VP', 'P', 'MP', 'F', 'MG', 'G', 'VG')
PM <- rbind(
 c('VG', 'G', 'MG'),
 c('MP', 'F', 'F'),
 c('MG', 'MP', 'F'),
 c('MG', 'MG', 'VG'),
 c('VP', 'P', 'G'),
 c('F', 'G', 'G'),
 c('F', 'MG', 'MG'),
 c('F', 'VG', 'G'),
 c('MG', 'MG', 'VG'),
 c('G', 'G', 'VG'),
 c('P', 'VP', 'MP'),
 c('F', 'MP', 'MG'),
 c('P', 'P', 'MP'),
 c('MG', 'VG', 'G'),
 c('MP', 'F', 'F'),
 c('MG', 'VG', 'G'),
 c('G', 'G', 'VG'),
 c('VG', 'MG', 'F'),
 c('G', 'VG', 'G'),
 c('MG', 'F', 'MP'),
 c('MP', 'P', 'P'),
 c('VP', 'F', 'P'),
 c('G', 'MG', 'MG'),
 c('P', 'MP', 'F')
)
alternatives <- c('site 1', 'site 2', 'site 3', 'site 4', 'site 5',
 'site 6')
result <- fuzzytopsis$new(PM, dictionaryPM, w, dictionaryW, alternatives)
}
\references{
Papathanasiou, Jason, Ploskas, Nikolaos. Multiple Criteria Decision Aid
 Methods, Examples and Python Implementations. Springer, 173 p., ISBN
 978-3-319-91648-4.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{Outranking}
\keyword{TOPSIS}
\keyword{analysis}
\keyword{anti-ideal}
\keyword{approaches}
\keyword{fuzzy}
\keyword{ideal}
\keyword{modelling}
\keyword{multicriteria}
\keyword{preference}
\keyword{set}
\keyword{solution}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{normalized performance matrix (direction of all criteria is
maximalize)}

\item{\code{dictionary_pm}}{dictionary of linguistic variables for criteria}

\item{\code{w}}{weght vector}

\item{\code{dictionary_w}}{dictionary for wights (single matrix or dataframe)}

\item{\code{alt}}{vector with names of the alternatives}

\item{\code{n}}{number of decision makers}

\item{\code{fuzzy_weights}}{fuzzy weights}

\item{\code{fuzzy_decision_matrix}}{fuzzy decision matrix}

\item{\code{wfndm}}{weighted normalized fuzzy decision matrix}

\item{\code{a_plus}}{a+}

\item{\code{a_minus}}{a-}

\item{\code{cc}}{Alternatives sorted by closeness coeficient}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fuzzytopsis-new}{\code{fuzzytopsis$new()}}
\item \href{#method-fuzzytopsis-compute}{\code{fuzzytopsis$compute()}}
\item \href{#method-fuzzytopsis-summary}{\code{fuzzytopsis$summary()}}
\item \href{#method-fuzzytopsis-clone}{\code{fuzzytopsis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzytopsis-new"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzytopsis-new}{}}}
\subsection{Method \code{new()}}{
public constructor, creates fuzzytopsis object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzytopsis$new(pm, dictionary_pm, w, dictionary_w, alt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{dictionary_pm}}{dictionary scale for performance matrix}

\item{\code{w}}{weights}

\item{\code{dictionary_w}}{dicrtionary scale for weights}

\item{\code{alt}}{alternative's names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dictionaryW <- rbind(
 c(0, 0, 0.1),
 c(0, 0.1, 0.3),
 c(0.1, 0.3, 0.5),
 c(0.3, 0.5, 0.7),
 c(0.5, 0.7, 0.9),
 c(0.7, 0.9, 1),
 c(0.9, 1, 1)
)
# V = very, L = low, M = medium, H = high
rownames(dictionaryW) <- c('VL', 'L', 'ML', 'M', 'MH', 'H', 'VH')
w <- rbind(
 c('H', 'VH', 'VH'),
 c('M', 'H', 'VH'),
 c('M', 'MH', 'ML'),
 c('H', 'VH', 'MH')
)
rownames(w) <- c('Investment cost', 'Employment needs', 'Social impact',
 'Environmental impact')
dictionaryPM <- rbind(
 c(0,0,1),
 c(0,1,3),
 c(1,3,5),
 c(3,5,7),
 c(5,7,9),
 c(7,9,10),
 c(9,10,10)
)
#V very, P poor, M medium, F fair, G good
rownames(dictionaryPM) <- c('VP', 'P', 'MP', 'F', 'MG', 'G', 'VG')
PM <- rbind(
 c('VG', 'G', 'MG'),
 c('MP', 'F', 'F'),
 c('MG', 'MP', 'F'),
 c('MG', 'MG', 'VG'),
 c('VP', 'P', 'G'),
 c('F', 'G', 'G'),
 c('F', 'MG', 'MG'),
 c('F', 'VG', 'G'),
 c('MG', 'MG', 'VG'),
 c('G', 'G', 'VG'),
 c('P', 'VP', 'MP'),
 c('F', 'MP', 'MG'),
 c('P', 'P', 'MP'),
 c('MG', 'VG', 'G'),
 c('MP', 'F', 'F'),
 c('MG', 'VG', 'G'),
 c('G', 'G', 'VG'),
 c('VG', 'MG', 'F'),
 c('G', 'VG', 'G'),
 c('MG', 'F', 'MP'),
 c('MP', 'P', 'P'),
 c('VP', 'F', 'P'),
 c('G', 'MG', 'MG'),
 c('P', 'MP', 'F')
)
alternatives <- c('site 1', 'site 2', 'site 3', 'site 4', 'site 5',
 'site 6')
result <- fuzzytopsis$new(PM, dictionaryPM, w, dictionaryW, alternatives)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzytopsis-compute"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzytopsis-compute}{}}}
\subsection{Method \code{compute()}}{
computes the FuzzyTOPSIS problem. Usually doesn't need to be run manually
 as it is called automatically by class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzytopsis$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzytopsis-summary"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzytopsis-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the Fuzzy TOPSIS resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzytopsis$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fuzzytopsis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fuzzytopsis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fuzzytopsis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
