% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabac.R
\name{mabac}
\alias{mabac}
\title{Multi-Attribute Border Approximation are Comparison}
\description{
Method was introduces by Pamucar and Cirovic in 2015. Its recommendations
 are based on alternatives distance from border aproximation area.

The precedure starts with normalizong performance matrix of alternatives
 (rows) in criteria (columns) using min-max normalization. Then a weighted
 normalized matrix is computed

\mjsdeqn{\hat{r}_{ij} = w_j + r_{ij}^*w_j}

Where r* is normalized perrformance matrix values.

Next border approximation area vector (g) is computed

\mjsdeqn{g_j = (\prod_{i=1}^m \hat{r}_{ij})^{1/m}}

m is number of alternatives being considered. Next the distance from border
 approximation area (q) is computed

\mjsdeqn{q_{ij} = \hat{r}_{ij} - g_j}

Total distance form border approximation area (S) can be computed by summing
 rows of matrix q. This measure is also directly applicable for purposes of
 ranking by softring the alternatives descentively using this measure.

\mjsdeqn{S_i = \sum_{j=1}^n q_{ij}}
}
\examples{

## ------------------------------------------------
## Method `mabac$new`
## ------------------------------------------------

alternatives <- c("A1", "A2", "A3", "A4")
criteria <- c("C1", "C2", "C3", "C4")
pm <- cbind(
  c(5, 1, 7, 10),
  c(54, 97, 72, 75),
  c(600, 200, 400, 1000),
  c(80, 65, 83, 40)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
minmax <- c("min", "min", "max", "max")
w <- rep(1, times = 4)
t <- mabac$new(pm, w, minmax)
}
\references{
Alinezhad, A., Khalili, J. New Methods and Applications in Multiple
 Attribute Decision Making (MADM). Springer Nature Switzerland AG, 2019,
 233 p., ISBN 978-3-030-15009-9
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MABAC}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{normalized performance matrix. Criteria (in columns) are
expected to be ordered from most influential to least influential.}

\item{\code{pm_orig}}{original (not normalized) performance matrix}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}

\item{\code{w}}{weights vector}

\item{\code{dist_border}}{total distance from border approximation area}

\item{\code{final_rank}}{final ranking of the alternatives based on dist_border}

\item{\code{result}}{dataframe with distance from border approximation area and rank}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mabac-new}{\code{mabac$new()}}
\item \href{#method-mabac-compute}{\code{mabac$compute()}}
\item \href{#method-mabac-summary}{\code{mabac$summary()}}
\item \href{#method-mabac-clone}{\code{mabac$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mabac-new"></a>}}
\if{latex}{\out{\hypertarget{method-mabac-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mabac$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria in columns, ordered from most to
least influential}

\item{\code{w}}{weights vector}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c("A1", "A2", "A3", "A4")
criteria <- c("C1", "C2", "C3", "C4")
pm <- cbind(
  c(5, 1, 7, 10),
  c(54, 97, 72, 75),
  c(600, 200, 400, 1000),
  c(80, 65, 83, 40)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
minmax <- c("min", "min", "max", "max")
w <- rep(1, times = 4)
t <- mabac$new(pm, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mabac-compute"></a>}}
\if{latex}{\out{\hypertarget{method-mabac-compute}{}}}
\subsection{Method \code{compute()}}{
Computed the MABAC using params provided in constructor. Usually run
 automatically by constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mabac$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mabac-summary"></a>}}
\if{latex}{\out{\hypertarget{method-mabac-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the MABAC method results and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mabac$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mabac-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mabac-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mabac$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
