% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcda_objective_weights.R
\name{mcda_objective_weights}
\alias{mcda_objective_weights}
\title{Compute weights of the criteria using objective weighting methods.}
\usage{
mcda_objective_weights(pm, method, minmax = "max", norm = "minmax")
}
\arguments{
\item{pm}{performance matrix}

\item{method}{weight computation method}

\item{minmax}{'min' or 'max' to specify cost or benefit criterion, max is
default value}

\item{norm}{specify normalization method, use constant from table above,
the 'minmax' is default value.}
}
\value{
weights for the criteria
}
\description{
The function for criteria weights computation using various objective
 weighting methods. At present time supported are:

\tabular{llc}{
       \bold{constant} \tab \bold{method} \tab \bold{normalization}\cr
       MW \tab Mean weighting method \tab N \cr
       SDW \tab Standard Deviation Weighting method \tab Y \cr
       SVW \tab Statistical Variance Weighting method \tab Y \cr
       EWM \tab Entropy Weight Method \tab N \cr
       CRITIC \tab Criterion Importance Through Intercriteria Correlation \tab N \cr
       GCW \tab Gini Coefficient Weighting \tab N \cr
       MEREC \tab Method based on Removal Effects of Criteria \tab N \cr
       CILOS \tab Criterion Impact LOSs \tab N \cr
       IDOCRIW \tab Integrated Determination of Objective CRIteria Weights \tab N \cr
   }

In normalization column of previous table Y means that the analyst can also
 choose normalization method himself. All normalization methods from
 mcda_norm can be used. If no normalization method is provided, the function
 uses default one for the method, usually min-max normalization.

\bold{MW - Mean weighting Method}

Method presumes that the weights of all criteria are the same (there are no)
 meaningfull differences between them. Thus method approximates the weigts
 as a mean value:

\mjsdeqn{w_j = \frac{1}{n}}

where n is number of criteria in decision problem.

\bold{SDW - Standard Deviation Weighting Method}

Computes weights based on standard deviation of the criteria.

Steps:

1) normalize data, min-max method is default, but you can choose to use
 other normalization method

2) compute standard deviation (SD) of the population for the criteria:

\mjsdeqn{\sigma_j = \sqrt{\frac{\sum_{i=1}^m (F_{ij} - \overline{F_j})^2}{m}}}

For alll j in (1, 2, ..., n), where F`j is arithmetic mean of normalized
 values.

3) calcutate the weights by normalizing the sigma to max sigma

\mjsdeqn{w_j = \frac{\sigma_j}{\sum_{k = 1}^m \sigma_k}}

\bold{SVW - Statistical Variance Weighting method}

Allows to compute the weights of criteria based on differences in variance
 in performance of altrernatives in the criteria.

Steps:

1) normalize data, min-max method is default, but you can choose to use
 other normalization method

2) compute variance of normalized data. Again population variance is being used

\mjsdeqn{\sigma_j^2 = \frac{\sum_{i=1}^m (F_{ij} - \overline{F_j})^2}{m}}

3) Calculate the weights

\mjsdeqn{w_j = \frac{\sigma_j^2}{\sum_{k = 1}^m \sigma_k^2}}

\bold{EWM - Entropy Weight Method}

Allows us to establish weight system just based on evaluation of the entropy
 in input data of the performance matrix.

\mjsdeqn{EM_{ij} = \frac{PM_{ij}}{max(PM_{ij})_j}}

Then we compute probability of criteria to occur.

\mjsdeqn{p_{ij} = \frac{EM_{ij}}{\sum_{i=1}^n EM_{ij}}}

Where j = criteria, i = alternatives, n = number of alternatives.

Then we compute entropy E:

\mjsdeqn{E_j = -P \sum_{i=1}^n p_{ij} * ln (p_{ij})}

Where

\mjsdeqn{P = \frac{1}{ln(n)}}.

Value of E is then in interval of <0;1>.

Then we compute degree of divergence:

\mjsdeqn{div_j = | 1 - E_j|}

and from it entropy weights which are returned as result of the function:

\mjsdeqn{Ew_j = \frac{div_j}{\sum_{j=1}^m div_j}}

\bold{CRITIC - Criterion Importance Through Intercriteria Correlation}

Derives weight coefficients of criteria based on correlations.

Basis of the evaluation is correlation matrix R we compute using Pearson's
 correlation coefficien. This information can be used to compute the
 conflict created by the criterium with respect to other criteria:

\mjsdeqn{\sum_{k=1}^m (1 - r_{jk})}

Then we determine quality of the information provided by multiplying the
 result by standard deviation

\mjsdeqn{C_j = \sigma_j \sum_{k=1}^m (1 - r_{jk}) }

Finally we can compute the weight by normalizing the C indicator

\mjsdeqn{w_j = \frac{C_j}{\sum_{k=1}^m C_j}}

\bold{GCW - Gini Coefficient Weighting}

Computes weights based on Gini index.

First we compute Gini coefficient (Gj) based on pairwise absolute difference
 of the discrete values of criteria. (No normalization needed).

\mjsdeqn{G_j = \frac{\sum_{i=1}^m \sum_{k=1}^m |f_{ij} - f_{kj}|}{2m^2\overline{f_j}}}

Then we calculate weights:

\mjsdeqn{w_j = \frac{G_j}{\sum_{k=1}^m G_k}}

\bold{MEREC - Method based on Removal Effects of Criteria}

Computes weights based by considering the effect of criterium removal from
 the evaluation on overall performance of the alternatives.

We start with normalization of performance matrix. For benefit criteria we
 use:

\mjsdeqn{n_{ij} = \frac{min_k x_{kj}}{x_{ij}}}

for cost criteria

\mjsdeqn{n_{ij} = \frac{x_{ij}}{max_k x_{kj}}}

Then we compute overal performance of overal performance of the alternatives
 Si (step 3)

\mjsdeqn{S_i = ln(1 + (\frac{1}{m} \sum_j |ln(n_{ij})|)}

In step 4 the calculation of performance alternatives by removig each
 attribute (Sij') is performed

\mjsdeqn{S_{ij}^\prime = ln(1 + (\frac{1}{m} \sum_{k,k \ne j} |ln(n_{ik})|))}

Then a Removal effect (Ej) is calculated (step 5).

\mjsdeqn{E_j = \sum_j |S_{ij}^\prime - S_i|}

Finally the final weights (wj) are coputed in step 6.

\mjsdeqn{w_j = \frac{E_j}{\sum_k E_k}}

\bold{CILAS - Criterion Impact LOSs}

Weight derivation based on measure impact loss.

The method is interesting as it was first coined by Mirkin in 1974, but only
 in general terms. True formulation was performed by Zavadskas and Podvezko
 in 2016 for purposes of their method IDOCRIW. For their purposes they named
 the method CILOS.

As first step the method requires all criteria to be maximized. Minimized
 criteria are expected to be translated to maximalization using

\mjsdeqn{r_{ij} = \frac{min_i r_{ij}}{r_{ij}}}

Next the whole performance matrix needs to be normalized.

\mjsdeqn{x_{ij} = \frac{r_{ij}}{\sum_{i = 1}^n r_{ij}}}

Next matrix A is constructed. To do that we need to identify maximal values
 \mjseqn{x_j} in each columns of the normalized performance matrix and rows
 \mjseqn{k} on which they reside.

\mjsdeqn{a_{ij} = x_{kj}}

meaning for example, that if in first column of the performance matrix the
 maximum was on third row, then

\mjsdeqn{a_{11} = x_{31}}

In short we construct the matrix A by copying rows from normalized
 performance matrix X corresponding to the row the maximum in the column
 was. That also means, that if we identified in such maximums in multiple
 columns, then the rows will repeat in matrix A.

Next we compute matrix of relative losses P

\mjsdeqn{p_{ij} = \frac{x_j - a_{ij}}{x_j}}

Since the matrix A will always have identified maximums in its diagonal, the
 matrix P will always have zeros in its diagonal.

Finaly we construct weight system matrix (F). We start with matrix P and
 minus sum the columns in the diagonal. The diagonal of F will thus always
 have negative numbers and sum of the column in F = 0.

We will use the F as a basis for weight derivation

\mjsdeqn{Fq^T = 0}

by solving the equation to get q and normalizing it we get the desired
 weights.

Please note that the system of equation is clearly homogenous system of
 equations with trivial solution q = 0, which is not useful for purposes of
 weight establishments. For establishment of weights non-trivial solutions
 need to be found.

This implementation of the method uses singular valuable decomposition (SVD) of
 matrix F for this purpose.

Final values of the weights is achieved by normalizing the solution.

\bold{IDOCRIW - Integrated Determination of Objective CRIteria Weights}

Introfuced by Zavadskas and Podvezko in 2016. Formaly it combines two other
 methods for weight establishment: EWM and CILOS, in following manner:

\mjsdeqn{w_j = \frac{q_j \cdot ew_j}{\sum_{i = 1}^n q_j \cdot ew_j}}

where q are weights established using CILOS method and ew are weights
 derived using EWM method.
}
\examples{
alternatives <- c("A1", "A2", "A3", "A4", "A5")
c1 <- c(0.185, 0.3, 0.72, 0.44, 0.98)
c2 <- c(3.14, 1, 7.4, 6.4, 8.2)
c3 <- c(420, 750, 300, 634, 248)
pm <- cbind(c1, c2, c3)
rownames(pm) <- alternatives
minmax <- c("max", "min", "max")
t <- mcda_objective_weights(pm, method = "MW", minmax = minmax, norm = "minmax")

}
\references{
Kumar, Raman; Bilga, Paramjit Singh Singh, Sehijpal. Multi objective
 optimization using different methods of assigning weights to energy
 consumption responses, surface roughness and material removal rate during
 rough turning operation. Journal of Cleaner Production, vol. 16, pp. 45-57,
 DOI: 10.1016/j.jclepro.2017.06.077.

Diakonlaki, D., Mavrotas, G., Papayannadis, J. (1995). Datamining Objective
 Weights in Multiple Criteria Problems: the CRITIC Method. Computers and
 Operations Research, 22(7), pp. 763-770.

Bhangale, P. P., Agrawal, V. P., & Saha, S. K. (2004). Attribute based
 specification, comparison and selection of a robot. Mechanism and Machine
 Theory, 39(12), 1345-1366.

Keshavarz-Ghorabaee, M., Amiri, M., Zavadskas, E. K., Turskis, Z., &
 Antucheviciene, J. (2021). Determination of objective weights using a new
 method based on the removal effects of criteria (MEREC). Symmetry, 13(4),
 525.

Alinezhad, A., Khalili, J. New Methods and Applications in Multiple
 Attribute Decision Making (MADM). Springer Nature Switzerland AG, 2019,
 233 p., ISBN 978-3-030-15009-9

Zavadskas, E. K., Podvezko, V. Integrated Determination of Objective
 Criteria Weights in MCDM. International Journal of Information Technology &
 Decision Making, vol. 15, DOI: 10.1142/S0219622016500036.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MW}
\keyword{weighting}
