% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melchior.R
\name{melchior}
\alias{melchior}
\title{Méthode d'ELimination et de CHoix Incluant les relations d'ORdre}
\description{
Method is based on specification of preferences on alternatives in criteria.
 Methods works with 3 types of preference: 1) weak Preference (Q) in situation
 when thre is good reson to prefer one alternative, but not to be strictly
 prefering it. 2) strict preference (P) - it is justified to prefer one
 alternative. 3) Veto (V): value of performance difference gj(b)-gj(a), from
 which the proposition aSb will not be accepted.

For computational point of view also indifference threshold qj and a preference
 threshold pj  with pj > qj >= 0.

Outranking relationship is established by followwing the rules:

\tabular{llc}{
   \bold{relationship} \tab \bold{condition} \cr
   aPj+b \tab If gj(a) > gj(b) + pj \cr
   aQj+b \tab If gj(a) > gj(b) + pj and gj(a) > gj(b) + qj \cr
   aIj+b \tab If gj(a) > gj(b) + qj and gj(a) > gj(b) \cr
   aEjb \tab If gj(a) = gj(b) \cr
   aPj-b \tab If bPj+a \cr
   aQj-b \tab If bQj+a \cr
   aIj-b \tab If bIj+a
}

For string outranking (SF), there are no criteria for which b is strictly
 preferable to a; Criterion i for which b is weakly preferable to a must be
 masked by more important criteria for which a enjoys strict preference.

Weak outranking (Sf) it is necessary that the criteria i for which b has the
 advantage must be masked by criteria j at least as important in favor of a.
}
\examples{

## ------------------------------------------------
## Method `melchior$new`
## ------------------------------------------------

pm <- rbind(
  c(39.6, 19.8, 4.5),
  c(71.5, 34.6, 7.5),
  c(67.2, 62.6, 5.2),
  c(66.3, 61, 4.7)
)
alt <- c(paste0("A", 1:4))
cri <- c(paste0("C", 1:3))
colnames(pm) <- cri
rownames(pm) <- alt
minmax <- c("max", "max", "min")
p <- c(10, 13, 0.5)
q <- c(20, 20, 1)
v <- c(40, 50, 4)
t <- melchior$new(pm, minmax, p, q, v)
}
\references{
COSTA, Igor Pinheiro de Araújo et al. Algorithm Selection for Machine Learning
Classification:An Application of the MELCHIOR Multicriteria Method. In: Modern
 Management based on Big Data II and Machine Learning and Intelligent Systems
 III, IOS Press, 2021, pp. 154-161, doi:10.3233/FAIA210243
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MELCHIOR}
\keyword{outranking}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria (in columns) are expected to be
ordered from most influential to least influential.}

\item{\code{minmax}}{vector of optimization direction for criteria (min/max)}

\item{\code{p}}{preference threshold}

\item{\code{q}}{indifference threshold}

\item{\code{v}}{veto threshold}

\item{\code{outranking}}{data frame with outranking relationship}

\item{\code{pairwise_outranking}}{criteria paiwise comparison to establish
outraniking relationship}

\item{\code{result}}{dataframe with final score and rank}

\item{\code{outranking_graph}}{visualization of the network of strong and weak
outranking connections}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-melchior-new}{\code{melchior$new()}}
\item \href{#method-melchior-compute}{\code{melchior$compute()}}
\item \href{#method-melchior-summary}{\code{melchior$summary()}}
\item \href{#method-melchior-graph}{\code{melchior$graph()}}
\item \href{#method-melchior-clone}{\code{melchior$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-melchior-new"></a>}}
\if{latex}{\out{\hypertarget{method-melchior-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{melchior$new(pm, minmax = "max", p, q, v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria in columns}

\item{\code{minmax}}{vector of optimization direction (min/max, max is default)}

\item{\code{p}}{preference threshold}

\item{\code{q}}{indifference threshold}

\item{\code{v}}{veto threshold}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- rbind(
  c(39.6, 19.8, 4.5),
  c(71.5, 34.6, 7.5),
  c(67.2, 62.6, 5.2),
  c(66.3, 61, 4.7)
)
alt <- c(paste0("A", 1:4))
cri <- c(paste0("C", 1:3))
colnames(pm) <- cri
rownames(pm) <- alt
minmax <- c("max", "max", "min")
p <- c(10, 13, 0.5)
q <- c(20, 20, 1)
v <- c(40, 50, 4)
t <- melchior$new(pm, minmax, p, q, v)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-melchior-compute"></a>}}
\if{latex}{\out{\hypertarget{method-melchior-compute}{}}}
\subsection{Method \code{compute()}}{
computes MELCHIOR model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{melchior$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-melchior-summary"></a>}}
\if{latex}{\out{\hypertarget{method-melchior-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the MELCHIOR method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{melchior$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-melchior-graph"></a>}}
\if{latex}{\out{\hypertarget{method-melchior-graph}{}}}
\subsection{Method \code{graph()}}{
functions prepares network diagram of outranking relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{melchior$graph()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
plot_ly network graph with outranking relations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-melchior-clone"></a>}}
\if{latex}{\out{\hypertarget{method-melchior-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{melchior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
