% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msim.R
\name{msim}
\alias{msim}
\title{Modified Similarity Index}
\description{
uses modeified \link{topsis}. The modification start in later steps of the
 procedure, namely step 6, when degree of conflict bertween each alternative
 and ideal and antiideal se coumputed.

\mjsdeqn{cos e_i^+ = \frac{\sum_{j=1}^m y_{ij}A_j^+}{\sqrt{\sum_{j=1}^m y_{ij}^2}\sqrt{\sum_{j=1}^m A_j^{+2}}}}

\mjsdeqn{cos e_i^- = \frac{\sum_{j=1}^m y_{ij}A_j^-}{\sqrt{\sum_{j=1}^m y_{ij}^2}\sqrt{\sum_{j=1}^m A_j^{-2}}}}

Determine degree of similarity between each pair of alternatives and ideal
 and antiideal variant.

\mjsdeqn{S_i^+ = \frac{cos e_i^+ \sqrt{\sum_{j=1}^m y_{ij}^2}}{|A_j^+|}}

\mjsdeqn{S_i^- = \frac{|A_j^+|}{cos e_i^- \sqrt{\sum_{j=1}^m y_{ij}^2}}}

Finally overall performance score is computed:

 \mjsdeqn{P_i = \frac{S_i^+}{S_i^+ + S_i^-}}
}
\examples{

## ------------------------------------------------
## Method `msim$new`
## ------------------------------------------------

alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
minmax <- "max"
t <- msim$new(M, w, minmax)
}
\references{
Chakrakborty, S., Chatterjee, P. Das, P. P. Multi-Criteria Decision Making
 Methods in Manufacturing Environments: Models and Applications. CRC Press,
 Boca Raton, 2024, 450 p., ISBN: 978-1-00337-703-0
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{w}}{numeric weight vector}

\item{\code{minmax}}{vector of optimization directions - min/max}

\item{\code{result}}{dataframe with computational results: degree of conflict
with ideal (cos_e_plus) and antiideal (cos_e_minus), similarity to ideal
(s_plus), similarity to antiideal (s_minus), performance (Ri), rank}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-msim-new}{\code{msim$new()}}
\item \href{#method-msim-compute}{\code{msim$compute()}}
\item \href{#method-msim-summary}{\code{msim$summary()}}
\item \href{#method-msim-clone}{\code{msim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msim-new"></a>}}
\if{latex}{\out{\hypertarget{method-msim-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msim$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
minmax <- "max"
t <- msim$new(M, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msim-compute"></a>}}
\if{latex}{\out{\hypertarget{method-msim-compute}{}}}
\subsection{Method \code{compute()}}{
Computes ROV based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msim$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msim-summary"></a>}}
\if{latex}{\out{\hypertarget{method-msim-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the MSIM method results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msim$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-msim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
