% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oreste.R
\name{oreste}
\alias{oreste}
\title{ORESTE}
\description{
Simple MADM method for decision support in situation. Proposed by Roubens in
 1983.

The method is based on computation of the distance of the rank of the
 alternative A in criterium X to rank 0 (optimum).

Method takes as input perfomance matrix, criteria (in columns) must be
 ordered from the most important one to least important one. So for example
 first criterium is more important than the second, etc.

Next we create position matrix, in which we rank the alternatives for each
 criterium separately. This information is used for computation of block
 distances:

\mjsdeqn{d(0, A_{ij}) = \alpha r_{ij} + (1 - \alpha) r_j}

Where r ij is the ran of the alternative i in criterium j, rj then means
 column, where the criterium i located. The positioning is important as all
 critera are ordered by their significance, making criterium 1 the most
 important one, with r1 = 1 (second criterium r2 = 2, ...).

The results are writen to block distance matrix.

Value alpha presents succession rate. Its values must always be in interval
 (0;1), but usually we are choosing aplpha such that

\mjsdeqn{1/3 \le \alpha \le 1/2}

For final ranking we presume that lower distances to the theoretical optimum
 will be represenative of better rank of such alternative.

\mjsdeqn{d(0, A_{ij}) \le d(0, A_{i'j})}

\mjsdeqn{R(A_{ij}) \le R(A_{ij})}

Value of R is derived from values of computed block distances. We order all
 of these from lowest to highest and rank these values. We then substitute
 the values in block distance matrix with the rank assigned to this distance
 in previous step.

We can compute the final value for ranking by aggregating these distance
 ranks for the alternative across the criteria.

\mjsdeqn{R(A_{i}) = \sum_{j = 1}^n R(A_{ij})}

Resulting R(A) can be directly used for ranking purposes (highest value is
 best).
}
\examples{

## ------------------------------------------------
## Method `oreste$new`
## ------------------------------------------------

alternatives <- c("A1", "A2", "A3")
criteria <- c("C1", "C2", "C3")
pm <- rbind(
  c(4, 3, 3),
  c(2, 1.2, 2),
  c(1, 1.5, 4)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
minmax <- c("max", "min", "max")
alpha <- 0.4
t <- oreste$new(pm, minmax, alpha)
}
\references{
Alinezhad, A., Khalili, J. New Methods and Applications in Multiple
 Attribute Decision Making (MADM). Springer Nature Switzerland AG, 2019,
 233 p., ISBN 978-3-030-15009-9
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{MADM}
\keyword{ORESTE}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria (in columns) are expected to be
ordered from most influential to least influential.}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}

\item{\code{alpha}}{succession rate in interval <0, 1>, usually in <0.33, 0.5>}

\item{\code{finalRank}}{- final ranking of the alternatives}

\item{\code{ra}}{total ranking of the alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-oreste-new}{\code{oreste$new()}}
\item \href{#method-oreste-compute}{\code{oreste$compute()}}
\item \href{#method-oreste-summary}{\code{oreste$summary()}}
\item \href{#method-oreste-clone}{\code{oreste$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oreste-new"></a>}}
\if{latex}{\out{\hypertarget{method-oreste-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{oreste$new(pm, minmax = "max", alpha = 0.4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria in columns, ordered from most to
least influential}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}

\item{\code{alpha}}{succession rate in interval <0, 1>, usually in <0.33, 0.5>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c("A1", "A2", "A3")
criteria <- c("C1", "C2", "C3")
pm <- rbind(
  c(4, 3, 3),
  c(2, 1.2, 2),
  c(1, 1.5, 4)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
minmax <- c("max", "min", "max")
alpha <- 0.4
t <- oreste$new(pm, minmax, alpha)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oreste-compute"></a>}}
\if{latex}{\out{\hypertarget{method-oreste-compute}{}}}
\subsection{Method \code{compute()}}{
computes ORESTE model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{oreste$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oreste-summary"></a>}}
\if{latex}{\out{\hypertarget{method-oreste-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the ORESTE method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{oreste$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oreste-clone"></a>}}
\if{latex}{\out{\hypertarget{method-oreste-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{oreste$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
