% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piv.R
\name{piv}
\alias{piv}
\title{Proximity Indexed Value (PIV)}
\description{
Introduced by Mufazzel and Muzakkin (2018) to mitigate rank reversal
 problem. The procedure takes ideal solution and evaluates distance of the
 alternatives to it.

In comparison TOPSIS evaluates both ideal and intiideal solution.

This simplification introduces some limitations for the results (compared
 to TOPSIS). Ignoring closeness to antiideal solution is not advised in
 decision situations with risk considerations. Succesible to inflation of
 extreme values to name few.

The procedure starts with 1) normalization of performance matrix using vector
 normalization. 2) Apply weights. 3) determine ideal solution A* and
 4) calculate overall proximity value of Di:

\mjsdeqn{D_i = \sum_{j=1}^n |v_{ij} - v_j^+ |}

and use it to rank the alternatives.

Note: since PIV is so close to TOPSIS this class just computes TOPSIS and
 publishes parts relevant to PIV.
}
\examples{

## ------------------------------------------------
## Method `piv$new`
## ------------------------------------------------

PM <- cbind(
   c(8,7,2,1),
   c(5,3,7,5),
   c(7,5,6,4),
   c(9,9,7,3),
   c(11,10,3,7),
   c(6,9,5,4)
)
colnames(PM) <- c('Site 1', 'Site 2', 'Site 3', 'Site 4', 'Site 5',
   'Site 6')
rownames(PM) <- c('Investment costs (million EUR)',
                  'Employment needs (hundred employees)',
                  'Social impact (1-7)',
                  'Environmental impact (1-7)')
PM <- t(PM)
minmax <- 'max'
w <- c(0.4, 0.4, 0.1, 0.2)
result <- piv$new(PM, w, minmax)
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{PIV}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not transformed) performance matrix of
alternatives in criteria}

\item{\code{pm}}{performance matrix of alternatives in criteria (all criteria
are maximized)}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{normPM}}{normalized performance matrix}

\item{\code{weightPM}}{weighted normalized performance matrix}

\item{\code{A_ideal}}{positive ideal solution}

\item{\code{D_ideal}}{alternative closeness to ideal variant}

\item{\code{ranks}}{ranks derived from distance to ideal solution (D_ideal) measure}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-piv-new}{\code{piv$new()}}
\item \href{#method-piv-summary}{\code{piv$summary()}}
\item \href{#method-piv-clone}{\code{piv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-piv-new"></a>}}
\if{latex}{\out{\hypertarget{method-piv-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for TOPSIS object. Performas validation of input
 parameters and uses them to compute the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{piv$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion.
Only numeric values expercted. Rows and columns are expected to be
named.}

\item{\code{w}}{vector containing the weights of the criteria.}

\item{\code{minmax}}{criteria MinMax Vector containing the preference direction
on each of the criteria. "min" (resp."max") indicates that the
criterion has to be minimized (maximized). Default value: max}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Instance of topsis class uncluding inputs and model results in class's
 properties
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{PM <- cbind(
   c(8,7,2,1),
   c(5,3,7,5),
   c(7,5,6,4),
   c(9,9,7,3),
   c(11,10,3,7),
   c(6,9,5,4)
)
colnames(PM) <- c('Site 1', 'Site 2', 'Site 3', 'Site 4', 'Site 5',
   'Site 6')
rownames(PM) <- c('Investment costs (million EUR)',
                  'Employment needs (hundred employees)',
                  'Social impact (1-7)',
                  'Environmental impact (1-7)')
PM <- t(PM)
minmax <- 'max'
w <- c(0.4, 0.4, 0.1, 0.2)
result <- piv$new(PM, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-piv-summary"></a>}}
\if{latex}{\out{\hypertarget{method-piv-summary}{}}}
\subsection{Method \code{summary()}}{
Creates summary for PIV model and sends it to console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{piv$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-piv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-piv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{piv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
