% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualiflex.R
\name{qualiflex}
\alias{qualiflex}
\title{QUALIFLEX}
\description{
Simple MADM method for decision support. Proposed by Paelinck in 1975.
 Method presumes we are using independent criteria for comparison for
 which weights are known.

The method is based on permutation approach and tries to identify the
 comparison permutation, which is closest to the values provided in the
 performance matrix.

Method is using following steps:

step 1) initial permutation of the alternatives. Number of permutation
 generated is always equal to m!, where m is number of alternatives (i.e.
 m = 3, m! = 6)

step 2) initial ranking of the alternatives - provided performance matrix
 is ranked separately for each criterium

step 3) dominant and dominated values - for each permutation in each
 criterium the ranking provided by permutation is compared with the one in
 performance matrix. Where the permutation agrees with performance matrix
 value 1, when it disagrees -1 and when the alternatives in performance
 matrix are equal 0 is provided.

For example permutation 1 = A2, A1, A3, but true ranking in criterium is
 A1 > A2 = A3, hence A2 < A1 = -1

step 4) The results are consolidated into matrix of permutation values of
 criteria.

step 5) permutation values of alternatives is computed as weighted average
 of permutation values of criteria.

step 6) final ranking is established based on the permutation or
 permutations with highest permutation value.
}
\examples{

## ------------------------------------------------
## Method `qualiflex$new`
## ------------------------------------------------

alternatives <- c("A1", "A2", "A3")
criteria <- c("C1", "C2", "C3")
pm <- rbind(
  c(1.2, 18, 32000),
  c(2, 23, 24000),
  c(2, 15, 25000)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
minmax <- c("min", "max", "max")
w <-c(1, 1, 1)
t <- qualiflex$new(pm, minmax, w)
}
\references{
Alinezhad, A., Khalili, J. New Methods and Applications in Multiple
 Attribute Decision Making (MADM). Springer Nature Switzerland AG, 2019,
 233 p., ISBN 978-3-030-15009-9
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{QUALIFLEX}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria (in columns) are expected to be
ordered from most influential to least influential.}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}

\item{\code{w}}{weight of the criteria}

\item{\code{finalRank}}{final ranking of the alternatives}

\item{\code{permutation_val}}{vermutation value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-qualiflex-new}{\code{qualiflex$new()}}
\item \href{#method-qualiflex-compute}{\code{qualiflex$compute()}}
\item \href{#method-qualiflex-summary}{\code{qualiflex$summary()}}
\item \href{#method-qualiflex-clone}{\code{qualiflex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-qualiflex-new"></a>}}
\if{latex}{\out{\hypertarget{method-qualiflex-new}{}}}
\subsection{Method \code{new()}}{
class constructor, validates data and computes the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qualiflex$new(pm, minmax = "max", w)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix. Criteria in columns, ordered from most to
least influential}

\item{\code{minmax}}{optimization direction vector for criteria. Either min or
max. Can be substituted by single min or max if optimization direction
is same for all criteria. The value needs to be provided for all
numeric criteria (ordinal criteria are already considered processed).}

\item{\code{w}}{weight of the criteria}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of the class including computed model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c("A1", "A2", "A3")
criteria <- c("C1", "C2", "C3")
pm <- rbind(
  c(1.2, 18, 32000),
  c(2, 23, 24000),
  c(2, 15, 25000)
)
rownames(pm) <- alternatives
colnames(pm) <- criteria
minmax <- c("min", "max", "max")
w <-c(1, 1, 1)
t <- qualiflex$new(pm, minmax, w)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-qualiflex-compute"></a>}}
\if{latex}{\out{\hypertarget{method-qualiflex-compute}{}}}
\subsection{Method \code{compute()}}{
computes ORESTE model based on parameters of the fields of the class.
 Usually run automatically as part of class inititation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qualiflex$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-qualiflex-summary"></a>}}
\if{latex}{\out{\hypertarget{method-qualiflex-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the ORESTE method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qualiflex$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-qualiflex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-qualiflex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qualiflex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
