% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram}
\alias{ram}
\title{Root Assesment Method}
\description{
A simple method similar to ones like \link{moora}, \link{moosra} and
 \link{copras}.

The method first normalizes all values using max version of vector
 normalization, then the weights are applied.

Method then separately computes scores for beneficial and cost criteria
 (S+i, S-i). Aggregation procedure is what makes the method differenc to
 the likes of MOORA, etc.:

\mjsdeqn{RI_i = \sqrt[2+S_{-i}]{2 + S_{+i}}}

The RI measdure is durectly usable for ranking purposes.
}
\references{
SOTOUDEH-ANVARI, Alireza. Root Assessment Method (RAM): A novel
 multi-criteria decision making method and its applications in sustainability
 challenges. Journal of Cleaner Production. 2023, Vol. 423, pp. 138695,
 available from: https://doi.org/10.1016/j.jclepro.2023.138695, ISSN 0959-6526.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{result}}{data frame consolidating results}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ram-new}{\code{ram$new()}}
\item \href{#method-ram-compute}{\code{ram$compute()}}
\item \href{#method-ram-summary}{\code{ram$summary()}}
\item \href{#method-ram-clone}{\code{ram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ram-new"></a>}}
\if{latex}{\out{\hypertarget{method-ram-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ram$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{weights vector}

\item{\code{minmax}}{minmax vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{beta}}{beta}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ram-compute"></a>}}
\if{latex}{\out{\hypertarget{method-ram-compute}{}}}
\subsection{Method \code{compute()}}{
computes the model bas of class properties. Usually we do not run the
 computation manually (it is run from class' constructor).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ram$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ram-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ram-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the RAM method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ram$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model including ranking.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
