% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rim.R
\name{rim}
\alias{rim}
\title{Reference Ideal Model}
\description{
Uses \link{topsis} approach to recommending optimal solution to the problem.
 Unlike TOPSIS, the RIM method uses its own approach to normalization (RIM
 normalization, see \link{rim_norm}).

After normalization step RIM follows TOPSIS approach. It applies weights,
 calculates the variation to the normalized reference ideal for each
 alternative.

\mjsdeqn{I_i^+ = \sqrt{\sum_{j=1}^n (y_{ij} - w_j)^2}}

\mjsdeqn{I_i^- = \sqrt{\sum_{j=1}^n (y_{ij})^2}}

The y is normalized weighted performance matrix. Finally we compute relative
 index for each alternative.

\mjsdeqn{R_i = \frac{I_i^-}{I_i^+ + I_i^-}}

Final ranking of the alternatives can be derived from Ri by sorting it
 descending.
}
\examples{

## ------------------------------------------------
## Method `rim$new`
## ------------------------------------------------

pm <- data.frame(
  age = c(25, 30, 35, 40, 50),
  pay = c(40000, 45000, 38000, 44000, 37000),
  qualification = c(1, 2, 3, 5, 5))
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
A <- c(23, 35000, 1)
B <- c(60, 50000, 5)
C <- c(30, 35000, 4)
D <- c(35, 42000, 5)
w <- c(0.25, 0.45, 0.3)
t <- rim$new(pm, A, B, C, D, w)
}
\references{
CABLES, E., LAMATA, M. T., VERDEGAY, J. L. RIM-reference ideal method in
 multicriteria decision making. Information Sciences. 2016, Vol. 337–338,
 pp. 1–10, available from: https://doi.org/10.1016/j.ins.2015.12.011,
 ISSN 0020-0255.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{RIM}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{normPM}}{normalized performance matrix}

\item{\code{A}}{numeric vector of lower ranges of criteria}

\item{\code{B}}{numeric vector of upper ranges of the criteria}

\item{\code{C}}{numeric vector of lower limit of reference ideal}

\item{\code{D}}{numeric vector of upper limit of reference ideal}

\item{\code{w}}{numeric weight vector}

\item{\code{result}}{data frame with variances (I+, I-), relative index and
ranking of the alternatives}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rim-new}{\code{rim$new()}}
\item \href{#method-rim-compute}{\code{rim$compute()}}
\item \href{#method-rim-summary}{\code{rim$summary()}}
\item \href{#method-rim-clone}{\code{rim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rim-new"></a>}}
\if{latex}{\out{\hypertarget{method-rim-new}{}}}
\subsection{Method \code{new()}}{
public constructor allowing the user to construct Borta count decision
 analysis problem and compute it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rim$new(pm, A, B, C, D, w)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion. only
numeric values expercted. Rows and columns are expected to be named.}

\item{\code{A}}{numeric vector of lower ranges of criteria}

\item{\code{B}}{numeric vector of upper ranges of the criteria}

\item{\code{C}}{numeric vector of lower limit of reference ideal}

\item{\code{D}}{numeric vector of upper limit of reference ideal}

\item{\code{w}}{numeric weight vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialized R6 class with computed results
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm <- data.frame(
  age = c(25, 30, 35, 40, 50),
  pay = c(40000, 45000, 38000, 44000, 37000),
  qualification = c(1, 2, 3, 5, 5))
rownames(pm) <- c("A1", "A2", "A3", "A4", "A5")
A <- c(23, 35000, 1)
B <- c(60, 50000, 5)
C <- c(30, 35000, 4)
D <- c(35, 42000, 5)
w <- c(0.25, 0.45, 0.3)
t <- rim$new(pm, A, B, C, D, w)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rim-compute"></a>}}
\if{latex}{\out{\hypertarget{method-rim-compute}{}}}
\subsection{Method \code{compute()}}{
Computes RIM based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rim$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rim-summary"></a>}}
\if{latex}{\out{\hypertarget{method-rim-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the RIM method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rim$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
