% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rov.R
\name{rov}
\alias{rov}
\title{Range of Value Method}
\description{
Developed by Yakowitz in 1993. Method is based on evaluation of separate
 utility functions for benefit (ui+) and cost (ui-) criteria.

Evaluation is performed on normalized performance matrix (using min-max
 normalization).

\mjsdeqn{u_i^+ =\sum_{j \in benefit} w_j \cdot r_{ij}}

\mjsdeqn{u_i^- =\sum_{j \in cost} w_j \cdot r_{ij}}

To establish rank, we compute uverage of the ui+ and ui-, with higher, the
 better values:

\mjsdeqn{u_i =\frac{u_i^+ + u_i^-}{2}}
}
\examples{

## ------------------------------------------------
## Method `rov$new`
## ------------------------------------------------

alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
minmax <- "max"
t <- rov$new(M, w, minmax)
}
\references{
Chakrakborty, S., Chatterjee, P. Das, P. P. Multi-Criteria Decision Making
 Methods in Manufacturing Environments: Models and Applications. CRC Press,
 Boca Raton, 2024, 450 p., ISBN: 978-1-00337-703-0
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ROV}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix (alternatives in rows, criteria in columns)}

\item{\code{w}}{numeric vector of weight}

\item{\code{minmax}}{vector with optimization direction (min/max) for the
criteria}

\item{\code{result}}{data frame with computational result: utility for benefit
(ui_plus), cost (ui_minus), average utility (ui) and rank based on ui}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rov-new}{\code{rov$new()}}
\item \href{#method-rov-compute}{\code{rov$compute()}}
\item \href{#method-rov-summary}{\code{rov$summary()}}
\item \href{#method-rov-clone}{\code{rov$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rov-new"></a>}}
\if{latex}{\out{\hypertarget{method-rov-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rov$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{w}}{vector of weights, its sum must be equal to 1}

\item{\code{minmax}}{minmax vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{alternatives <- c('BLR', 'BOM', 'DEL', 'MNL', 'HYD', 'GRU', 'DUB',
  'KRK', 'MAA', 'EZE')
criteria <- c('tlnt', 'stab', 'cost', 'infl', 'tm-zn', 'infr', "life")
M <- rbind(
  c(0.8181818, 0.1814159, 1.0000000, 0.1198582, 0, 0.6, 0.750),
  c(1.0000000, 0.1814159, 0.6666667, 0.1198582, 0, 0.6, 0.375),
  c(1.0000000, 0.1814159, 0.8333333, 0.1198582, 0, 0.6, 0.125),
  c(0.8181818, 0.0000000, 1.0000000, 0.3482143, 0, 0.6, 0.375),
  c(0.1818182, 0.1814159, 1.0000000, 0.1198582, 0, 0.2, 0.375),
  c(0.1818182, 0.1814159, 0.5000000, 0.1198582, 0, 0.2, 0.125),
  c(0.0000000, 1.0000000, 0.0000000, 0.5741667, 1, 1.0, 1.000),
  c(0.3636364, 0.7787611, 0.6666667, 1.0000000, 1, 0.0, 0.500),
  c(0.4545455, 0.1814159, 0.9166667, 0.1198582, 0, 0.4, 0.000),
  c(0.1818182, 0.6283186, 0.5833333, 0.0000000, 0, 0.4, 0.125)
)
rownames(M) <- alternatives
colnames(M) <- criteria
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
minmax <- "max"
t <- rov$new(M, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rov-compute"></a>}}
\if{latex}{\out{\hypertarget{method-rov-compute}{}}}
\subsection{Method \code{compute()}}{
Computes ROV based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rov$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rov-summary"></a>}}
\if{latex}{\out{\hypertarget{method-rov-summary}{}}}
\subsection{Method \code{summary()}}{
summary of the ROV method resutls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rov$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
basic information on the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rov-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rov-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rov$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
