% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seca.R
\name{seca}
\alias{seca}
\title{Simultaneous Evaluation of Criteria and Alternatives method}
\description{
The SECA (Simultaneous Evaluation of Criteria and Alternatives) method is an
 objective Multi-Criteria Decision-Making (MCDM) tool designed to determine
 alternative rankings and criteria weights concurrently. Unlike methods that
 rely on subjective human input for weighting, SECA utilizes a mathematical
 optimization model that balances two key vectors: the internal consistency
 of the alternatives (captured through standard deviation) and the conflict
 between criteria (captured through correlation). By maximizing these two
 measures, the method minimizes the bias inherent in decision-maker
 preferences, making it particularly effective for complex engineering or
 management problems where an impartial, data-driven perspective is
 required.

The procesure start with normalization of decision matrix using max
 normaliation (see \link{mcda_norm}). Then standard deviation of the criteria
 and correlation between each pair of vectors of criteria (r_ij) and is used
 for computing a degree of conflict:

\mjsdeqn{\pi_j = \sum_{l=1}^n(1 - r_{jl})}

Values of standard deviation and degree of conflict are vector normalized,
 thus establishing reference points.

The formulate solution to the decision problem multi-objective non-linar
 problem needs to be solved:

\mjsdeqn{max S_j = \sum_{j=1}^n w_j\sigma_j \sum_{j=1}^n w_j \pi_j}
}
\references{
Shankar Chakraborty, Prasenjit Chatterjee, Partha Protim Das. SIMULTANEOUS
 EVALUATION OF CRITERIA AND ALTERNATIVES(SECA) METHOD. In Multi-Criteria
 Decision-Making Methods in Manufacturing Environments. Apple Academic
 Press: New York, 2023, pp. 325-335, ISBN 9781003377030.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix}

\item{\code{w}}{vector of weights}

\item{\code{score}}{to evaluate the results}

\item{\code{minmax}}{vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{beta}}{beta (0.5 by default)}

\item{\code{ref_points}}{dataframe with information on computed reference points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-seca-new}{\code{seca$new()}}
\item \href{#method-seca-compute}{\code{seca$compute()}}
\item \href{#method-seca-summary}{\code{seca$summary()}}
\item \href{#method-seca-clone}{\code{seca$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seca-new"></a>}}
\if{latex}{\out{\hypertarget{method-seca-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for the class. Checks validity of input parameters
 and performs computation of of the model based on them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seca$new(pm, minmax = "max", beta = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{normalized performance matrix}

\item{\code{minmax}}{minmax vector specifying optimalization direction for the
criteria. Values max/min are expected. If all criteria are optimalized
in same direction the vector can be replaced by single value. Max value
is default.}

\item{\code{beta}}{beta}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seca-compute"></a>}}
\if{latex}{\out{\hypertarget{method-seca-compute}{}}}
\subsection{Method \code{compute()}}{
performs computation of SECA model based on class properties. Usually
 is not run manually as constructor calls this method automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seca$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seca-summary"></a>}}
\if{latex}{\out{\hypertarget{method-seca-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the SECA method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seca$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seca-clone"></a>}}
\if{latex}{\out{\hypertarget{method-seca-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seca$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
