% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_e34.R
\name{sensitivity_e34}
\alias{sensitivity_e34}
\title{function generalizing sensitivity checking for Electre III and IV methods}
\usage{
sensitivity_e34(object, steps = 100)
}
\arguments{
\item{object}{object representing a solution the sensitivity should be
tested for}

\item{steps}{how many steps should sensitivity testing take. Interval
for testing will be split to steps segments. The higher number of steps
the smaller the step will be, the more granular the testing will be.
Set to 100 by default}
}
\value{
returns dataframes sens_p, sens_q and sens_v with sensitivity limit for
 the criteria
}
\description{
sensitivity checking in Electre III \code{\link{electre3}} and IV
 \code{\link{electre4}} methods. The difference is in support of weights
 (Electre IV does not support these).

This function generalizes common computation logic into this function.

This function is intended for package internal use only.

sensitivity analysis is performed on values of preference (p), indefference
 (q) and veto (v) thresholds.

Since the thresholds (all of them) are specfied separately for each
 criterion, their sensitivity needs to be also evaluated seprately.
 This is being realized in this function by generating separate
 dataframe for each threshold (sens_p, sens_q, sens_v) with following
 structure:

criterium; from; default; to

From and to limits are limist of of solution stability, meaning that
 going under it (from) or over it (to) will produce different result.

Sensitivity is tested from default value (the value used to produce
 result) going down to 0 for from column and up to maximal performance
 in the criterium j (max(pm[j])). If no change in result is detected,
 value "insens." is inserted into dataframe.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{ELECTRE}
\keyword{III}
\keyword{IV}
\keyword{sensitivity}
\keyword{testing}
