% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotis.r
\name{spotis}
\alias{spotis}
\title{Stable Preference Ordering Super Intelligence}
\description{

}
\examples{

## ------------------------------------------------
## Method `spotis$new`
## ------------------------------------------------

# Example from the book (see references)
PM <- rbind(
   c(10.5, -3.1, 1.7),
   c(-4.7, 0, 3.4),
   c(8.1, 0.3, 1.3),
   c(3.2, 7.3, -5.3)
)
rownames(PM) <- c('A1', 'A2', 'A3', 'A4')
colnames(PM) <- c('C1', 'C2', 'C3')
minmax <- c('max', 'min', 'max')
w <- c(0.2, 0.3, 0.5)
bounds <- rbind(
   c(-5, 12),
   c(-6, 10),
   c(-8, 5)
)
rownames(bounds) <- colnames(PM)
colnames(bounds) <- c('min', 'max')
result <- SPOTIS(PM, w, minmax, bounds)
}
\references{
Dezert et al. The SPOTIS Rank Reversal Free Method for Multi-Criteria
  Decisionmaking Support. In: IEEE 23rd International Conferencew on
  Information Fusion (FUSION). Rustenburg, South Africa, 2020, pp. 1-8,
  doi: 10.23919/FUSION45008.2020.9190347
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{SPOTIS}
\keyword{reversal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{bounds}}{dataframe of bounds of criteria (criteria in rows with
min and max columns)}

\item{\code{score_raw}}{vector of average distance to ideal solution}

\item{\code{score}}{vector of average distance to ideal solution ordered}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spotis-new}{\code{spotis$new()}}
\item \href{#method-spotis-compute}{\code{spotis$compute()}}
\item \href{#method-spotis-summary}{\code{spotis$summary()}}
\item \href{#method-spotis-clone}{\code{spotis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spotis-new"></a>}}
\if{latex}{\out{\hypertarget{method-spotis-new}{}}}
\subsection{Method \code{new()}}{
public constructor allowing the user to construct SPOTIS decision
 analysis problem and compute it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spotis$new(pm, w, minmax = "max", bounds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{performance matrix of alternatives in criteria}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{bounds}}{dataframe of bounds of criteria (criteria in rows with
min and max columns)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
initialized R6 class with computed results for SPOTIS
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example from the book (see references)
PM <- rbind(
   c(10.5, -3.1, 1.7),
   c(-4.7, 0, 3.4),
   c(8.1, 0.3, 1.3),
   c(3.2, 7.3, -5.3)
)
rownames(PM) <- c('A1', 'A2', 'A3', 'A4')
colnames(PM) <- c('C1', 'C2', 'C3')
minmax <- c('max', 'min', 'max')
w <- c(0.2, 0.3, 0.5)
bounds <- rbind(
   c(-5, 12),
   c(-6, 10),
   c(-8, 5)
)
rownames(bounds) <- colnames(PM)
colnames(bounds) <- c('min', 'max')
result <- SPOTIS(PM, w, minmax, bounds)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spotis-compute"></a>}}
\if{latex}{\out{\hypertarget{method-spotis-compute}{}}}
\subsection{Method \code{compute()}}{
Computes SPOTIS model based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spotis$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spotis-summary"></a>}}
\if{latex}{\out{\hypertarget{method-spotis-summary}{}}}
\subsection{Method \code{summary()}}{
prepares summary of the SPOTIS method resutls and outputs them
 to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spotis$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spotis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spotis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spotis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
