% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topsis.R
\name{topsis}
\alias{topsis}
\title{TOPSIS decision support method}
\description{
The acronym TOPSIS stands for: Technique of Order Preference Similarity to
 the Ideal Solution. As name suggests TOPSIS provides its guidance based on
 evaluation of the similarity to both ideal and anti-ideal variant of the
 solution.

Original method uses 5 steps for the procedure. In first step the procedure
 normalizes values in performance matrix nad applies weights to it in step
 2. In step 3 ideal variant \mjseqn{A^*} and anti-ideal variant \mjseqn{A^-}
 is computed as maximums and minimums of the criteria in performance matrix.

In step 4 distance to ideal \mjseqn{D^*} and anti-ideal variant \mjseqn{D^-}
 is computed and in step 5 used to compute closenes criterium

\mjsdeqn{C_i = \frac{D^-_i}{D^-_i + D^*_i}}

Criterium C is then directly usable to rank alternatives. C is always in
 interval of 0-1, the closer the value is to 1, the closer it is to ideal
 variant.
}
\examples{

## ------------------------------------------------
## Method `topsis$new`
## ------------------------------------------------

PM <- cbind(
   c(8,7,2,1),
   c(5,3,7,5),
   c(7,5,6,4),
   c(9,9,7,3),
   c(11,10,3,7),
   c(6,9,5,4)
)
colnames(PM) <- c('Site 1', 'Site 2', 'Site 3', 'Site 4', 'Site 5',
   'Site 6')
rownames(PM) <- c('Investment costs (million EUR)',
                  'Employment needs (hundred employees)',
                  'Social impact (1-7)',
                  'Environmental impact (1-7)')
PM <- t(PM)
minmax <- 'max'
w <- c(0.4, 0.4, 0.1, 0.2)
result <- topsis$new(PM, w, minmax)
}
\references{
ALAOUI, Mohamed El. Fuzzy TOPSIS: Logic, Approaches, and Case Studies. Boca
 Raton: CRC Press, 2021. 216 s. ISBN 978-0-367-76748-8.

Papathanasiou, Jason, Ploskas, Nikolaos. Multiple Criteria Decision Aid
 Methods, Examples and Python Implementations. Springer, 173 p., ISBN
 978-3-319-91648-4.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{TOPSIS}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pm_orig}}{original (not transformed) performance matrix of
alternatives in criteria}

\item{\code{pm}}{performance matrix of alternatives in criteria (all criteria
are maximized)}

\item{\code{w}}{vector of weights}

\item{\code{minmax}}{direction (max or min) of criteria optimization}

\item{\code{closeness_ord}}{ordered list of alternatives using closeness (C)
criterium}

\item{\code{normPM}}{normalized performance matrix}

\item{\code{weightPM}}{weighted normalized performance matrix}

\item{\code{A_ideal}}{positive ideal solution}

\item{\code{A_anti}}{anti ideal solution}

\item{\code{D_ideal}}{alternative closeness to ideal variant}

\item{\code{D_anti}}{alternative closeness to anti-ideal variant}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-topsis-new}{\code{topsis$new()}}
\item \href{#method-topsis-compute}{\code{topsis$compute()}}
\item \href{#method-topsis-summary}{\code{topsis$summary()}}
\item \href{#method-topsis-clone}{\code{topsis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topsis-new"></a>}}
\if{latex}{\out{\hypertarget{method-topsis-new}{}}}
\subsection{Method \code{new()}}{
Public constructor for TOPSIS object. Performas validation of input
 parameters and uses them to compute the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{topsis$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pm}}{Matrix or data frame containing the performance table. Each
row corresponds to an alternative, and each column to a criterion.
Only numeric values expercted. Rows and columns are expected to be
named.}

\item{\code{w}}{vector containing the weights of the criteria.}

\item{\code{minmax}}{criteria MinMax Vector containing the preference direction
on each of the criteria. "min" (resp."max") indicates that the
criterion has to be minimized (maximized). Default value: max}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Instance of topsis class uncluding inputs and model results in class's
 properties
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{PM <- cbind(
   c(8,7,2,1),
   c(5,3,7,5),
   c(7,5,6,4),
   c(9,9,7,3),
   c(11,10,3,7),
   c(6,9,5,4)
)
colnames(PM) <- c('Site 1', 'Site 2', 'Site 3', 'Site 4', 'Site 5',
   'Site 6')
rownames(PM) <- c('Investment costs (million EUR)',
                  'Employment needs (hundred employees)',
                  'Social impact (1-7)',
                  'Environmental impact (1-7)')
PM <- t(PM)
minmax <- 'max'
w <- c(0.4, 0.4, 0.1, 0.2)
result <- topsis$new(PM, w, minmax)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topsis-compute"></a>}}
\if{latex}{\out{\hypertarget{method-topsis-compute}{}}}
\subsection{Method \code{compute()}}{
Computes TOPSIS model based on parametrs specified in constructor.
 Normally this methods does not need to be run manually as constructor
 calls it automatically.

Manual re-computation si required only if the user changes class' fields
 without using the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{topsis$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topsis-summary"></a>}}
\if{latex}{\out{\hypertarget{method-topsis-summary}{}}}
\subsection{Method \code{summary()}}{
Creates summary for TOPSIS model and sends it to console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{topsis$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topsis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-topsis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{topsis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
