% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_electre_pqv}
\alias{validation$validate_electre_pqv}
\title{Validate consistency of thresholds}
\arguments{
\item{p}{preference threshold}

\item{q}{indifference threshold}

\item{v}{veto threshold}

\item{cri}{vector of names for criteria}
}
\description{
Validate consistency of thresholds
}
