% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_fuzzy_consistency}
\alias{validation$validate_fuzzy_consistency}
\title{function to check consistency of matrix againts dictionary of values}
\arguments{
\item{m}{matrix to be checked}

\item{dict}{dictionary to check matrix against}

\item{fn}{type of fuzzy number (i.e. 3 for triangular, 4 for trapeziodal)}

\item{msg}{message to guide user on what is being examined i.e. PM,
weight matrix, etc.}
}
\description{
function to check consistency of matrix againts dictionary of values
}
