% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_invalid_val}
\alias{validation$validate_invalid_val}
\title{Validate that the matrix or dataframe has only valid values}
\arguments{
\item{m}{matrix or dataframe to check}

\item{valid_val}{vector of valid values to check}

\item{msg}{Identification of what are we checking (i.e. performance matrix)}
}
\description{
Vector of valid values must contain full set of valid values to check
 against. If other value is detected the error will be generated and
 computation run will be stoped.
}
