% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_matrix_numeric}
\alias{validation$validate_matrix_numeric}
\title{Validates that the matrix or dataframe has only numeric elements}
\arguments{
\item{pm}{matrix or dataframe to check}
}
\description{
Validates that the matrix or dataframe has only numeric elements
}
