% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_no_elements_vs_cri}
\alias{validation$validate_no_elements_vs_cri}
\title{Validates that the number of elements in the vector is same as number of
 criteria}
\arguments{
\item{vect}{vector to check}

\item{ncri}{number of criteria}

\item{msg}{identification of what are we checking to use in error message}

\item{test_numeric}{set to TRUE to test vect as numeric vector}
}
\description{
Validates that the number of elements in the vector is same as number of
 criteria
}
\examples{
w <- c(0.125, 0.2, 0.2, 0.2, 0.175, 0.05, 0.05)
ncri <- 7
validation$validate_no_elements_vs_cri(w, ncri, "weights", TRUE)
}
