% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_pm}
\alias{validation$validate_pm}
\title{Validate performance matrix (internal use only)}
\arguments{
\item{pm}{performance matrix}
}
\description{
Function for validation of the performance matrix. Numeric values are
 expected. There must be at least two alternatives and criteria for the
 problem do fall under decision making.

If validation rules are violated, the run of function will stop with
 error message.
}
