% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_promethee_thresholds}
\alias{validation$validate_promethee_thresholds}
\title{validates consistency of preference, indifference and intermediate
 thresholds pro PROMETHEE based approaches}
\arguments{
\item{p}{preference threshold}

\item{q}{indifference threshold}

\item{s}{intermediate threshold}

\item{ncri}{number of criteria}
}
\value{
adjusted intermediate threshold
}
\description{
validates consistency of preference, indifference and intermediate
 thresholds pro PROMETHEE based approaches
}
