% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_promethee_thresholds}
\alias{validation$validate_promethee_thresholds}
\title{validates consistency of preference, indifference and intermediate
 thresholds pro PROMETHEE based approaches}
\arguments{
\item{p}{preference threshold}

\item{q}{indifference threshold}

\item{s}{intermediate threshold}

\item{ncri}{number of criteria}
}
\value{
adjusted intermediate threshold
}
\description{
validates consistency of preference, indifference and intermediate
 thresholds pro PROMETHEE based approaches
}
\examples{
p <- c(10, 0, 450, 50, 0, 0) #indifference threshold
q <- c(0, 30, 50, 10, 0, 0) #prefference threshold
s <- c(0,0,0,0,0,5) #intermediate threshold
shape <- c('U-shape', 'V-shape', 'linear', 'level', 'default', 'Gaussian')
sj <- validation$validate_promethee_thresholds(p, q, s, shape, 6)
}
