% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validate_vector_progression}
\alias{validation$validate_vector_progression}
\title{Validate the numeric progresion in provided vector}
\arguments{
\item{vect}{numeric vector with values to check}
}
\description{
Presumes that the elements in the vector are ordered ascending. This is
 usefull for various threshold checks.
}
