% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation$validation_vector_in_interval}
\alias{validation$validation_vector_in_interval}
\title{Validates tahl all values in the vactor are in specified bounds}
\arguments{
\item{vect}{numeric vector to check}

\item{from}{lower bound of the checking interval}

\item{to}{upper bound for the checking interval}

\item{msg}{identification of object we are checking (i.e. indifference
threshold)}
}
\description{
Validates tahl all values in the vactor are in specified bounds
}
\examples{
c_minus <- 0.65
c_zero <- 0.75
c_plus <- 0.85
d_minus <- 0.25
d_plus <- 0.5
thres <- c(d_minus, d_plus, c_minus, c_zero, c_plus)
validation$validation_vector_in_interval(thres, 0, 1,
                                         "C-, C0, C+, D-, D+")
}
