% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisp.R
\name{wisp}
\alias{wisp}
\title{Simple weight sum weighting method}
\description{
Simple methos for alternatives rank establishment considering weighted sum
 producs.

The inputs of method are formulated as usual for this type of decision
 problem by creading matrix of alternatives (in rows) performance in
 selected criteria (columns).

The values of the performance need to be normalized using:

\mjsdeqn{r_{ij} = \frac{x_{ij}}{max_i x_{ij}}}

The normalization is performed using same equation for both benefit and
 cost criteria as the method is based on analyzing the difference of these.

Then we need to estables four utility measures:

1) weighted sum

\mjsdeqn{u_i^{wsd} = \sum_{j \in \Omega_{max}}r_{ij}w_j - \sum_{j \in \Omega_{min}} r_{ij}w_j}

2) weighted product

\mjsdeqn{u_i^{wpd} = \prod_{j \in \Omega_{max}} r_{ij}w_j - \prod_{j \in \Omega_{min}} r_{ij}w_j}

3) weighted sum ratio

\mjsdeqn{u_i^{wsr} = \frac{\sum_{j \in \Omega_{max}}r_{ij}w_j}{\sum_{j \in \Omega_{min}} r_{ij}w_j}}

4) weighted product ratio

\mjsdeqn{u_i^{wpr} = \frac{\prod_{j \in \Omega_{max}} r_{ij}w_j}{\prod_{j \in \Omega_{min}} r_{ij}w_j}}

From these values four utility measures need to be derived. For weighted sum

\mjsdeqn{\overline{u_i^{wsd}} = \frac{1 + u_i^{wsd}}{1 + max_i u_i^{wsd}}}

weighted products

\mjsdeqn{\overline{u_i^{wpd}} = \frac{1 + u_i^{wpd}}{1 + max_i u_i^{wpd}}}

weighted sum ratio

\mjsdeqn{\overline{u_i^{wsr}} = \frac{1 + u_i^{wsr}}{1 + max_i u_i^{wsr}}}

weighted product ratio

\mjsdeqn{\overline{u_i^{wpr}} = \frac{1 + u_i^{wpr}}{1 + max_i u_i^{wpr}}}

Overall metric for performance is derived from these four measures by
 averaging them.

\mjsdeqn{u_i = (\overline{u_i^{wsd}} + \overline{u_i^{wpd}} + \overline{u_i^{wsr}} + \overline{u_i^{wpr}}) / 4}

Using this metric we can sort the alternatives from best to worst.
}
\references{
Stanujkic, D., Popovic, G., Karabasevic, D., Meidute-Kavaliauskiene, I., &
 Ulutaş, A. (2021). An integrated simple weighted sum product method—WISP.
 IEEE Transactions on Engineering Management.
}
\author{
Pavel Šenovský \email{pavel.senovsky@vsb.cz}
}
\keyword{WISP}
\keyword{WSM}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wisp-new}{\code{wisp$new()}}
\item \href{#method-wisp-clone}{\code{wisp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wisp-new"></a>}}
\if{latex}{\out{\hypertarget{method-wisp-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wisp$new(pm, w, minmax = "max")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wisp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wisp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wisp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
